### 2022-08-15 #553 @antaldaniel
dataset: Create Data Frames that are Easier to Exchange and Reuse
Submitting Author Name: Daniel Antal
Submitting Author Github Handle: @antaldaniel
Repository:  https://github.com/dataobservatory-eu/dataset/
Version submitted: 0.1.7
Submission type: Standard

Version accepted: TBD
---

```
Package: dataset
Title: Create Data Frames that are Easier to Exchange and Reuse
Version: 0.1.7.3
    person(given = "Daniel", family = "Antal",
           email = "daniel.antal@dataobservatory.eu",
           comment = c(ORCID = "0000-0001-7513-6760")
           )
Description: The aim of the 'dataset' package is to make tidy datasets easier to release,
    exchange and reuse. It organizes and formats data frame 'R' objects into well-referenced,
    well-described, interoperable datasets into release and reuse ready form. A subjective
    interpretation of the  W3C  DataSet recommendation and the datacube model  <https://www.w3.org/TR/vocab-data-cube/>,
    which is also used in the global Statistical Data and Metadata eXchange standards,
    the application of the connected Dublin Core <https://www.dublincore.org/specifications/dublin-core/dcmi-terms/>
    and DataCite <https://support.datacite.org/docs/datacite-metadata-schema-44/> standards
    preferred by European open science repositories to improve the findability, accessibility,
    interoperability and reusability of the datasets.
    assertthat,
    ISOcodes,
    utils
    covr,
    declared,
    eurostat,
    here,
    kableExtra,
    rdflib,
    readxl,
    statcodelists,
    tidyr
```

You can find the package website on [dataset.dataobservatory.eu](https://dataset.dataobservatory.eu/). The article [Motivation: Make Tidy Datasets Easier to Release Exchange and Reuse](https://dataset.dataobservatory.eu/articles/motivation.html) will eventually be condensed into a JOSS paper.  It has a major development dilemma.

	- [x ] data deposition

Open science repositories and analyst comupters are full with datasets that have no provenance, structural or referential data. We believe that whenever possible, metadata should be machine-recorded when possible, and should not be detached from an R object.
There are several R packages that have overalapping goals or functionality to `dataset`, but they use a different philosophy. When exporting to different files, they should be written as exported, but no sooner, and preferably into the file that contains the data.

This package is intended to give a common foundation to the rOpenGov reproducible research packages.  It mainly serves communities that want to reuse statistical data (using the SDMX statistical (meta)data exchange sources, like Eurostat, IMF, World Bank, OECD...) or release new datasets from primary social sciences data that can be integrated into an SDMX compatible API or placed on a knowledge graph.  Our main aim is to provide a clear publication workflow to the European open science repository Zenodo, and clear serialization strategies to RDF application.

The [dataspice package](https://CRAN.R-project.org/package=dataspice) aims to create well-defined and referenced datasets, but follows a different schema and a different publication strategy. The dataset package follows the more restrictive W3C/SDMX "DataSet" definition within the datacube model, which is better suited to synchronize with statistical data sources. Unlike dataset, it uses a manual metadata entry from CSV files. (See the documentation of the [dataspice](https://docs.ropensci.org/dataspice/) package.)

The `dataset` package aims for a higher level of reproducibality, and does not detach the metadata from the R object's attributes (it is aimed to be used in other reproducible research pacakges that will directly record provenance and other transactional metadata into the attributes.) We aim to bind together `dataspice` and `dataset` by creating export functions to csv files that contain the same metadata that dataspice records.  Generally, dataspice seems to be better suited to raw, observational data, while dataset for statistically processed data.

The intended use of `dataset` is to start correctly record referential, structural and provenance metadata retrieved by various reproducible science packages that interact with statistical data (such as the rOpenGov packages [eurostat](https://CRAN.R-project.org/package=eurostat) and [iotables](https://CRAN.R-project.org/package=iotables), or the [oecd](https://CRAN.R-project.org/package=oecd) package.

Neither `dataset` or `dataspice` are very suitable of or documenting social sciences survey data, which are usually held in datasets. Our aim is to connect `dataset`, [declared](https://CRAN.R-project.org/package=declared) and [DDIwR](https://CRAN.R-project.org/package=DDIwR) to create such datasets with DDI codebook metadata.  They will create a stable new foundation of the [retroharmonize](https://CRAN.R-project.org/package=retroharmonize) package to create new, well-documented and harmonized statistical datasets from the observational datasets of social sciences surveys.

The [zen4R]((https://CRAN.R-project.org/package=zen4R)) package provides reproducible export functionality to the zenodo open science repository. Interacting with `zen4R` may be intimidating for the casual R user as it uses R6 classes. Our aim to provide an export function that completely wraps the workings of `zen4R` when releasing the dataset.

In our experience, while the tidy data standards make reuse more efficient by eliminating unnecessary data processing steps before analysis or placement in a relational database, the application of DataSet definition and the datacube model with the information science metadata standards make reuse more efficient with exchanging and combining the data with other data in different datasets.

Yes

- [x ] I have read the [rOpenSci packaging guide](https://devguide.ropensci.org/building.html).
- [x ] I have read the [author guide](https://devdevguide.netlify.app/authors-guide.html) and I expect to maintain this package for at least 2 years or to find a replacement.

- [x ] does not violate the Terms of Service of any service it interacts with.
- [ x] has a CRAN and OSI accepted license.
- [ x] contains a [README with instructions for installing the development version](https://ropensci.github.io/dev_guide/building.html#readme).
- [ x] includes [documentation with examples for all functions, created with roxygen2](https://ropensci.github.io/dev_guide/building.html#documentation).
- [x ] contains a vignette with examples of its essential functions and uses.
- [ x] has a [test suite](https://ropensci.github.io/dev_guide/building.html#testing).

- [x ] Do you intend for this package to go on CRAN? -> Yes, I started the CRAN publication process, but opted to stop and get feedback from rOpenSic first
- [ ] Do you intend for this package to go on Bioconductor? -> Don't know.

- [ x] I agree to abide by [rOpenSci's Code of Conduct](https://ropensci.org/code-of-conduct/) during the review process and in maintaining my package should it be accepted.
