+++ vignettes/motivation.Rmd
@@ -21,11 +21,39 @@ knitr::opts_chunk$set(
 )
 ```
 
-The aim of the dataset package is to make tidy datasets easier to release, exchange and reuse. It organizes and formats data frame R objects into well-referenced, well-described, interoperable datasets into release and reuse ready form. We apply a subjective, R language compatible interpretation of the  W3C [Data Cube Vocabulary](https://www.w3.org/TR/vocab-data-cube/) based on the statistical SDMX data cube model^[RDF Data Cube Vocabulary, W3C Recommendation 16 January 2014  <https://www.w3.org/TR/vocab-data-cube/>, Introduction to SDMX data modeling <https://www.unescap.org/sites/default/files/Session_4_SDMX_Data_Modeling_%20Intro_UNSD_WS_National_SDG_10-13Sep2019.pdf>]. We enrich these statistical data exchange definition with similar definitions from information science and the exchange of scientific resources on open science repositories by applying the [Dublin Core](https://www.dublincore.org/specifications/dublin-core/dcmi-terms/) and [DataCite](https://support.datacite.org/docs/datacite-metadata-schema-44/) metadata specifications for R attributes to improve the findability, accessibility, interoperability and reusability of the datasets.
+The aim of the 'dataset' package is to make tidy datasets easier to release, exchange and reuse. It organizes and formats data frame 'R' objects into well-referenced, well-described, interoperable datasets into release and reuse ready form. It applies a subjective, R-language compatible interpretation of the  W3C [Data Cube Vocabulary](https://www.w3.org/TR/vocab-data-cube/) based on the statistical SDMX data cube model^[RDF Data Cube Vocabulary, W3C Recommendation 16 January 2014  <https://www.w3.org/TR/vocab-data-cube/>, Introduction to SDMX data modeling <https://www.unescap.org/sites/default/files/Session_4_SDMX_Data_Modeling_%20Intro_UNSD_WS_National_SDG_10-13Sep2019.pdf>]. We enrich these statistical data exchange definition with similar definitions from information science and the exchange of scientific resources on open science repositories by applying the [Dublin Core](https://www.dublincore.org/specifications/dublin-core/dcmi-terms/) and [DataCite](https://support.datacite.org/docs/datacite-metadata-schema-44/) metadata specifications for R attributes to improve the findability, accessibility, interoperability and reusability of the datasets.
 
 We organize and format standard tabular data R objects (data.fame, data.table, tibble, or even well-structured lists like json and json-ld) become highly interoperable and can be placed into relational databases, semantic web applications, archives, repositories.  This puts the [FAIR](https://www.go-fair.org/fair-principles/) principles into the practice of an R environment workflow: they make these tabular data objects more findable, more accessible, interoperable and reusable. 
 
-While the tidy data standards make reuse more efficient by eliminating unnecessary data processing steps before analysis or placement in a relational database, the application of DataSet definition and the datacube model with the information science metadata standards make reuse more efficient with exchanging and combining the data with other data in different datasets. 
+
+## Statement of need
+
+Open science repositories and analyst comupters are full with datasets that have no provenance, structural or referential data. We believe that whenever possible, metadata should be machine-recorded when possible, and should not be detached from an R object. 
+
+There are several R packages that have overalapping goals or functionality to `dataset`, but they use a different philosophy. When exporting to different files, they should be written as exported, but no sooner, and preferably into the file that contains the data.  
+
[…]
