+++ README.md
@@ -6,8 +6,8 @@
 <!-- badges: start -->
 
 [![lifecycle](https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6969683.svg)](https://zenodo.org/record/6950435#.YukDAXZBzIU)
-[![devel-version](https://img.shields.io/badge/devel%20version-0.1.6-blue.svg)](https://github.com/dataobservatory-eu/dataset)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6992467.svg)](https://zenodo.org/record/6950435#.YukDAXZBzIU)
+[![devel-version](https://img.shields.io/badge/devel%20version-0.1.7-blue.svg)](https://github.com/dataobservatory-eu/dataset)
 [![dataobservatory](https://img.shields.io/badge/ecosystem-dataobservatory.eu-3EA135.svg)](https://dataobservatory.eu/)
 [![Follow
 rOpenGov](https://img.shields.io/twitter/follow/ropengov.svg?style=social)](https://twitter.com/intent/follow?screen_name=ropengov)
@@ -78,7 +78,7 @@ dublincore(my_iris_dataset)
 #> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"     
 #> 
 #> $dimensions
-#> [1] names       isDefinedBy codeListe  
+#> [1] names       isDefinedBy codeList   
 #> <0 rows> (or 0-length row.names)
 #> 
 #> $measures
@@ -112,7 +112,7 @@ dublincore(my_iris_dataset)
 #> [1] "American Iris Society"
 #> 
 #> $Date
-#> [1] "2022-08-06"
+#> [1] "2022-08-15"
 #> 
 #> $Creator
 #> [1] "Edgar Anderson [aut]"
