### 2022-08-15 #553 @mpadge
@antaldaniel Then you can start the checks yourself by calling `@ropensci-review-bot check package`
