### 2022-08-15 #553 @antaldaniel
@@ropensci-review-bot check package
