### 2022-08-16 #553 @adamhsparks
Hi again, @antaldaniel. If you could please address the issues that the bot flagged with the ✖️, then I can proceed with your submission.
