+++ man/attributes_measures.Rd
@@ -28,3 +28,10 @@ Do not confuse with \code{base::\link{attributes}}, which applies to the attribu
 of the entire dataset, and not each observation (measurement) row.
 See the W3C and SDMX definition of a \href{https://www.w3.org/TR/vocab-data-cube/#dsd-dimensions}{attribute}.
 }
+\examples{
+df <- data.frame ( sex = c("M", "F"), value = c(1,2), unit = c("NR", "NR"))
+dimensions(df, sdmx_attributes = "sex") <- "sex"
+measures(df) <- "value"
+attributes_measures(df) <- "unit"
+attributes_measures(df)
+}
