+++ man/measures.Rd
@@ -25,6 +25,11 @@ Measures of a dataset
 See the W3C and SDMX definition of a \href{https://www.w3.org/TR/vocab-data-cube/#dsd-dimensions}{measure}.
 }
 \examples{
+df <- data.frame ( sex = c("M", "F"), value = c(1,2), unit = c("NR", "NR"))
+dimensions(df, sdmx_attributes = "sex") <- "sex"
+measures(df) <- "value"
+attributes_measures(df) <- "unit"
+measures(df)
 df <- data.frame ( sex = c("M", "F"), value = c(1,2))
 measures(df) <- "value"
 measures(df)
