+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Date: 2022-08-15
-Version: 0.1.7
+Date: 2022-08-17
+Version: 0.1.7.0001
 Authors@R: 
     person(given = "Daniel", family = "Antal", 
            email = "daniel.antal@dataobservatory.eu", 
@@ -32,6 +32,7 @@ Imports:
     rlang,
     utils
 Suggests: 
+    covr,
     declared,
     dplyr,
     eurostat,
