### 2022-08-17 #553 @antaldaniel
Hi @adamhsparks I hope I managed to add these things, with the following exception.

✔️does not have a 'codemeta.json' file -> added with codematar.
✔️does not have a 'contributing' file -> added CONTRIBUTING.md
✔️ These functions do not have examples: [attributes_measures]. -> added
✖️ Function names are duplicated in other packages

I tried to avoid duplications while keeping in mind rOpenSci duplication guildelines, and at this point, I do not see which are the dupblications and if there is any sensible way to resolve them.

Your guidelines state "Avoid function name conflicts with base packages or other popular ones (e.g. ggplot2, dplyr, magrittr, data.table)"  The package currently has no name conflict with any packages that I was thinking of to be used together, and I do not know how to test for this.  (Sorry if this is somewhere in the [1.3 Package API](https://devguide.ropensci.org/building.html?q=verb#function-and-argument-naming)

✔️ Package has no continuous integration checks -> added
✖️ Package coverage is 67.8% (should be at least 75%)

I do not see a sensible way to achieve 75%+ codecov coverage with a metadata package that is in an early development page, still has development questions open (see [](), hence the submission here before the first CRAN release).  For example, in the target category, other metadata management pacakges like codemetar has a 42% coverage, EML has 65%,  both below the current coverage before the first release of dataset.
