### 2022-08-17 #553 @mpadge
It's in the check results. Under "*4. Other Checks*", you'll see a "Details of other checks (click to open)". You can also generate those yourself by running:
``` r
library(pkgcheck)
checks <- pkgcheck("/<path>/<to>/<dataset-pkg>")
checks_md <- checks_to_markdown(checks, render = TRUE)
```
That will automatically open a HTML-rendered version of the checks, just like the above. You can use that repeatedly as you work through the issues highlighted above.
