### 2022-08-17 #553 @antaldaniel
@mpadge Oh, really, sorry for asking the obvious.

I would like to comment here on the issue then in substance.  The main development question of the package, which aims to make R objects standard datasets (as defined by W3C and SDMX), is to add structural and referential metadata, is if the best way to do this is to create an s3 object or not (see the dilemma [here](https://dataset.dataobservatory.eu/articles/motivation.html).)

In the current stage, it is a pseudo object inherited from data.frame, but it can be seen also as a utility to any data.frame, tibble, and data.table (or similar tabular format) R objects.  The functions, which have duplicates in other packages, are  following a very simple naming convention.  I think that these is the cleanest API interface that I can think of, for example, the

[subject()](https://dataset.dataobservatory.eu/reference/subject.html) gets the metadata attribute `Subject` and the [`subject<-`()](https://dataset.dataobservatory.eu/reference/subject.html) sets it.  As DataCite, Dublin Core and schema.org has dozens of potential attributes, to me the easiest is to use in a slightly modified form the name of the attribute to set/get its value.

All these functions are lowercase to manipulate a camelCase standard attribute.  Except for the SDMX attribute 'attribute', which would create a conflict with the base R 'attributes()' function.
