### 2022-08-18 #553 @adamhsparks
Hi @antaldaniel,
I can understand the difficulty in writing tests for such a non-standard package. But I've had a look at `covr::report()` for "dataobservatory-eu/dataset". I think that there is still low-hanging fruit here that can be covered to get your code-coverage up to 75% that we ask for.

For instance, [Lines 40-43](https://github.com/dataobservatory-eu/dataset/blob/0820db303af1c5fc6a1e0f76c00d700bbf60b5d8/R/language.R#L40) are covered but [Lines 44-45](https://github.com/dataobservatory-eu/dataset/blob/0820db303af1c5fc6a1e0f76c00d700bbf60b5d8/R/language.R#L44) aren't. These are seemingly the same except for checking on 2 or 3 letter ISO codes, unless I'm mistaken.

Or the message response within the `stop()` functions in the same file aren't checked.

Could I ask that you have another look and see if you can't further improve the coverage a bit more?
