### 2022-08-19 4c2907d
further unit tests
