### 2022-08-19 280a4ce
further unit tests
