### 2022-08-19 373e04e
further unit tests
