+++ tests/testthat/test-dataset.R
@@ -3,9 +3,8 @@ dimensions(df, sdmx_attributes = "sex") <- "sex"
 measures(df) <- "value"
 attributes_measures(df) <- "unit"
 
-attributes_measures(df)
-
 test_that("dimensions() work", {
+  expect_true(inherits(dimensions(df), 'data.frame'))
   expect_equal(dimensions(df)$names, c("sex"))
   expect_equal(dimensions(df)$isDefinedBy, c("https://purl.org/linked-data/cube|https://raw.githubusercontent.com/UKGovLD/publishing-statistical-data/master/specs/src/main/vocab/sdmx-attribute.ttl"))
 })
@@ -15,6 +14,7 @@ test_that("measures() work", {
 })
 
 test_that("attributes_measures() work", {
+  expect_true(inherits(attributes_measures(df), 'data.frame'))
   expect_equal(attributes_measures(df)$names, c("unit"))
 })
 
