### 2022-08-19 #553 @antaldaniel
Hi @adamhsparks  I went up to 71.27%, but further changes are not very productive. I  did not extensively cover two areas, one is the constructor for the dataset()  itself, where I expect potentially breaking changes, and in the file I/O areas, where I think I would like to come up with a more general solution, and also avoid test being run on CRAN later.   As the overwrite function and its messages make the most branches, this is a bit of a play with %, as the very same copied test is tested again and again.

Do you have a good solution to include download and file I/O tests that run fast enough or cause no disruption when later run on CRAN?
