### 2022-08-19 9dea33e
remove two not used internal functions
