### 2022-08-19 a01f5d8
remove disused code elements
