### 2022-08-19 af8cf99
remove disused old internal functions
