+++ R/dataset.R
@@ -256,78 +256,5 @@ attributes_measures <- function(x) attr(x, "attributes")
   invisible(x)
 }
 
-#' @keywords internal
-dimensions_names <- function(x, dimensions)   {
 
-  if ( is.character(dimensions)) {
-    dimensions <- which(names(x) %in% dimensions)
-    not_in_dataset <- (! (which (dimensions %in% names(x))))
-    if (length(not_in_dataset>0)) {
-      warning("dimensions_get(x, dimensions, ...): ", paste(dimensions[not_in_dataset], collapse = ", "), " not in dataset.")
-    }
-  }
-
-  if ( is.numeric(dimensions) ) {
-    not_in_dataset <- dimensions[which (! dimensions %in% 1:ncol(x))]
-
-    if (length(not_in_dataset>0)) {
-      warning("dimensions_get(x, dimensions, ...): ", paste(not_in_dataset, collapse = ", "), " not in dataset.")
-    }
-
-    if ( all(dimensions %in% not_in_dataset)) {
-      stop("dimensions_get(x, dimensions, ...): none of ", paste(dimensions, collapse = ", "), " in dataset.")
-    }
-    names(x)[as.integer(dimensions[! dimensions %in% not_in_dataset])]
-  }
-}
-
-#' @keywords internal
-measures_names <- function(x, measures)   {
-
-  if ( is.character(measures)) {
-    measures <- which(names(x) %in% measures)
-    not_in_dataset <- (! (which (measures %in% names(x))))
-    if (length(not_in_dataset>0)) {
-      warning("measures_get(x, measures, ...): ", paste(measures[not_in_dataset], collapse = ", "), " not in dataset.")
-    }
-  }
-
-  if ( is.numeric(measures) ) {
-    not_in_dataset <- measures[which (! measures %in% 1:ncol(x))]
-
-    if (length(not_in_dataset>0)) {
-      warning("measures_get(x, measures, ...): ", paste(not_in_dataset, collapse = ", "), " not in dataset.")
-    }
-
-    if ( all(measures %in% not_in_dataset)) {
-      stop("measures_get(x, measures, ...): none of ", paste(measures, collapse = ", "), " in dataset.")
-    }
-    names(x)[as.integer(measures[! measures %in% not_in_dataset])]
-  }
-}
-
-#' @keywords internal
-attributes_names <- function(x, attributes)   {
-
-  if ( is.character(attributes)) {
-    attributes <- which(names(x) %in% attributes)
-    not_in_dataset <- (! (which (attributes %in% names(x))))
-    if (length(not_in_dataset>0)) {
-      warning("attributes_get(x, attributes, ...): ", paste(attributes[not_in_dataset], collapse = ", "), " not in dataset.")
-    }
-  }
-
-  if ( is.numeric(attributes) ) {
-    not_in_dataset <- attributes[which (! attributes %in% 1:ncol(x))]
-
-    if (length(not_in_dataset>0)) {
-      warning("attributes_get(x, attributes, ...): ", paste(not_in_dataset, collapse = ", "), " not in dataset.")
-    }
-
-    if ( all(attributes %in% not_in_dataset)) {
-      stop("attributes_get(x, attributes, ...): none of ", paste(attributes, collapse = ", "), " in dataset.")
-    }
-    names(x)[as.integer(attributes[! attributes %in% not_in_dataset])]
-  }
-}
 
