+++ NAMESPACE
@@ -53,14 +53,12 @@ export(subject)
 export(subject_create)
 export(version)
 importFrom(assertthat,assert_that)
-importFrom(rlang,get_expr)
 importFrom(utils,bibentry)
 importFrom(utils,citation)
 importFrom(utils,download.file)
 importFrom(utils,object.size)
 importFrom(utils,person)
 importFrom(utils,read.csv)
-importFrom(utils,sessionInfo)
 importFrom(utils,toBibtex)
 importFrom(utils,write.csv)
 importFrom(utils,write.table)
