+++ README.md
@@ -117,7 +117,7 @@ dublincore(my_iris_dataset)
 #> [1] "American Iris Society"
 #> 
 #> $Date
-#> [1] "2022-08-17"
+#> [1] "2022-08-19"
 #> 
 #> $Creator
 #> [1] "Edgar Anderson [aut]"
