+++ DESCRIPTION
@@ -29,7 +29,6 @@ LazyData: true
 Imports: 
     assertthat,
     ISOcodes,
-    rlang,
     utils
 Suggests: 
     covr,
