### 2022-08-19 #553 @antaldaniel
@adamhsparks  I am much above your treshold, and apologies for the trivial error.  I wanted to omit some issues in the dataset() construtor, but I did not realize that it had some old code that had been rewritten - the test were omitting them, of course, but they sat at the bottom of the file.  It is now 81.2% covered, I know that it has to improve, but I'd prefer to do it when some issues are resolved in a clear direction (see my comment above.)
