### 2022-08-22 #553 @antaldaniel
For clarity, apart from adding the README badge, I made a few URL correction and added a paragraph to the Movtivation vignette.
