### 2022-08-22 e0ca094
Addressing a few CRAN check NOTES - changing the version to 0.1.7.3 - fixing URLs - adding a reference to surveydataset.dataobservatory.eu, an offspring specifically dedicated to social sciences data.
