### 2022-09-12 #553 @msperlin
---
title: "review"
  rmarkdown::md_document:
    pandoc_args: [
      "--wrap=none"
    ]
---

- [-] **Installation:** Installation succeeds as documented.
        - the install command should be `remotes::install_github('dataobservatory-eu/dataset')`
- [-] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.
        - see comments below for Readme.Rmd

---

Overall, I find the package interesting and well done. In my assessment, the R code is good and passes all CHECKS and tests. However, as with any work in development, there are some points to improve. Please find my suggestions next.

* **Package name**
        -I find the name of the package too generic and not informative. My suggestion is to name the package for something related to its context, such as "w3cdataset", "dataregistry", "metadata", or "datacube".

* **Readme.Rmd**:
        - the install command should be `remotes::install_github('dataobservatory-eu/dataset')`
        - I fell the motivation for the package in the Readme.Rmd is short. A new user that only reads the Reame.Rmd is probably going to have a hard time understanding why he/she should use the package for its own data. However, the vignettes have rich content and could be copied to the Readme.Rmd file. A couple of examples of applications from the vignettes would be useful in "selling" the package to the public. For example, applications of the package within companies or research institutions would be very interesting to read.

* **spelling**
        - On CHECK, the spelling package returns many spelling errors. You should check the vocabulary and add a text file at "inst" for removing the false-positives. More details [here](https://devguide.ropensci.org/building.html?q=spellin#testing).

* **printing**
        - the default print case for a "dataset" object is just the content and attributes of the object as a list (see first example at "Getting started" -- Readme.Rmd). Perhaps is worth a while to bring a new print method with package cli? The same is valid for function `dublincore`, which just prints the object as a list. While only a change in aesthetics, it might set the package apart from others and help attract new users.

* **function dataset_export**
        - why not use readr::write_csv for data.table::fwrite for reading the csv files? These are much faster and flexible than base::read.csv.
        - why we don't see a dataset_import_csv() function? I suspect many users will expect this function to be available.

* **function dataset()**
        - you should add some checking for major inputs. For example, input `x` should be of class _data.frame_ or _list_.

* **Vignettes**
        - /vignetts/datasets.Rmd: Line 87 show a code that is not executed or used (it actually fails to run).

* **unit tests**
        - test-dataset_download.R: since you need an internet connection, you should add testthat::skip_if_offline() at top of code.
