### 2022-09-13 #553 @antaldaniel
Thank you very much for the review.  You must see this package as a very early concept, and this will be the platform of many pacages of the rOpenGov collaboration (many of which could also be part of rOpenSci).

For me, the most important questions are what I put into the dilemmas:
- should the package build a new dataset class, and if yes, should it be inherited from data.frame or tibble?
-  or alternatively just create functions that can improve a data.frame, tibble, or data.table object.

The criteria for this is best integration with the rOpenSci packages that handle metadata, or create a binding to RDF, and connection to zen4R that uses R6 classes.

I do not find the w3c dataset or the datacube names very descriptive, because we will provide conversion to not only w3c but sdmx formats, and not only datacubes, but any dataset.

The smaller remarks on the functions are welcome, but again, these are just dummy functions to decide into which direction to develop the package to connect well with metadata, RDF, and open science deposition tools.
