### 2022-09-13 #553 @msperlin
Daniel,

> should the package build a new dataset class, and if yes, should it be inherited from data.frame or tibble? or alternatively just create functions that can improve a data.frame, tibble, or data.table object.

This is a design choice. Im my oppinion, you'll be fine in any direction, as long as the code/input is flexible towards existing classes.

> I do not find the w3c dataset or the datacube names very descriptive, because we will provide conversion to not only w3c but sdmx formats, and not only datacubes, but any dataset.

Ok.

> The smaller remarks on the functions are welcome, but again, these are just dummy functions to decide into which direction to develop the package to connect well with metadata, RDF, and open science deposition tools.

Fair enough. But, core functions will be mostly simillar in any design choice you make, so it seems right to find their best version right now. If you are not willing to change them because they are "dummy" functions, it would be nice to let us know before the review.
