### 2022-09-18 #553 @antaldaniel
I am a little bit puzzled, because you suggest to submit packages before they are on CRAN and developed to a detail that it is hard to make braking changes. So I did, and I provided an enormous amount of documentaiton, and development questions, particularly how to integrate best the package with `dataspice` and `rdflib`, which are rOpenSci packages, and our package will add a lot to their functionality, plus bridge them to several dozen rOpenGov packages.  .

https://dataset.dataobservatory.eu/articles/index.html
Particularly here: https://dataset.dataobservatory.eu/articles/motivation.html

As the package is intended to work well with two rOpenSci and several rOpenGov packages (which, in turn, would be very low hanging fruits to bring into rOpenSci as well, some of them are very well established packages with many thousand users in Europe), I really did not want to start to work out I/O function before clarifying what would be the best structure to import into.
