### 2022-09-18 #553 @antaldaniel
@msperlin Thank you for your comment on the unit test and the spelling, which I will implement now.  I will also create a new README.

This leads me to the comment of @melvidoni . This package is not supposed to work in isolation, as it is a metadata package, and the main development dilemma is how best connect to the `dataspice`, `rdflib`,and `zen4r` packages (these are rOpenSci packages) , as outlined in the documentation.  I see these as the real export routes, and not writing a csv file.

I believe that the other comments are all related to the basic question that I have asked in the documentation of the package, i.e. what should be the s3 class of the dataset open, which  as a consequences leads to the question of a suitable print method and a unit test for parameter x. I will just rephrase the dilemma that is implicilty left open, and unless there is any further comment on the issues below, I will leave the current design untouched: dataset() will be an s3 constructor inherited from data.frame, and I will remain as close to base R as possible.

**Checking input x** -- again this goes back to the dilemma of dependencies, and the main class, and I am not really sure what should be the limitations on the parameter x. @msperlin's suggestion seems to side with the argument to make it a data.frame like object. This is close to the current design, where dataset is an s3 class inherited from data.frame., and the test could be in this case for anything inherited from data.frame.

An alternative is not to make any real restrictions, and do not treat the datasets as an s3 class. In this case x can be any R object, however, this relates to a very basic design question: should the dataset be an s3 class with methods, including, of course, a nicer print method as you suggested, or it should work with any R objects?  Dataset can be placed into a list or a JSON object, a table, too.  There could be strong arguments to allow any R objet that has at least a two-dimensional structure to be x, i.e. the data container of the dataset. And the current quasi-methods of the dataset class remain helper functions.

**Import comments**
As for imports, and related to the README, the first use cases will be reproducible research packages, like the rOpenGov pacakges, which will hand over a tibble or a data.frame, not a file.  Further potential import routes could be the rOpenSci data access packages. The aim would be to make the output of such packages available for machine to machine communication on the web, hence improving interoperability, and also findability and reusability. This relates to the remarks in the review about 'selling' the package -- this package will be the foundation of many packages that have a wide use.  As this will be an upstream package, and a dependency of other pacakges, the README will mainly aim at package developers.

Re:  why we don't see a dataset_import_csv() ? Going back to the previous dilemma, this function in the current design would have no added value apart from possibly retainin the path where of the original csv file. The dataset() is an s3 class constructor that is not working from a file but an R object (which is most likely a data.frame).   I see value in importing if the source file has valuable metadata.  A good use case for this is the use of primary data collected by social science surveys. We are developing parallel a package for this, as in that case not only W3C/schema.org/SDMX but also DDI compliance is needed [DDI defines data documentation].  This is explained in some detail in the  [The survey Class](https://dataset.dataobservatory.eu/articles/survey.html) article.  The general question, again, if this is practical at all using s3 classes?

Tthere is already a csv based metadata package in rOpenSci, i.e. `dataspice`.  Connecting the two packages would be valuable indeed, because `dataspice` currently offers a not really reproducible, multiple-csv file based solution to the problems that dataset solves with the dataset() constructor.  The creators of  `dataspice`  use a different dataset definition, and this could be easily reconciled. The `dataspice` could be recommended for users who are seeking a spreadsheet based, manual solution and  `dataset` to those who seek a programmatic approach.  There are plenty of other synergies, `dataset` could rely on the html output of `dataspice`, while `dataset` could provide a release route for `dataspice` in machine-to-machine situations.

As for using readr as a dependency, I would like to keep the core dataset package as lean as possible, and rather reduce than increase the dependencies of the package.   I have an alternate version that could work with any i/o function treating the suggested readr::write_csv and data.table::fwrite as .f function parameters, similarly to purrr.  However, I stayed with base R for now as it is undecided what should be the expected class of x, and if the datasets should be an s3 class at all, and if yes, should they be tibbles or data frames.  Moving towards tidyverse would make the use of write_csv logical, of course.  But again, I think that the exporting should be possibly done by  `dataspice`, `rdflib`,and `zen4r`.  Currently the dataset is, after all, a data.frame that can be written into a file with any package that user likes.
