### 2022-09-19 #553 @romanflury
## Package Review

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   \[\] **Installation:** Installation succeeds as documented. as mentioned above, `remotes` does not export the function `install_package()` and `remotes::install_github('dataobservatory-eu/dataset')` should be used.
-   ☒ **Functionality:** Any functional claims of the software been confirmed.
-   ☒ **Performance:** Any performance claims of the software been confirmed.
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing:

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

In general, I like the idea of the package dataset and how the functions are implemented. In my opinion, it will be a valuable addition to the rOpenSci package network. Especially the manual page of each function is written in great detail and care, which eases the use of this package immensely. All major points from the checklist above are fulfilled.

Based on the author’s comments in this issue, it is my understanding that at this development state, not a full fletched review of this package is asked but the discussion of design and structure on a metalevel. In the following, you can find my personal suggestions and some minor comments for other improvements.

#### what should be the s3 class of the dataset

I support inheriting the S3 constructor from data.frame and keeping everything as close to base R as possible. In my opinion, this is the most flexible approach and the way to go if one seeks to have a generic package. But at the end of the day, this is probably also a matter of taste.

#### export routes

You mention this package is to be understood as a metadata package and defined as export routes the packages dataspice, rdflib,and zen4r. Maybe you could document how and where the dataset package fits into the entire process provided by these packages and what are best practices, e.g., in the spirit of <https://cran.r-project.org/web/packages/dataspice/vignettes/overview.html> section 2.

In the context of processing a dataset, do you think it is possible to streamline this process by concatenating multiple functions? Or do you see the package more as a toolbox of functions? Dependent on this answer, one could bring back the discussion about the name of the package. In my opinion, the name could be improved to indicate its specific purpose, as mentioned above by @msperlin.

#### dublincore\_add()

This function allows to specify the Creator attribute and thereby uses the function person(), which allows many different roles. However, for the attribute Creator this role should at least by default be “aut” or even limited to a few roles (<https://stat.ethz.ch/R-manual/R-devel/library/utils/html/person.html>).

#### geolocation()

It would be valuable if not only the geolocation as a string but also geoLocation points as described in <https://support.datacite.org/docs/datacite-metadata-schema-v44-recommended-and-optional-properties> section 18ff.

I hope this helps with deciding on the general strategy and continuing with the development of this package!
Best wishes, roman
