### 2022-09-19 #553 @antaldaniel
@romanflury Thank you very much, it is very helpful and will make the changes.
In general, this is a reverse package engineering as we are moving out from well-established packages a few functions that we think could support other similar packages.  So it is indeed a very early development phase, but with some mature elements that we would just like to generalize for a wider user base.

About your question to concatenating functions, if I understand correctly, we would like to do three things.

a) Genearlize the tidy dataset concept, and make sure that the user creates truly tidy dataset.  This may be easier done with dplyr and tidyr, and probably merit the switch to tibble to data.frame.  We have a bunch of packages at rOpenGov that already do this, and we want to create a 'container' for them.
b) Make sure that the dataset has enough metadata that it can truly be machine processed and reused (this very much overlaps with dataspice, but we use definitions of global libraries and statistical agencies, to me dataspice is more of a 'genereric' schema, probably corporate oriented. This is done, but as @msperlin pointed out, for example, it prints ugly.
c) Then send the data through the zen4r [into an open repository] or the rdflib serialization way [place it on a web 3.0 knowledge graph], or publish via dataspice [which is a good web 1.0 solution for the same.] The csv export is in fact an exception for manual release, such as upload to the Zenodo where zen4r provides a programatic upload.  Another easy export route is ddbplyr or rsqlite into relational databases.

The problem with data.frames, or generally, with a 'traditional' aproach of tables or databases (multiple tables) is that they are missing semantic information and various administrative information (reuse rights, authorship, etc.) that would make the data farmes reusable by another user or entity.  We want to create an object that 'carries'  a dataset in a way that it can be machine read, released, reused.

The reason why I prefer the more generic 'dataset' name is that we currently apply to important 'dataset' definitions, that of the W3C and SDMX, and we could easily add the third, i.e the schema.org definition used by `dataspice`.  So, after all, the package makes sure that the packaging of a dataset as stored in a data.frame, a data.table or a  tibble actually conforms the 'dataset' definitions of schema.org (general), SDMX (statistical agencies and applications) and W3C (semantic web, knowledge graphs), and the 'table' definition of relational databases.

The closest mental concept to what we are doing is the tidy dataset, but not so much on focusing on tidiness but on interoperabiilty and reusability helped by standard metadata and semantic description.  Of course, any reference to the tidyverse or tidyr would be rather superstitious.  The datacube is very particular to SDMX and has a bit of a1990s feeling.  The 'metadata' is not descriptive as there are many forms of metadata, and we do not only work with w3c dataset definitions, but SDMX, too, and eventually schema.org, too.  I actually had considered all data variations of @msperlin earlier, but I think that they are not descriptive enough. While the consensus is on an inherited s3 class from data.frame, in fact, the problem that we solve could be a problem of any R object that has a dataset structure, which means that it is organized in rows and columns, but lacks the documenation/metadata/semantic information to be cited, confirmed to be lawfully reused, and understood by a machine to import it into a database or find it.
