### 2022-09-20 #553 @melvidoni
Dear all,

First of all, thanks @msperlin and @romanflury for your thoughtful reviews.

@antaldaniel I took the discussion of this package to the Associate Editors, and we arrived to the following decisions:

Regarding package names, as long as a package name it's not offensive, you are allowed to name it as whatever you want. Nevertheless, `dataset' is a very broad term that also refers to a type of data. The latter may make it especially difficult for less experienced R users, and thus a more descriptive/creative name could possible bridge this gap.

I personally have a formal background in software engineering (meaning, I _am_ a software engineer). At first, I thought your idea of 'dummy functions' was a nice workaround to create 'interfaces' (in the sense of object-oriented programming), something I understand R does not provide (yet). Nevertheless, _ the peer review process is not equipped to develop a consensus on common standards across packages._

You established that this is "a metadata package that is in an early development page, still has development questions open". Moreover, you have also mentioned other packages you used as inspiration multiple times. Do note that _it is undesirable to look at packages that have been onboarded a long time ago, to justify their not meeting goals at submission_. This is mostly because the onboarding rules evolved over time, and the standards we have now, may not resemble those of "early-stage-rOpenSci".

Therefore, the final decision was to **pause the review process** to give @antaldaniel time to contact the authors of other packages. Do note that such feedback would be outside of this review. Until then, the review will be marked as `on-hold` on a following comment.
