### 2022-09-24 dataobservatory-eu/dataset#4 @antaldaniel
Nicer print method
the default print case for a "dataset" object is just the content and attributes of the object as a list (see first example at "Getting started" -- Readme.Rmd). Perhaps is worth a while to bring a new print method with package cli? The same is valid for function dublincore, which just prints the object as a list. While only a change in aesthetics, it might set the package apart from others and help attract new users.

[https://github.com/ropensci/software-review/issues/553#issuecomment-1244076662](https://github.com/ropensci/software-review/issues/553#issuecomment-1244076662)
