### 2022-09-24 dataobservatory-eu/dataset#6 @antaldaniel
dataset vignette line 87
- /vignetts/datasets.Rmd: Line 87 show a code that is not executed or used (it actually fails to run).
[software-review/issues/553](https://github.com/ropensci/software-review/issues/553#issuecomment-1244076662)
