### 2022-09-24 dataobservatory-eu/dataset#7 @antaldaniel
Reference rOpenSci Community Contributing Guide
including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).

[https://github.com/ropensci/software-review/issues/553#issuecomment-1244076662](https://github.com/ropensci/software-review/issues/553#issuecomment-1244076662)
