+++ man/language.Rd
@@ -16,7 +16,7 @@ language(x, iso_639_code = "639-3") <- value
 
 \item{value}{The language to be added to the object attributes, added by name, or
 as a 2- or 3-character code for the language. You can add a language code or language name,
-and the parameter is normalised to \code{tolower(language)}. (The ISO 639 standard capitalizes
+and the parameter is normalized to \code{tolower(language)}. (The ISO 639 standard capitalizes
 language names and uses lower case for the codes.)}
 }
 \value{
