### 2022-09-24 dataobservatory-eu/dataset#7 @antaldaniel
Inluded in 0.1.8
