### 2022-09-24 dataobservatory-eu/dataset#6 @antaldaniel
Inluded in 0.1.8
