### 2022-09-24 #553 @antaldaniel
Thank you very much for the comments and the useful experience.

I created the following milestones based on this review.

1. [First CRAN release](https://github.com/dataobservatory-eu/dataset/milestone/1) - all minor issues raised in this review are included [here](https://github.com/dataobservatory-eu/dataset/issues?q=is%3Aopen+is%3Aissue+milestone%3A%22First+CRAN+release%22). I solved the smaller, not conceptual issues and add @msperlin as reviewer contributor to [0.1.8](https://zenodo.org/record/7110256#.Yy8wSHZBzIU).
2. [rOpenGov review](https://github.com/dataobservatory-eu/dataset/milestone/2)

As several functions from heavily used rOpenGov packages are moving to dataset, this wil have to be done this year and that will create a basic usability, and it will at least lay out a blueprint to integrate with rOpenGov packages, which can be useful for any data retrieval packages in rOpenSci, too.

For the enhanced usability of the packages, I created these rOpenSci relevant milestones, which are meant to conceptually review how the packages can work together.

3. [Dataset on knowledge graphs](https://github.com/dataobservatory-eu/dataset/milestone/3)
4. Review the integration with [zen4r](https://github.com/eblondel/zen4R) and [dataspice](https://cran.r-project.org/web/packages/dataspice/index.html).
5. [Dataset and open science releases](https://github.com/dataobservatory-eu/dataset/milestone/4)

I expect this to be finished by the end of Q1 2023, when a draft concept paper will close this rather exploratory development phase.

We'd like to exchange data with statistical agencies and knowledge graphs on a large scale in the second half of 2023, I think that is when a first paper draft is due and the package will reach its more mature shape.
