### 2022-09-25 #553 @msperlin
Hi @antaldaniel,

Thanks. I appreciate the chance to contribute in making `datasets` better.

As for the future, it seems you still have some structural decisions to make and I'm sure you'll sort it out with time.

best,
