+++ man/dataset_export.Rd
@@ -51,7 +51,7 @@ my_iris_dataset <- dublincore_add(
 
 dataset_export_csv(my_iris_dataset, file = file.path(tempdir(), "my_iris.csv"))
 
-\dontrun{
+\donttest{
 read.csv(file.path(tempdir(), "my_iris.csv"), skip=20)
 read.csv(file.path(tempdir(), "my_iris.csv"))
 }
