+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Date: 2022-09-24
-Version: 0.1.8
+Date: 2022-12-01
+Version: 0.1.9
 Authors@R: 
     c(person(given = "Daniel", family = "Antal", 
            email = "daniel.antal@dataobservatory.eu", 
@@ -27,7 +27,7 @@ URL: https://github.com/dataobservatory-eu/dataset
 BugReports: https://github.com/dataobservatory-eu/dataset/issues
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.2
 Depends: 
     R (>= 2.10)
 LazyData: true
