+++ _pkgdown.yml
@@ -1,6 +1,10 @@
 url: ~
 template:
   bootstrap: 5
+  bslib:
+    base_font: {google: "Roboto"}
+    heading_font: {google: "Roboto Slab"}
+    code_font: {google: "JetBrains Mono"}
 
 home:
   links:
