+++ NEWS.md
@@ -1,4 +1,4 @@
-# dataset 0.1.8
+# dataset 0.1.9
 
 * Incorporating minor changes from the [rOpenSci](https://github.com/ropensci/software-review/issues/553) and CRAN peer-reviews.
 
