### 2022-12-04 b0ec2da
Print method works. Summary method does not work. Nicer printing for dublincore()
