+++ R/dublincore.R
@@ -103,7 +103,48 @@ dublincore <- function(x) {
   attributes_measures$class <- NULL
   attributes_measures$Size <- NULL
 
-  attributes_measures
+  simple_attributes <- attributes_measures
+  simple_attributes$dimensions <- NULL
+  simple_attributes$measures   <- NULL
+  simple_attributes$attributes <- NULL
+  simple_attributes$Type <- NULL
+  simple_attributes$Title <- NULL
+  simple_attributes$names <- NULL
+
+  cat(paste(paste0(names(unlist(attributes_measures$Title)), ": ", unlist(attributes_measures$Title)), collapse = " | "), "\n")
+  cat("Publiser: ", simple_attributes$Publisher, " | ")
+  cat("Source: ", simple_attributes$Source, " | ")
+  cat("Date: ", simple_attributes$Date, " | ")
+  cat("Language: ", simple_attributes$Language, " | ")
+  cat("Identifier: ", simple_attributes$Identifier, " | ")
+  cat("Rights: ", simple_attributes$Rights, " | ")
+  cat("Description: ", simple_attributes$Description, " | ")
+  cat("\n")
+
+  cat("names: ", paste(attributes_measures$names, collapse = ", "), "\n")
+
+  if(length(attributes_measures$dimensions$names)==0) {
+    cat("- dimensions: <none>\n")
+  } else {
+    cat("- dimensions:", paste0(attributes_measures$dimensions$names, " (", attributes_measures$dimensions$class, ") "), "\n")
+  }
+
+  if(length(attributes_measures$measures$names)==0) {
+    cat("- measures: <none>\n")
+  } else {
+    cat("- measures:", paste0(attributes_measures$measures$names, " (", attributes_measures$measures$class, ") "), "\n")
+  }
+
+  if(length(attributes_measures$attributes$names)==0) {
+    cat("- attributes: <none>\n")
+  } else {
+    cat("- attributes:", paste0(attributes_measures$attributes$names, " (", attributes_measures$attributes$class, ") "), "\n")
+  }
+
+
+
+
+  invisible(attributes_measures)
 }
 
 #' @rdname dublincore
@@ -122,11 +163,13 @@ dublincore_add <- function(x,
                            Relation = NULL,
                            Description = NULL,
                            Type = "DCMITYPE:Dataset",
-                           overwrite = TRUE) {
+                           overwrite = FALSE) {
 
   ## Set the Title property ------------------------------------------------
   if (is.null(attr(x, "Title"))) {
     dataset_title(x) <- Title
+  } else if ( is.na(attr(x, "Title")[1]) ) {
+    dataset_title(x, overwrite = TRUE) <- Title
   } else if ( overwrite ) {
     dataset_title(x, overwrite = TRUE) <- Title
   } else {
@@ -135,6 +178,8 @@ dublincore_add <- function(x,
 
   ## Set the Identifier property ---------------------------------------------
   if (is.null(attr(x, "Identifier"))) {
+    identifier(x) <- Identifier
+  } else if ( is.na(attr(x, "Identifier")) ) {
     identifier(x) <- Identifier
   } else if ( overwrite ) {
     identifier(x) <- Identifier
@@ -145,6 +190,8 @@ dublincore_add <- function(x,
   ## Set the Creator property ---------------------------------------------
   if (is.null(attr(x, "Creator"))) {
     creator(x) <- Creator
+  } else if ( is.na(attr(x, "Creator")) ) {
+    creator(x) <- Creator
   } else if ( overwrite ) {
     creator(x) <- Creator
   } else {
@@ -154,8 +201,10 @@ dublincore_add <- function(x,
   ## Set the Subject property ------------------------------------------------
[…]
