+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(print,dataset)
+S3method(summary,dataset)
 export("attributes_measures<-")
 export("creator<-")
 export("dataset_source<-")
