### 2022-12-04 dataobservatory-eu/dataset#4 @antaldaniel
0.1.9.0001 has a nicer print method.
