### 2022-12-10 4cf81b3
New replication dataset print, summary, subset, and [ methods.
