+++ R/iris_dataset.R
@@ -1,0 +1,25 @@
+#' Edgar Anderson's Iris Data
+#'
+#' This famous (Fisher's or Anderson's) iris data set gives the measurements in
+#' centimeters of the variables sepal length and width and petal length and
+#' width, respectively, for 50 flowers from each of 3 species of iris.
+#' The species are \emph{Iris setosa}, \emph{versicolor}, and \emph{virginica}.
+#' This is a replication of \code{datasets::\link[datasets:iris]{iris}} as
+#' \emph{dataset} s3 class.
+#'
+#' @details See \code{datasets::\link[datasets:iris]{iris}} for details.
+#'
+#' @format
+#' iris is a data frame with 150 cases (rows) and 5 variables (columns) named
+#' Sepal.Length, Sepal.Width, Petal.Length, Petal.Width, and Species.
+#' @references
+#' Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language.
+#' Wadsworth & Brooks/Cole.
+#' @source
+#' Fisher, R. A. (1936) The use of multiple measurements in taxonomic problems.
+#' Annals of Eugenics, \strong{7}, Part II, 179–188.
+#'
+#' The data were collected by Anderson, Edgar (1935).
+#' The irises of the Gaspe Peninsula,
+#' Bulletin of the American Iris Society, \strong{59}, 2–5.
+"iris_dataset"
