+++ R/dataset.R
@@ -85,8 +85,122 @@ dataset <- function(x,
   tmp_ds
 }
 
-#print   <- function(x, ...) { UseMethod("print", x)}
-#summary <- function(x, ...) { UseMethod("summary")}
+
+zzz <- function() {
+  new_dataset(y,
+              Dimensions = attr(x, "Dimensions"),
+              Measures = attr(x, "Dimensions"),
+              Attributes = attr(x, "Dimensions"),
+              sdmx_attributes = attr(x, "Dimensions"),
+              Title = dataset::dataset_title_create(paste0(attr(x, "Title")$Title[1] , " (subset)")),
+              Label = attr(x, "Label"),
+              Creator = attr(x, "Creator"),
+              Publisher = attr(x, "Publisher"),
+              Issued = attr(x, "Issued"),
+              Identifier = attr(x, "Identifier"),
+              Subject = attr(x, "Subject"),
+              Type = "DCMITYPE:Dataset" )
+}
+
+
+#' @rdname dataset
+#' @export
+is.dataset <- function(x) inherits(x, "dataset")
+
+
+#' @keywords internal
+new_dataset <- function(x,
+                        Dimensions = NULL,
+                        Measures = NULL,
+                        Attributes = NULL,
+                        sdmx_attributes = NULL,
+                        Title = NULL,
+                        Label = NULL,
+                        Creator = NULL,
+                        Publisher = NULL,
+                        Issued = NULL,
+                        Identifier = NULL,
+                        Subject = NULL,
+                        Source = NULL,
+                        Language = NULL,
+                        Format = NULL,
+                        Relation = NULL,
+                        Rights = NULL,
+                        Description = NULL,
+                        Type = "DCMITYPE:Dataset") {
+
+  stopifnot(inherits(x, "data.frame"))
+
+  tmp_ds <- dublincore_add(x,
+                           Title = Title,
+                           Creator = Creator,
+                           Identifier = Identifier,
+                           Publisher = Publisher,
+                           Subject = Subject,
+                           Language = Language,
+                           Format = Format,
+                           Rights = Rights,
+                           Relation = Relation,
+                           Description = Description,
+                           Type = Type,
+                           overwrite = TRUE)
+
+  if (is.null(Issued)) Issued <- Sys.Date()
+  attr(tmp_ds, "Date") <- Issued
+
+  if (!inherits(tmp_ds, "dataset")) {
+    attr(tmp_ds, "class") <- c("dataset", class(tmp_ds))
+  }
+
+  tmp_ds
+}
+
+#' @rdname dataset
+#' @export
+subset.dataset <- function(x, ...) {
+
+  subset_title <- paste0(dataset::dataset_title(x)$Title, " (subset)")
+  y <- NextMethod()
+  new_dataset(y,
+              Dimensions = attr(x, "Dimensions"),
+              Measures = attr(x, "Dimensions"),
+              Attributes = attr(x, "Dimensions"),
+              sdmx_attributes = attr(x, "Dimensions"),
+              Title = dataset_title_create(subset_title),
+              Label = attr(x, "Label"),
+              Creator = attr(x, "Creator"),
+              Publisher = attr(x, "Publisher"),
+              Issued = attr(x, "Issued"),
+              Identifier = attr(x, "Identifier"),
+              Subject = attr(x, "Subject"),
[…]
