+++ R/dataset_download.R
@@ -9,17 +9,16 @@
 #' invokes \code{utils::download.file}.
 #' @return A \code{\link{dataset}} with the downloaded dataset.
 #' @examples
-#'
-#' dest_file <- file.path(tempdir(), "5813772.csv")
-#'
 #' \donttest{
+#' dest_file <- file.path(tempdir(), "iris.csv")
 #' dataset_download(
-#'   url = "https://dataset.dataobservatory.eu/environmental_transfer_from_europe_test_dataset.csv",
-#'   Title = "Environmental Subsidies and Similar Transfers from Europe to the Rest of the World",
-#'   Dimensions = c("time", "geo"),
-#'   Measures = "value",
-#'   Attributes = c("unit", "obs_status", "method", "freq"),
-#'   Identifier = "https://doi.org/10.5281/zenodo.5813772",
+#'   url = "https://zenodo.org/record/7421899/files/iris.csv?download=1",
+#'   Dimensions = NULL,
+#'   Measures = c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width" ),
+#'   Attributes = "Species",
+#'   Title = "Iris Dataset",
+#'   Publisher = "American Iris Society",
+#'   Identifier = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x",
 #'   destfile = dest_file
 #' )
 #' }
