+++ DESCRIPTION
@@ -36,6 +36,7 @@ Imports:
     ISOcodes,
     utils
 Suggests: 
+    dataspice,
     covr,
     declared,
     dplyr,
