### 2022-12-10 13e9fa1
Add dataspice vignette
