+++ vignettes/dataspice.Rmd
@@ -14,43 +14,52 @@ knitr::opts_chunk$set(
 )
 ```
 
+The [ROpenSci](https://ropensci.org/) collective has an R package that aims to increase the metadata qualities of published datasets. Although the [dataspice](https://github.com/ropensci/dataspice) package is philosophically different—it is not aimed for programmatic use, and it does not use the DublinCore or DataCite metadata definitions used by libraries—it is easy to connect its functionality with `dataset` in a more programmatic way. 
+
 ```{r setup}
 library(dataset)
 library(here)
-description(iris_dataset) <- "This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris."
+description(iris_dataset) <- "This famous (Fisher's or Anderson's) iris data set 
+gives the measurements in centimeters of the variables sepal length and width 
+and petal length and width, respectively, for 50 flowers from each of 
+three species of iris."
 ```
 
+We create grammatically the auxiliary files of dataspice in the `tempdir()`, and then save the resulting JSON-LD file and the html overview on the `/datascpice` subdirectory. 
 
-```{r}
+```{r, eval=FALSE}
 library(dataspice)
-dataset_to_dataspice <- function(ds, 
-                                 fileName = "https://zenodo.org/record/7421899/files/iris.csv?download=1",
-                                 aux_path = file.path(tempdir()), 
-                                 geographicDescription = "Gaspe Peninsula, Canada",
-                                 dataspice_path = here::here("dataspice"), 
-                                 southBoundCoord = "47.976593", 
-                                 northBoundCoord = "49.2526866",
-                                 eastBoundCoord = "-64.504089", 
-                                 westBoundCoord = "-66.7947692",
-                                 startDate = "1935", 
-                                 endDate = "1935",
-                                 var_labels = c(
-                                   "Measurement of sepal lenght (cm)",
-                                   "Measurement of sepal width (cm)",
-                                   "Measurement of petal length (cm)",
-                                   "Measurement of petal width (cm)",
-                                   "Name of species, iris [species]"
-                                 ),
-                                 unitText = c(rep("centimeter", 4), NA_character_)
-                                 ) {
-access <- data.frame(
+dataset_to_dataspice <- function(
+    ds, 
+    fileName = "https://zenodo.org/record/7421899/files/iris.csv?download=1",
+    aux_path = file.path(tempdir()), 
+    geographicDescription = "Gaspe Peninsula, Canada",
+    dataspice_path = here::here("dataspice"), 
+    southBoundCoord = "47.976593", 
+    northBoundCoord = "49.2526866",
+    eastBoundCoord = "-64.504089", 
+    westBoundCoord = "-66.7947692",
+    startDate = "1935", 
+    endDate = "1935",
+    var_labels = c(
+      "Measurement of sepal lenght (cm)",
+      "Measurement of sepal width (cm)",
+      "Measurement of petal length (cm)",
+      "Measurement of petal width (cm)",
+      "Name of species, iris [species]"
+    ),
+    unitText = c(rep("centimeter", 4), NA_character_)
+) {
+  access <- data.frame(
   fileName = fileName,
   name = NA_character_,
   contentUrl	= NA_character_, 
   encodingFormat = NA_character_
 )  
 
[…]
