+++ README.Rmd
@@ -141,6 +141,8 @@ Our [datasets](https://dataset.dataobservatory.eu/reference/dataset.html):
 
 - [x] Correct exporting with FAIR metadata to the most used file formats and straightforward publication to open science repositories with correct bibliographical and use metadata. See [Export And Publish a dataset](https://dataset.dataobservatory.eu/articles/publish.html)
 
+- [x] Use programmatically the [dataspice](https://github.com/ropensci/dataspice) package to publish dataset documentation.
+
 - [x] Relatively lightweight in dependencies and easily works with data.frame, [tibble](https://tibble.tidyverse.org/) or [data.table](https://rstudio.github.io/DT/) R objects.
 
 
