+++ _pkgdown.yml
@@ -31,6 +31,11 @@ articles:
       - dataset
       - metadata
       - motivation
+  - title: "Compatibility"
+    desc: >
+     Works with dataspice.
+    contents:
+       - dataspice
   - title: "Planned functionality"
     desc: >
      This package is in an early development phase. The current dataset S3 class is inherited
