+++ R/dublincore.R
@@ -163,93 +163,116 @@ dublincore_add <- function(x,
                            overwrite = TRUE) {
 
   ## Set the Title property ------------------------------------------------
-  if (is.null(attr(x, "Title"))) {
-    dataset_title(x) <- Title
-  } else if ( is.na(attr(x, "Title")[1]) ) {
-    dataset_title(x, overwrite = TRUE) <- Title
-  } else if ( overwrite ) {
-    dataset_title(x, overwrite = TRUE) <- Title
-  } else {
-    message ("The dataset already has Title(s): ", dataset_title(x) )
+
+  if (! is.null(Title)) {
+    if (is.null(attr(x, "Title"))) {
+      dataset_title(x) <- Title
+    } else if ( is.na(attr(x, "Title")$Title) ) {
+      dataset_title(x, overwrite = TRUE) <- Title
+    } else if ( overwrite ) {
+      dataset_title(x, overwrite = TRUE) <- Title
+    } else {
+      message ("The dataset already has Title(s): ", dataset_title(x), ". Did not change (overwrite=FALSE)."  )
+    }
   }
 
   ## Set the Identifier property ---------------------------------------------
-  if (is.null(attr(x, "Identifier"))) {
-    identifier(x) <- Identifier
-  } else if ( is.na(attr(x, "Identifier")) ) {
-    identifier(x) <- Identifier
-  } else if ( overwrite ) {
-    identifier(x) <- Identifier
-  } else {
-    message ("The dataset already has an Identifier: ", identifier(x) )
+  if (! is.null(Identifier)) {
+    if (is.null(attr(x, "Identifier"))) {
+      identifier(x) <- Identifier
+    } else if ( is.na(attr(x, "Identifier")) ) {
+      identifier(x) <- Identifier
+    } else if ( overwrite ) {
+      identifier(x) <- Identifier
+    } else {
+      message ("The dataset already has an Identifier: ", identifier(x) )
+    }
+
   }
 
   ## Set the Creator property ---------------------------------------------
-  if (is.null(attr(x, "Creator"))) {
-    creator(x) <- Creator
-  } else if ( is.na(attr(x, "Creator")) ) {
-    creator(x) <- Creator
-  } else if ( overwrite ) {
-    creator(x) <- Creator
-  } else {
-    message ("The dataset already has a Creator: ",  creator(x) )
+  if (!is.null(Creator)) {
+    if (is.null(attr(x, "Creator"))) {
+      creator(x) <- Creator
+    } else if ( is.na(attr(x, "Creator")) ) {
+      creator(x) <- Creator
+    } else if ( overwrite ) {
+      creator(x) <- Creator
+    } else {
+      message ("The dataset already has a Creator: ",  creator(x) )
+    }
   }
+
 
   ## Set the Subject property ------------------------------------------------
-  if (is.null(attr(x, "Subject"))) {
-    subject(x) <- Subject
-  } else if ( is.na(attr(x, "Subject")) ) {
-    subject(x) <- Subject
-  } else if ( overwrite ) {
-    subject(x) <- Subject
-  } else {
-    message ("The dataset already has Subject(s): ", subject(x) )
+  if (! is.null(Subject)) {
+    if (is.null(attr(x, "Subject"))) {
+      subject(x) <- Subject
+    } else if ( is.na(attr(x, "Subject")) ) {
+      subject(x) <- Subject
+    } else if ( overwrite ) {
+      subject(x) <- Subject
+    } else {
+      message ("The dataset already has Subject(s): ", subject(x) )
+    }
   }
 
+
   ## Set the Source property ------------------------------------------------
-  if (is.null(attr(x, "Source"))) {
-    dataset_source(x) <- Source
- # } else if ( is.na(attr(x, "Subject")) ) {
[…]
