+++ R/source.R
@@ -1,8 +1,10 @@
 #' @title Get/set the Source of the object.
 #' @description Get/set the optional \code{Source} property as an attribute to an
-#' R object. DO not confuse with the base R \code{source()} function.
+#' R object. Do not confuse with the base R \code{source()} function.
 #' @details The \code{Source} is a related resource from which the described resource is
-#' derived. See \href{https://purl.org/dc/elements/1.1/source}{dct:source}.
+#' derived. See \href{https://purl.org/dc/elements/1.1/source}{dct:source}. In Datacite,
+#' the source is described by a \code{relatedIdentifierType} with the property
+#' \code{relationType="isDerivedFrom"}.
 #' @param value The \code{Source} as a character string of lengths one.
 #' @inheritParams dublincore
 #' @return The \code{Source} attribute as a character of length 1 is added to \code{x}.
