+++ man/dataset_source.Rd
@@ -23,11 +23,13 @@ The \code{Source} attribute as a character of length 1 is added to \code{x}.
 }
 \description{
 Get/set the optional \code{Source} property as an attribute to an
-R object. DO not confuse with the base R \code{source()} function.
+R object. Do not confuse with the base R \code{source()} function.
 }
 \details{
 The \code{Source} is a related resource from which the described resource is
-derived. See \href{https://purl.org/dc/elements/1.1/source}{dct:source}.
+derived. See \href{https://purl.org/dc/elements/1.1/source}{dct:source}. In Datacite,
+the source is described by a \code{relatedIdentifierType} with the property
+\code{relationType="isDerivedFrom"}.
 }
 \examples{
 iris_dataset <- iris

[…]
