+++ man/as.data.frame.Rd
@@ -1,0 +1,33 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dataset.R
+\name{as.data.frame}
+\alias{as.data.frame}
+\alias{as.data.frame.dataset}
+\title{Coerce dataset to a data frame}
+\usage{
+as.data.frame(x, row.names = NULL, optional = FALSE, ...)
+
+\method{as.data.frame}{dataset}(x, row.names = NULL, optional = FALSE, ...)
+}
+\arguments{
+\item{x}{A dataset.}
+
+\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}
+
+\item{optional}{logical. See also the \code{make.names} argument of the matrix method.}
+
+\item{...}{additional arguments to be passed to or from methods.}
+}
+\value{
+Coerces a dataset into a data.frame. The attributes are retained.
+}
+\description{
+Coerce a dataset into a data frame.
+}
+\seealso{
+Other dataset functions: 
+\code{\link{dataset_local_id}()},
+\code{\link{dataset_uri}()},
+\code{\link{dataset}()}
+}
+\concept{dataset functions}
