+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method("[",dataset)
+S3method(as.data.frame,dataset)
 S3method(print,dataset)
 S3method(subset,dataset)
 S3method(summary,dataset)
@@ -21,6 +22,7 @@ export("resource_type_general<-")
 export("rights<-")
 export("subject<-")
 export("version<-")
+export(as.data.frame)
 export(attributes_measures)
 export(bibentry_dataset)
 export(creator)
