+++ man/dataset_local_id.Rd
@@ -36,7 +36,6 @@ my_ds <- dataset (x = data.frame (
 }
 \seealso{
 Other dataset functions: 
-\code{\link{as.data.frame}()},
 \code{\link{dataset_uri}()},
 \code{\link{dataset}()}
 }
