+++ NEWS.md
@@ -1,4 +1,4 @@
-# dataset 0.1.9.0003
+# dataset 0.2.0
 
 * New methods for the `dataset()` s3 class: `print.dataset()`, `summary.dataset()`, `subset.dataset`, `[.dataset`.
 * New vignette on how to use the [dataspice](https://github.com/ropensci/dataspice) package programmatically for publishing dataset documentation.
