### 2023-03-18 2a8f67e
Do not evaluate on CRAN access to the Eurostat warehouse.
