+++ NEWS.md
@@ -1,3 +1,7 @@
+# dataset 0.2.1
+
+A minor correction to avoid vignettes downloading data from the Eurostat data warehouse on CRAN. Small readability improvements in the vignette articles.
+
 # dataset 0.2.0
 
 * New methods for the `dataset()` s3 class: `print.dataset()`, `summary.dataset()`, `subset.dataset`, `[.dataset`, `as.data.frame()`.
