### 2023-05-04 #553 @annakrystalli
Great to hear you have secured development funding! You are always welcome to advertise on the rOpenSci slack, especially in the `#jobs` channel. Blog posts are always a good idea but the rOpenSci blog is reserved for promoting packages once they have completed review so wouldn't be appropriate at this stage.
