### 2023-11-09 #553 @antaldaniel
After a very long time,  here is a conceptual working paper on the development with far more detailed specification than before, and some code ideas:

[Making Datasets Truly Interoperable in R](https://music.dataobservatory.eu/documents/open_music_europe/dataset-development/dataset-working-paper.html) is a working paper to accompany develop the package.

The working paper can be referenced with:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10091666.svg)](https://doi.org/10.5281/zenodo.10091666)

I am also looking for volunteer and potentially paid contributors to the package.

The source file is usually more recent:  [dataset-working-paper.qmd](https://github.com/dataobservatory-eu/dataset-development/stories/)`
