### 2023-11-10 #553 @annakrystalli
Thank you for the update @antaldaniel !

Good to hear you are making progress with the plans. Ultimately I feel the package will still remain on hold until it has been developed enough to be considered, if not ready, pretty close to release. That's when feedback from reviewers will be most useful and is also more aligned what is expected for reviewers to contribute their views on.

Let us know when you feel you have reached that stage!
