+++ NEWS.md
@@ -1,3 +1,6 @@
+# New devel branch
+
+
 # dataset 0.2.1
 
 A minor correction to avoid vignettes downloading data from the Eurostat data warehouse on CRAN. Small readability improvements in the vignette articles.
