+++ R/dataset.R
@@ -1,456 +1,140 @@
-#' @title Structure a data frame to dataset
-#' @description A DataSet is a collection of statistical data that corresponds to a defined structure.
-#' @details Loosely follows the \href{https://www.w3.org/TR/vocab-data-cube/}{The RDF Data Cube Vocabulary},
-#' but without the definition of data slices.\cr
-#' \code{\link{bibentry_dataset}} is a wrapper around \code{\link[utils:bibentry]{bibentry}} to correctly turn the
-#' metadata of the dataset into a bibentry object.\cr
-#' \code{as.data.frame} coerces a dataset into a data.frame in a way that the metadata attributes
-#' are retained.
-#' @param x A data.frame or inherited tibble, data.frame, or a structured list.
-#' @param Dimensions The name or column number of the dimensions within the dataset.
-#' @param Measures The name or column number of the measures within the dataset.
-#' @param Attributes The name or column number of the attributes within the dataset.
-#' @param sdmx_attributes The optional dimensions and attributes that conform with
-#' SDMX. \code{c("time", "geo")} will mark the "time" and "geo" attributes as conforming to
-#' sdmx. See \href{https://raw.githubusercontent.com/UKGovLD/publishing-statistical-data/master/specs/src/main/vocab/sdmx-attribute.ttl}{sdmx-attribute}.
-#' @param Label may be used to provide a human-readable version of the dataset's name.
-#' A text description (optionally with a language tag) as defined by
-#' \href{https://www.w3.org/TR/rdf-schema/#ch_label}{rdfs:label}.
-#' @inheritParams dublincore_add
-#' @param Subject Recommended for discovery in DataCite. Subject, keyword, classification code, or key
-#' phrase describing the resource. Similar to \href{https://purl.org/dc/elements/1.1/subject}{dct:subject}. \cr
-#' Use \code{\link{subject}} to properly add a key phrase from a controlled vocabulary
-#' and create structured Subject objects with \code{\link{subject_create}}.
-#' @param Type It is set by default to \href{https://purl.org/dc/dcmitype/Dataset}{DCMITYPE:Dataset}.
-#' @param Issued Corresponds to \href{https://purl.org/dc/elements/1.1/date}{dct:date}.
-#' @param ... Other parameters for the \code{print}, \code{summary} and \code{as.data.frame} methods.
-#' @importFrom utils toBibtex
-#' @return A data frame-like object with structural and referential metadata.
-#' @family dataset functions
+#' @title Create a dataset
+#'
+#' @param x An R object that contains the data of the dataset (a data.frame or
+#' inherited from data.frame, for example, tibble, tsibble, data.table).
+#' @param author A single person or a vector of persons as authors, declared with
+#' \code{\link[utils:person]{person}}.
+#' @param title The title of the dataset.
+#' @param identifier The permanent identifier, for example, the DOI of the dataset.
+#' If left empty (NULL), receives the DataCite standard value \code{':tba'}, or to
+#' be announced later.
+#' @param publisher The organisation or person that publishes the dataset.
+#' If left empty (NULL), receives the DataCite standard value \code{':tba'}, or to
+#' be announced later.
+#' @param year The year of the creation of the dataset. If left empty (NULL),
+#' the current year.
+#' @param version The version of the dataset. If left emtpy (NULL), defaults to
+#' \code{'0.1.0'}
+#' @param subject The subject of the data frame, as a subject type.
[…]
