+++ R/subject.R
@@ -1,95 +1,129 @@
-#' @title Get/Add subject(s) to a dataset
-#' @description Add one or more subject terms to the dataset's metadata.
-#' @details In the Dublin Core elements, \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#https://purl.org/dc/elements/1.1/subject}{dct::subject} is
-#' defined Typically, the subject will be represented using keywords, key phrases, or classification codes. It is
-#' recommended as a best practice to use a controlled vocabulary. \cr
-#' In DataCite, subjects are defined as key phrases from a controlled library.
-#' @param x An R object
-#' @param value Subject terms, or a Subject object created by \code{\link{subject_create}}.
-#' @param term A term, or a character vector of multiple terms.
-#' @param subjectScheme The scheme to which the term corresponds. If there are multiple terms,
-#' provide the subjectScheme(s) in the same order. Optional.
-#' @param schemeURI The URI(s) of the subject identifier scheme. If there are multiple terms,
-#' provide the schemeURIs in the same order as the terms. Optional.
-#' @param valueURI The URI of the subject term. If there are multiple terms,
-#' provide the valueURIs in the same order as the terms. Optional.
-#' @param overwrite Defaults to \code{FALSE}, in which case new subject(x) <- "Subject" calls
-#' are binding further Subjects to the already set Subject properties.
-#' @return The subjects as a data.frame of terms
+#' @title Create a subject
+#' @param term A subject term, for example, \code{"Data sets"}.
+#' @param subjectScheme The name of the subject scheme or classification code or
+#' authority if one is used. It is a namespace.
+#' @param schemeURI The URI of the subject identifier scheme, for example
+#' \code{"http://id.loc.gov/authorities/subjects"}
+#' @param valueURI The URI of the subject term.
+#' \code{"https://id.loc.gov/authorities/subjects/sh2018002256"}
+#' @param prefix An abbreviated prefix of a scheme URI, for example,
+#' \code{"lcch:"} representing \code{"http://id.loc.gov/authorities/subjects"}.
+#' Widely used namespaces (schemes) have conventional abbreviations.
+#' @param classificationCode The classificationCode subproperty may be used for
+#' subject schemes, like ANZSRC, which do not have valueURIs for each subject term.
+#' @return A named list with the subject term, the subject scheme, URIs and prefix.
 #' @examples
-#' x <- data.frame( geo = c("AL", "MK"),
-#'                 value = c(1,2))
-#' my_subject <- subject_create (
-#'                   term = c("R (Computer program language)",
-#'                            "Questionnaires--Computer programs"),
-#'                   subjectScheme = rep("LC Subject Headings", 2),
-#'                   schemeURI = rep("http://id.loc.gov/authorities/subjects",2),
-#'                   valueURI = c("https://id.loc.gov/authorities/subjects/sh2002004407.html",
-#'                                "http://id.worldcat.org/fast/1085693/")
-#' )
-#'  subject(x) <- my_subject
-#'  subject(x)
+#' Subject ("Dataset",
+#'         schemeURI = "http://id.loc.gov/authorities/subjects",
+#'         valueURI = "https://id.loc.gov/authorities/subjects/sh2018002256",
+#'         subjectScheme = "LCCH",
+#'         prefix = "lcch:")
 #'
-#'  y <- data.frame()
-#'  subject(y) <- "R (Computer program language)"
-#'  subject(y) <- "Questionnaires--Computer programs"
-#'  subject(y)
+#'  ds <- dataset(iris,
+#'         title = "The iris Dataset",
+#'         author = c(
+#'            person(family ="Anderson",
+#'            given ="Edgar",
+#'            role = "aut")
+#'            ),
[…]
