+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 S3method(as_dataset,data.frame)
 S3method(describe,dataset)
+S3method(is.subject,subject)
 S3method(print,dataset)
 S3method(summary,dataset)
 export("creator<-")
@@ -14,8 +15,8 @@ export("language<-")
 export("publication_year<-")
 export("publisher<-")
 export("rights<-")
-export("subject<-")
 export("version<-")
+export(Subject)
 export(as_dataset)
 export(bibentry_dataset)
 export(creator)
@@ -42,8 +43,6 @@ export(publisher)
 export(related_item)
 export(rights)
 export(size)
-export(subject)
-export(subject_create)
 export(version)
 importFrom(assertthat,assert_that)
 importFrom(stats,setNames)

[…]
