### 2023-11-17 4057ae3
dataset_bibentry() function
