+++ NAMESPACE
@@ -23,6 +23,7 @@ export(creator)
 export(datacite)
 export(datacite_add)
 export(dataset)
+export(dataset_bibentry)
 export(dataset_download)
 export(dataset_export)
 export(dataset_export_csv)
