### 2023-11-20 5770173
bibentry modifications
