+++ R/dataset_bibentry.R
@@ -22,10 +22,13 @@
 #'          url = "https://en.wikipedia.org/wiki/Iris_flower_data_set",
 #'          resourceType = "Dataset"
 #'          )
+#'
 #' print(ds, style="text")
 #' print(ds, style="Bibtex")
 #' @export
 
 dataset_bibentry <- function(ds) {
+
   attr(ds, "DataBibentry")
+
 }
