+++ tests/testthat/test-as_bibentry.R
@@ -1,0 +1,19 @@
+
+
+myBibentry <- as_bibentry(iris,
+                          author = c(person(family ="Anderson", given ="Edgar",
+                                            role = "aut")),
+                          title = "The iris Dataset")
+
+
+print(myBibentry, "Bibtex")
+
+test_that("as_bibentry() internal function works", {
+  expect_equal(myBibentry$title, "The iris Dataset")
+  expect_equal(myBibentry$language, ":unas")
+  expect_equal(myBibentry$publisher, ":tba")
+  expect_equal(myBibentry$identifier, ":tba")
+  expect_equal(myBibentry$version, "0.1.0")
+  expect_equal(myBibentry$description, ":unas")
+  expect_equal(myBibentry$year, substr(as.character(Sys.Date()),1,4))
+})
