+++ man/dataset_bibentry.Rd
@@ -34,6 +34,7 @@ ds <- dataset(iris,
          url = "https://en.wikipedia.org/wiki/Iris_flower_data_set",
          resourceType = "Dataset"
          )
+
 print(ds, style="text")
 print(ds, style="Bibtex")
 }
