### 2023-11-20 9d3b173
Authorship with creator() and unit tests
