+++ man/creator.Rd
@@ -5,19 +5,19 @@
 \alias{creator<-}
 \title{Get/set the Creator of the object.}
 \usage{
-creator(x)
+creator(ds)
 
-creator(x, overwrite = TRUE) <- value
+creator(ds, overwrite = TRUE) <- value
 }
 \arguments{
-\item{x}{An R object, such as a data.frame, a tibble, or a data.table.}
-
 \item{overwrite}{If the attributes should be overwritten. In case it is set to \code{FALSE},
 it gives a message with the current \code{Creator} property instead of overwriting it.
 Defaults to \code{TRUE} when the attribute is set to \code{value} regardless of previous
 setting.}
 
 \item{value}{The \code{Creator} as a \code{utils::\link[utils]{person}} object.}
+
+\item{x}{An R object, such as a data.frame, a tibble, or a data.table.}
 }
 \value{
 The Creator attribute as a character of length 1 is added to \code{x}.
@@ -37,9 +37,17 @@ code, use the property Contributor/contributorType/hostingInstitution for the co
 repository.
 }
 \examples{
-iris_dataset <- iris
-creator(iris_dataset) <- person("Anderson", "Edgar", role = "aut")
-creator(iris_dataset)
+ds <- dataset(iris,
+              title = "The iris Dataset",
+              author = c(
+                person(family ="Anderson",
+                       given ="Edgar",
+                       role = "aut")
+              ),
+              identifier = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x",
+              year = "1935",
+              version = "1.0",
+              description = "The famous dataset that is distributed with R.")
 }
 \seealso{
 Other Reference metadata functions: 
