### 2023-11-21 c841910
Create a datacite s3 class with DataCite bibliographic entries
