+++ R/datacite.R
@@ -1,7 +1,7 @@
-#' @title Add DataCite metadata to an object
+#' @title Add DataCite metadata to a bibentry object
 #'
 #' @description Add metadata conforming the \href{https://schema.datacite.org/}{DataCite Metadata Schema}
-#' to datasets, i.e. structured R data.frame or list objects, for an accurate and consistent identification
+#' to a \code{utils::\link[utils]{bibentry}} object, for an accurate and consistent identification
 #' of a resource for citation and retrieval purposes.
 #'
 #' @details DataCite is a leading global non-profit organisation that provides persistent identifiers
@@ -72,15 +72,14 @@
 #' being registered.
 #' @param overwrite If pre-existing metadata properties should be overwritten,
 #' defaults to \code{TRUE}.
-#' @return An R object with at least the mandatory DataCite attributes.
-#' @importFrom utils person
+#' @return A \code{utils::\link[utils]{bibentry}} object DataCite attributes.
+#' @importFrom utils person bibentry
 #' @source \href{https://support.datacite.org/docs/schema-mandatory-properties-v43}{DataCite 4.3 Mandatory Properties} and
 #' \href{https://support.datacite.org/docs/schema-optional-properties-v43}{DataCite 4.3 Optional Properties}
 #' @family metadata functions
 #' @export
 #' @examples
-#' my_iris <- datacite_add(
-#'    x = iris,
+#' datacite(
 #'    Title = "Iris Dataset",
 #'    Creator = person(family = "Anderson", given = "Edgar", role = "aut"),
 #'    Publisher = "American Iris Society",
@@ -88,146 +87,96 @@
 #'    Geolocation = "US",
 #'    Language = "en")
 #'
-#' datacite(my_iris)
 
-datacite <- function(x) {
+datacite <- function(Title,
+                     Creator,
+                     Identifier = NULL,
+                     Publisher = NULL,
+                     PublicationYear = NULL,
+                     Subject = NULL,
+                     Type = "Dataset",
+                     Contributor = NULL,
+                     DateList = NULL,
+                     Language = NULL,
+                     AlternateIdentifier = NULL,
+                     RelatedIdentifier = NULL,
+                     Format = NULL, Version = NULL, Rights = NULL,
+                     Description = NULL, Geolocation = NULL,
+                     FundingReference = NULL) {
 
-  attributes_measures <- attributes(x)
-  attributes_measures$row.names <- NULL
-  attributes_measures$class <- NULL
+  DateList <- ifelse (is.null(Date), ":tba", as.character(DateList))
+  Format <- ifelse (is.null(Format), ":tba", as.character(Format))
+  AlternateIdentifier <- ifelse (is.null(AlternateIdentifier), ":unas", AlternateIdentifier)
+  RelatedIdentifier <- ifelse (is.null(RelatedIdentifier), ":unas", RelatedIdentifier)
+  Rights <- ifelse (is.null(Rights), ":tba", as.character(Rights))
+  Geolocation <- ifelse (is.null(Geolocation), ":unas", as.character(Geolocation))
+  FundingReference <- ifelse (is.null(FundingReference), ":unas", as.character(FundingReference))
 
-  attributes_measures
+  new_datacite(Title = Title,
+               Creator = Creator,
+               Identifier = Identifier,
+               Publisher = Publisher,
+               PublicationYear = PublicationYear,
+               Subject = Subject,
+               Type = "Dataset",
+               Contributor = Contributor,
+               DateList = DateList,
+               Language = Language,
+               AlternateIdentifier = AlternateIdentifier,
+               RelatedIdentifier = RelatedIdentifier,
+               Forma = Format,
+               Version = Version,
+               Rights = Rights,
+               Description = Description,
[…]
