+++ tests/testthat/test-datacite.R
@@ -1,11 +1,4 @@
-my_ds <- dataset (data.frame(sex = c("M", "F"),
-                    value = c(1,2)
-                    ),
-         Measures = "value",
-         Dimensions = "sex")
-
-iris_ds <- datacite_add(
-   x = iris,
+iris_datacite <- datacite(
    Title = "Iris Dataset",
    Creator = person(family="Anderson", given ="Edgar", role = "aut"),
    Publisher = "American Iris Society",
@@ -14,7 +7,8 @@ iris_ds <- datacite_add(
    Language = "en")
 
 test_that("datacite() works", {
-  expect_equal(datacite(iris_ds)$Language, 'eng')
-  expect_equal(datacite(iris_ds)$Creator, person ( given = "Edgar", family = "Anderson", role = "aut"))
-  expect_true(grepl("KiB", attr(iris_ds, "Size")))
+  expect_true(is.datacite(iris_datacite))
+  expect_equal(iris_datacite$language, 'en')
+  expect_equal(iris_datacite$geolocation, 'US')
+  expect_equal(iris_datacite$rights, ':tba')
 })
