+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 S3method(as_dataset,data.frame)
 S3method(describe,dataset)
+S3method(is.datacite,datacite)
 S3method(is.subject,subject)
 S3method(print,dataset)
 S3method(summary,dataset)
@@ -21,7 +22,6 @@ export(as_dataset)
 export(bibentry_dataset)
 export(creator)
 export(datacite)
-export(datacite_add)
 export(dataset)
 export(dataset_bibentry)
 export(dataset_download)

[…]
