### 2023-11-21 42e8438
create a DublinCore S3 class with methods
