+++ R/dublincore.R
@@ -1,6 +1,7 @@
-#' @title Add DublinCore metadata to an object
+#' @title Add DublinCore metadata to a bibentry object
 #'
-#' @description Add metadata conforming the \href{https://schema.datacite.org/}{DataCite Metadata Schema}
+#' @description Add metadata conforming the
+#' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/format/}{DCMI Metadata Terms}.
 #' to datasets, i.e. structured R data.frame or list objects, for an accurate and consistent identification
 #' of a resource for citation and retrieval purposes.
 #'
@@ -18,11 +19,11 @@
 #' structured R list.
 #' @details The \code{ResourceType} property will be by definition "Dataset".
 #' The \code{Size} attribute (e.g. bytes, pages, inches, etc.) will automatically added to the dataset.
-#' @param Title \href{https://purl.org/dc/elements/1.1/title}{dct:title}, a name given to the resource.
+#' @param title \href{https://purl.org/dc/elements/1.1/title}{dct:title}, a name given to the resource.
 #' \code{\link{datacite}} allows the use of alternate titles, too. See \code{\link{dataset_title}}.
-#' @param Creator An entity primarily responsible for making the resource. \href{https://purl.org/dc/elements/1.1/creator}{dct:creator}
+#' @param creator An entity primarily responsible for making the resource. \href{https://purl.org/dc/elements/1.1/creator}{dct:creator}
 #' Corresponds to \code{Creator} in \code{\link{datacite}}. See \code{\link{creator}}.
-#' @param Identifier An unambiguous reference to the resource within a given context.
+#' @param identifier An unambiguous reference to the resource within a given context.
 #' Recommended practice is to identify the resource by means of a string conforming to an
 #' identification system. Examples include International Standard Book Number (ISBN),
 #' Digital Object Identifier (DOI), and Uniform Resource Name (URN).
@@ -30,7 +31,7 @@
 #' \href{https://www.ukoln.ac.uk/metadata/dcmi-ieee/identifiers/index.html}{registered URI schemes maintained by IANA}.
 #' More details: \href{https://www.ukoln.ac.uk/metadata/dcmi-ieee/identifiers/}{Guidelines for using resource identifiers in Dublin Core metadata and IEEE LOM}.
 #' Similar to \code{Identifier} in \code{\link{datacite}}. See \code{\link{identifier}}.
-#' @param Publisher Corresponds to \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#publisher}{dct:publisher}
+#' @param oublisher Corresponds to \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#publisher}{dct:publisher}
 #' and Publisher in DataCite.
 #' The name of the entity that holds, archives, publishes prints, distributes, releases,
 #' issues, or produces the resource. This property will be used to formulate the citation,
@@ -39,15 +40,15 @@
 #' "holds, archives, publishes, prints, distributes, releases, issues, or produces" the
 #' code, use the property Contributor/contributorType/hostingInstitution for the code
 #' repository. See \code{\link{publisher}}.
-#' @param Subject In \href{https://purl.org/dc/elements/1.1/subject}{dct:subject}. In
+#' @param subject In \href{https://purl.org/dc/elements/1.1/subject}{dct:subject}. In
 #' \code{\link{datacite}} it is a recommended property for discovery. In DataCite, a more complex
 #' referencing is used. See \code{\link{subject}} and create structured Subject objects with
 #' \code{\link{subject_create}}.
-#' @param Date Corresponds to a point or period of time associated with an event in the
[…]
