+++ NAMESPACE
@@ -3,6 +3,7 @@
 S3method(as_dataset,data.frame)
 S3method(describe,dataset)
 S3method(is.datacite,datacite)
+S3method(is.dublincore,dublincore)
 S3method(is.subject,subject)
 S3method(print,dataset)
 S3method(summary,dataset)
@@ -34,7 +35,6 @@ export(dataset_title_create)
 export(dataset_uri)
 export(description)
 export(dublincore)
-export(dublincore_add)
 export(geolocation)
 export(identifier)
 export(is.dataset)

[…]
