### 2023-11-21 212a2a8
Unit testing subject field and Dublin Core object.
