+++ man/subject.Rd
@@ -1,32 +1,32 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/Subject.R
-\name{Subject}
-\alias{Subject}
+% Please edit documentation in R/subject.R
+\name{subject}
 \alias{subject}
 \alias{subject<-}
 \alias{is.subject}
 \alias{is.subject.subject}
 \title{Create a subject}
 \usage{
-Subject(
-  term,
-  schemeURI = NULL,
-  valueURI = NULL,
-  prefix = NULL,
-  subjectScheme = NULL,
-  classificationCode = NULL
-)
+subject(ds)
 
 subject(ds)
 
-subject(ds, subject) <- value
+subject(ds) <- value
 
 is.subject(x)
 
 \method{is.subject}{subject}(x)
 }
 \arguments{
+\item{ds}{A dataset which has or needs a subject field in the metadata
+attributes.}
+
+\item{x}{An object that is tested if it has a class "subject".}
+
 \item{term}{A subject term, for example, \code{"Data sets"}.}
+
+\item{subjectScheme}{The name of the subject scheme or classification code or
+authority if one is used. It is a namespace.}
 
 \item{schemeURI}{The URI of the subject identifier scheme, for example
 \code{"http://id.loc.gov/authorities/subjects"}}
@@ -38,18 +38,10 @@ is.subject(x)
 \code{"lcch:"} representing \code{"http://id.loc.gov/authorities/subjects"}.
 Widely used namespaces (schemes) have conventional abbreviations.}
 
-\item{subjectScheme}{The name of the subject scheme or classification code or
-authority if one is used. It is a namespace.}
-
 \item{classificationCode}{The classificationCode subproperty may be used for
 subject schemes, like ANZSRC, which do not have valueURIs for each subject term.}
 
-\item{ds}{A dataset which has or needs a subject field in the metadata
-attributes.}
-
-\item{subject}{A subject field created by \code{\link{Subject}}.}
-
-\item{x}{An object that is tested if it has a class "subject".}
+\item{subject}{A subject field created by \code{\link{subject}}.}
 }
 \value{
 A named list with the subject term, the subject scheme, URIs and prefix.
@@ -60,7 +52,7 @@ The Subject field of the dataset's attributes.
 Create a subject
 }
 \examples{
-Subject ("Dataset",
+subject ("Dataset",
         schemeURI = "http://id.loc.gov/authorities/subjects",
         valueURI = "https://id.loc.gov/authorities/subjects/sh2018002256",
         subjectScheme = "LCCH",
