+++ NAMESPACE
@@ -18,9 +18,7 @@ export("publication_year<-")
 export("publisher<-")
 export("rights<-")
 export("version<-")
-export(Subject)
 export(as_dataset)
-export(bibentry_dataset)
 export(creator)
 export(datacite)
 export(dataset)
@@ -44,11 +42,11 @@ export(publisher)
 export(related_item)
 export(rights)
 export(size)
+export(subject)
 export(version)
 importFrom(assertthat,assert_that)
 importFrom(stats,setNames)
 importFrom(utils,bibentry)
-importFrom(utils,citation)
 importFrom(utils,download.file)
 importFrom(utils,object.size)
 importFrom(utils,person)
