### 2023-11-21 90880b3
Removing old functions that are no longer in use
