+++ NEWS.md
@@ -1,5 +1,9 @@
 # New devel branch
 
+The devel branch contains new code that is not is validated, but as a whole the package is not working consistently.
+
+0.2.2: New s3 classes for DataCite and Dublin Core bibliographic entries.
+
 
 # dataset 0.2.1
 
