+++ R/dataset_bibentry.R
@@ -5,6 +5,8 @@
 #' For further information on
 #' the \code{bibentry} class see \code{utils::\link[utils]{bibentry}}.
 #' @param ds A dataset object created by \code{\link{dataset}}.
+#' @param style Defaults to \code{"text"}, alternatives: \code{"Bibtex"},
+#' \code{"citation"}, \code{"html"}, \code{"R"}.
 #' @return The \code{\link[utils]{bibentry}} object of the dataset.
 #' @importFrom utils bibentry
 #' @examples
@@ -23,12 +25,19 @@
 #'          resourceType = "Dataset"
 #'          )
 #'
-#' print(ds, style="text")
-#' print(ds, style="Bibtex")
+#' dataset_bibentry(ds, style="text")
+#' dataset_bibentry(ds, style="Bibtex")
 #' @export
 
-dataset_bibentry <- function(ds) {
+dataset_bibentry <- function(ds, style = "text") {
 
-  attr(ds, "DataBibentry")
+  DataBibentry <- attr(ds, "DataBibentry")
 
+  if( ! style %in% c("Bibtex", "citation", "html", "latex", "R", "text") ) {
+    message('dataset_bibentry(ds, style="..." must be any of "text", "Bibtex", "citation",\n"html", "latex", "R".  Using "text".')
+    DataBibentry
+  } else {
+    print(DataBibentry, style = style)
+    invisible(DataBibentry)
+  }
 }
