+++ NAMESPACE
@@ -8,8 +8,8 @@ S3method(is.subject,subject)
 S3method(print,dataset)
 S3method(summary,dataset)
 export("creator<-")
-export("dataset_source<-")
 export("dataset_title<-")
+export("datasource<-")
 export("description<-")
 export("geolocation<-")
 export("identifier<-")
@@ -27,10 +27,10 @@ export(dataset_download)
 export(dataset_export)
 export(dataset_export_csv)
 export(dataset_local_id)
-export(dataset_source)
 export(dataset_title)
 export(dataset_title_create)
 export(dataset_uri)
+export(datasource)
 export(description)
 export(dublincore)
 export(geolocation)
