### 2023-11-29 69a30d7
Adding new unit tests and documentation
