+++ man/size.Rd
@@ -29,7 +29,7 @@ attr(iris_dataset, "Size")
 \seealso{
 Other Reference metadata functions: 
 \code{\link{creator}()},
-\code{\link{dataset_source}()},
+\code{\link{datasource}()},
 \code{\link{description}()},
 \code{\link{geolocation}()},
 \code{\link{identifier}()},
