+++ man/description.Rd
@@ -38,7 +38,7 @@ description(iris_dataset)
 \seealso{
 Other Reference metadata functions: 
 \code{\link{creator}()},
-\code{\link{dataset_source}()},
+\code{\link{datasource}()},
 \code{\link{geolocation}()},
 \code{\link{identifier}()},
 \code{\link{language}},
