+++ man/datasource.Rd
@@ -1,20 +1,13 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/source.R
-\name{dataset_source}
-\alias{dataset_source}
-\alias{dataset_source<-}
+% Please edit documentation in R/datasource.R
+\name{datasource}
+\alias{datasource}
 \title{Get/set the Source of the object.}
 \usage{
-dataset_source(x)
-
-dataset_source(x, overwrite = TRUE) <- value
+datasource(x)
 }
 \arguments{
-\item{x}{An R object of type data.frame, or inherited data.table, tibble; alternatively a well
-structured R list.}
-
-\item{overwrite}{If pre-existing metadata properties should be overwritten,
-defaults to \code{TRUE}.}
+\item{x}{An object that is tested if it has a class "dublincore".}
 
 \item{value}{The \code{Source} as a character string of lengths one.}
 }
@@ -32,9 +25,8 @@ the source is described by a \code{relatedIdentifierType} with the property
 \code{relationType="isDerivedFrom"}.
 }
 \examples{
-iris_dataset <- iris
-dataset_source(iris_dataset) <- "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x"
-dataset_source(iris_dataset)
+datasource(iris_dataset) <- "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x"
+datasource(iris_dataset)
 }
 \seealso{
 Other Reference metadata functions: 
