### 2023-11-29 c37ebb6
Title setting with Bibentry objects requires a different solution, and alternative titles are not continued in this function.
