+++ NAMESPACE
@@ -28,7 +28,6 @@ export(dataset_export)
 export(dataset_export_csv)
 export(dataset_local_id)
 export(dataset_title)
-export(dataset_title_create)
 export(dataset_uri)
 export(datasource)
 export(description)
