### 2023-11-29 0d05fe7
Update of manuals
