+++ man/dataset_title.Rd
@@ -56,3 +56,4 @@ dataset_title(my_iris)
  dataset_title(y) <- "Questionnaires--Computer programs"
  dataset_title(y)
 }
+\keyword{internal}
