+++ man/dataset_bibentry.Rd
@@ -4,10 +4,13 @@
 \alias{dataset_bibentry}
 \title{Get the bibliographic entries of a dataset}
 \usage{
-dataset_bibentry(ds)
+dataset_bibentry(ds, style = "text")
 }
 \arguments{
 \item{ds}{A dataset object created by \code{\link{dataset}}.}
+
+\item{style}{Defaults to \code{"text"}, alternatives: \code{"Bibtex"},
+\code{"citation"}, \code{"html"}, \code{"R"}.}
 }
 \value{
 The \code{\link[utils]{bibentry}} object of the dataset.
@@ -35,6 +38,6 @@ ds <- dataset(iris,
          resourceType = "Dataset"
          )
 
-print(ds, style="text")
-print(ds, style="Bibtex")
+dataset_bibentry(ds, style="text")
+dataset_bibentry(ds, style="Bibtex")
 }
