### 2023-12-02 b540d24
Work with the DataStructureDefinition
