+++ man/initialise_dsd.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/initialise_dsd.R
+\name{initialise_dsd}
+\alias{initialise_dsd}
+\alias{initialize_dsd}
+\title{Initialise a DataStructure (internal)}
+\usage{
+initialise_dsd(df, col)
+
+initialize_dsd(df, col)
+}
+\arguments{
+\item{df}{An object inherited from data.frame.}
+
+\item{col}{The column number that needs to be initialised.}
+}
+\value{
+A list containing the data structure definition.
+}
+\description{
+Initialise a DataStructure (internal)
+}
+\keyword{internal}
