### 2023-12-04 6928c8b
Updated rights, publishing metadata management.
