+++ R/related_item.R
@@ -8,7 +8,6 @@
 #' @param schemeURI  See \href{https://support.datacite.org/docs/datacite-metadata-schema-v44-recommended-and-optional-properties#12d-schemeuri}{schemeURI}.
 #' @param schemeType  See \href{https://support.datacite.org/docs/datacite-metadata-schema-v44-recommended-and-optional-properties#12e-schemetype}{schemeType}.
 #' @param resourceTypeGeneral The general type of a resource or file.
-#' See \code{\link{resource_type_general}} for allowed values and validation.
 #' @param Volume The volume of the related item (optional).
 #' @param Issue The issue number of the related item (optional).
 #' @param Edition The edition of the related item (optional).
