### 2023-12-04 b04dbb3
New unit tests for title, publisher, rights.
