+++ tests/testthat/test-resource_type.R
@@ -1,15 +1,0 @@
-x <- data.frame()
-resource_type(x) <- "Dataset"
-
-y <- data.frame()
-resource_type(y) <- "Census Data"
-
-test_that("resource_type() works", {
-  expect_equal(resource_type(y)$resourceType, "Census Data")
-  expect_equal(resource_type(x)$resourceType, "Dataset")
-  expect_equal(resource_type(y)$resourceTypeGeneral, "Dataset")
-  expect_equal(attr(y, "Type"), list(
-    resourceType = "Census Data",
-    resourceTypeGeneral = "Dataset" ))
-})
-
