### 2023-12-04 82811bb
All Dublin Core metadata functions work
