+++ man/language.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/Language.R
+% Please edit documentation in R/language.R
 \name{language}
 \alias{language}
 \alias{language<-}

[…]
