+++ R/dataset_download.R
@@ -2,9 +2,7 @@
 #' @description A wrapper around \code{\link{dataset}} and
 #' an import function.
 #' @importFrom utils download.file read.csv
-#' @inheritParams utils::download.file
 #' @inheritParams dataset
-#' @inheritParams dublincore
 #' @param type A file type.  Currently only \code{csv} is implemented that
 #' invokes \code{utils::download.file}.
 #' @return A \code{\link{dataset}} with the downloaded dataset.
@@ -25,72 +23,45 @@
 #' @export
 
 
-dataset_download  <- function(Title,
-                              Dimensions = NULL,
-                              Measures = NULL,
-                              Attributes = NULL,
-                              Identifier = NULL,
-                              url,
+dataset_download  <- function(url,
+                              title,
+                              author,
+                              destfile,
                               type = "csv", ...) {
 
-  dataset_download_csv(url = url,
-                       Title = Title,
-                       Dimensions = Dimensions,
-                       Measures = Measures,
-                       Attributes = Attributes,
-                       Identifier = Identifier,
-                       type = "csv")
+  if (type == "csv") {
+    dataset_download_csv(url = url,
+                         title = title,
+                         author = author,
+                         destfile = destfile,
+                         ...)
+  } else {
+    stop ('Currently only dataset_download(..., type = "csv" is implemented')
+  }
+
+
 }
 
 
 #' @rdname dataset_download
+#' @inheritParams utils::download.file
 #' @keywords internal
 dataset_download_csv  <- function(url,
-                                  Title,
-                                  Dimensions = NULL,
-                                  Measures = NULL,
-                                  Attributes = NULL,
-                                  Identifier = NULL,
-                                  type = "csv",
+                                  title,
+                                  author,
                                   destfile = NULL,
-                                  method =  'auto',
-                                  quiet = TRUE,
-                                  mode = 'wb',
-                                  cacheOK = TRUE) {
+                                  ...) {
 
   if (is.null(destfile)) destfile <- tempfile()
 
-  download.file(url = url,
-                destfile = destfile,
-                method = method,
-                quiet = quiet,
-                mode = mode,
-                cacheOK = cacheOK)
+  download.file(url = url, destfile=destfile, ... = ...)
 
   tmp <- read.csv(destfile)
 
-  if(!all(Dimensions %in% names(tmp))) {
-    stop("Not all Dimensions are present in ", destfile)
-  }
-
-  if(!all(Measures %in% names(tmp))) {
-    stop("Not all Measures are present in ", destfile)
-  }
-
-  if(!all(Attributes %in% names(tmp))) {
-    stop("Not all Attributes are present in ", destfile)
-  }
-
   ds <- dataset ( x = tmp,
-                  Dimensions = Dimensions,
-                  Measures = Measures,
-                  Attributes = Attributes,
-                  Title  = Title )
-
-  ds <- dublincore_add (ds,
-                        Identifier = Identifier,
-                        Source = url)
+                  title = title,
+                  author = author,
+                  ...)
 
   ds
 }
-
