+++ tests/testthat/test-dataset_title.R
@@ -1,9 +1,12 @@
+myiris <- iris_dataset
+
 test_that("dataset_title() works", {
-  expect_true(is.character(dataset_title(iris_dataset)))
-  expect_equal(dataset_title(iris_dataset), "Iris Dataset")
+  expect_true(is.character(dataset_title(myiris)))
+  expect_equal(dataset_title(myiris), "Iris Dataset")
 })
 
-test_that("set_dataset_title() works", {
-  expect_true(is.character(dataset_title(iris_dataset)))
-  expect_equal(dataset_title(set_dataset_title(iris_dataset, "The Famous Iris Dataset" )), "The Famous Iris Dataset")
+test_that("dataset_title() <- assignment works", {
+  expect_true(is.character(dataset_title(myiris)))
+  expect_warning(dataset_title(myiris) <-"New title")
+  expect_equal(dataset_title(myiris, overwrite = TRUE) <-"The Famous Iris Dataset", "The Famous Iris Dataset")
 })
