+++ README.Rmd
@@ -33,117 +33,70 @@ rlang::check_installed("here")
 
 The primary aim of dataset is create well-referenced, well-described, interoperable datasets from data.frames, tibbles or data.tables that translate well into the W3C DataSet definition within the [Data Cube Vocabulary](https://www.w3.org/TR/vocab-data-cube/) in a reproducible manner. The data cube model in itself is is originated in the _Statistical Data and Metadata eXchange_, and it is almost fully harmonized with the Resource Description Framework (RDF), the standard model for data interchange on the web^[RDF Data Cube Vocabulary, W3C Recommendation 16 January 2014  <https://www.w3.org/TR/vocab-data-cube/>, Introduction to SDMX data modeling <https://www.unescap.org/sites/default/files/Session_4_SDMX_Data_Modeling_%20Intro_UNSD_WS_National_SDG_10-13Sep2019.pdf>]. 
 
-A mapping of R objects into these models has numerous advantages: 
+The development version of the `dataset` package is very significantly differnet from the CRAN release. The documentation has not been rewritten yet!
 
-1. Makes data importing easier and less error-prone; 
-2. Leaves plenty of room for documentation automation, resulting in far better reusability and reproducibility; 
-3. The publication of results from R following the [FAIR](https://www.go-fair.org/fair-principles/) principles is far easier, making the work of the R user more findable, more accessible, more interoperable and more reusable by other users;
-4. Makes the placement into relational databases, semantic web applications, archives, repositories possible without time-consuming and costly data wrangling (See [From dataset To RDF](https://dataset.dataobservatory.eu/articles/RDF.html)).
-
-Our package functions work with any structured R objects (data.fame, data.table, tibble, or well-structured lists like json), however, the best functionality is achieved by the (See [The dataset S3 Class](https://dataset.dataobservatory.eu/articles/dataset.html)), which is inherited from `data.frame()`.
-
-## Installation
-
-You can install the development version of dataset from Github:
-
-``` r
-remotes::install_github('dataobservatory-eu/dataset')
-```
-
-or install from CRAN:
-``` r
-install.packages('dataset')
-```
-
-## Getting started
-
-The `dataset()` constructor creates a dataset from a data.frame or similar object. 
-
-```{r initialize}
+```{r initialise}
 library(dataset)
-my_iris_dataset <- dataset(
-  x = iris, 
-  Dimensions = NULL, 
-  Measures = c("Sepal.Length", "Sepal.Width",  "Petal.Length", "Petal.Width"), 
-  Attributes = "Species", 
-  Title = "Iris Dataset", 
-  Issued = 1936
+iris_ds <- dataset(
+  x = iris,
+  title = "Iris Dataset",
+  author = person("Edgar", "Anderson", role = "aut"),
+  publisher = "American Iris Society",
+  source = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x",
+  date = 1935,
+  language = "en",
+  description = "This famous (Fisher's or Anderson's) iris data set."
 )
-
-is.dataset(my_iris_dataset)
 ```
 
-Then you add the metadata:
+It is mandatory to add a `title`, `author` to a dataset, and if the `date` is not specified, the current date will be added. 
 
-```{r dublincore}
-my_iris_dataset <- dublincore_add(
-  x = my_iris_dataset,
-  Creator = person("Edgar", "Anderson", role = "aut"),
-  Publisher = "American Iris Society",
-  Source = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x",
-  Date = 1935,
-  Language = "en"
-)
+As the dataset at this point is just created, if it is not published yet, the `identifer` receives the default `:tba` value, a `version` of 0.1.0 and the `:unas` (unassigned) `publisher` field.
 
-print(my_iris_dataset)
[…]
