+++ NAMESPACE
@@ -6,7 +6,9 @@ S3method(is.datacite,datacite)
 S3method(is.dublincore,dublincore)
 S3method(is.subject,subject)
 S3method(print,dataset)
+S3method(set_var_labels,dataset)
 S3method(summary,dataset)
+S3method(var_labels,dataset)
 export("creator<-")
 export("dataset_title<-")
 export("datasource<-")
@@ -18,18 +20,18 @@ export("publication_year<-")
 export("publisher<-")
 export("rights<-")
 export("version<-")
+export(DataStructure)
+export(DataStructure_update)
 export(as_dataset)
 export(creator)
 export(datacite)
 export(dataset)
 export(dataset_bibentry)
 export(dataset_download)
-export(dataset_export)
-export(dataset_export_csv)
-export(dataset_local_id)
 export(dataset_title)
 export(dataset_uri)
 export(datasource)
+export(describe)
 export(description)
 export(dublincore)
 export(geolocation)
@@ -50,5 +52,3 @@ importFrom(utils,download.file)
 importFrom(utils,object.size)
 importFrom(utils,person)
 importFrom(utils,read.csv)
-importFrom(utils,write.csv)
-importFrom(utils,write.table)
