+++ README.Rmd
@@ -33,7 +33,7 @@ rlang::check_installed("here")
 
 The primary aim of dataset is create well-referenced, well-described, interoperable datasets from data.frames, tibbles or data.tables that translate well into the W3C DataSet definition within the [Data Cube Vocabulary](https://www.w3.org/TR/vocab-data-cube/) in a reproducible manner. The data cube model in itself is is originated in the _Statistical Data and Metadata eXchange_, and it is almost fully harmonized with the Resource Description Framework (RDF), the standard model for data interchange on the web^[RDF Data Cube Vocabulary, W3C Recommendation 16 January 2014  <https://www.w3.org/TR/vocab-data-cube/>, Introduction to SDMX data modeling <https://www.unescap.org/sites/default/files/Session_4_SDMX_Data_Modeling_%20Intro_UNSD_WS_National_SDG_10-13Sep2019.pdf>]. 
 
-The development version of the `dataset` package is very significantly differnet from the CRAN release. The documentation has not been rewritten yet!
+The development version of the `dataset` package is very significantly differnet from the CRAN release. The documentation has not been rewritten yet! You can follow the discussion of this package on [rOpenSci](https://github.com/ropensci/software-review/issues/553).
 
 ```{r initialise}
 library(dataset)
