+++ vignettes/motivation.Rmd
@@ -65,13 +65,13 @@ A mapping of R objects into these models has numerous advantages: it makes data 
 ## Data semantics
 The tidy data principles are seemingly very simple: each variable is a column, each observation is a row, and place all new units of measurement into a new table. However, the steps of tidying up a messy dataset require practice, like solving a Rubik cube–--take a long time to do quicky.
 
-The tidy data principles frame Codd's relational algebra, particularly the third normal form (3NF) into statistical language, and make the storage and reuse of data in relational databases optional. It suggests a format for tabular data and data semantics that makes a very important, and usually the most time (resource) consuming part of data analysis far more efficient, because the analyst does not have to reinvent the wheel all the time.The data cube model can be seen as an extension of the tidy model that makes the storage, reuse, and exchange of data among different database, or their placement on a knowledge graph optimal.
+The tidy data principles frame Codd's relational algebra, particularly the third normal form (3NF), into statistical language and make the storage and reuse of data in relational databases optional. It suggests a format for tabular data and data semantics that makes a very important and usually the most time-consuming (resource-consuming) part of data analysis far more efficient because the analyst does not have to reinvent the wheel all the time. The data cube model can be seen as an extension of the tidy model that makes the storage, reuse, and exchange of data among different database or their placement on a knowledge graph optimal.
 
-The wording of the tidy data principles is not sufficiently exact for the data cube model.  THe 3NF form stipulates that the (e.g. table columns) are functionally dependent on solely the primary key, which means that we must have a primary key that is locally unique in the table. The primary key must be a unique combination of the dimension and attribute columns, including the row identifier (rowid), if it exists. This means a lot more than the uniqueness of the measurement unit.  
+The wording of the tidy data principles is not sufficiently exact for the data cube model. The 3NF form stipulates that the (e.g. table columns) are functionally dependent solely on the primary key, which means that we must have a primary key that is locally unique in the table. If it exists, the primary key must be a unique combination of the dimension and attribute columns, including the row identifier (rowid). This means a lot more than the uniqueness of the measurement unit.  
 
-A usual dimension is the reference time period and the reference geographical area.  Consider a dataset that contains population growth numbers from a county of a country, collected in the time period of 2010..2015 quarterly, and in 2016:2020 annually, with the county borders changing in 2014, and the unit of measure changed from the number of people in thousands to the number of people in millions. The meausurement of the reference time period and the aggregation area has changed, too.
+A usual dimension is the reference time period and the reference geographical area. Consider a dataset that contains population growth numbers from a county of a country, collected in the time period of 2010..2015 quarterly and in 2016:2020 annually, with the county borders changing in 2014, and the unit of measure changed from the number of people in thousands to the number of people in millions. The measurements of the reference time period and the aggregation area has changed, too.
 
-```{r}
+```{r dataset-constructor}
 library(dataset)
 my_dataset <- dataset (x = data.frame (
   time = rep(c(2019:2022),4),
@@ -80,53 +80,49 @@ my_dataset <- dataset (x = data.frame (
   value = c(1,3,2,4,2,3,1,5, NA_real_, 4,3,2,1, NA_real_, 2,5),
   unit = rep("NR",8),
   freq = rep("A",8)),
-  Dimensions = c("time", "geo", "sex"),
-  Measures = "value",
-  Attributes = c("unit", "freq"),
-  sdmx_attributes = c("sex", "time", "freq"),
-  Title = "Example dataset",
-  Creator = person("Jane", "Doe")
+  title = "Example dataset",
+  author = person("Jane", "Doe")
 )
 
-my_dataset <- dataset_local_id(my_dataset)
 my_dataset
 ```
+
 ### The data cube model
[…]
