### 2023-12-07 841f5d3
Improvements with DataCite and DublinCore metadata handling, new tests.
