+++ tests/testthat/test-dublincore.R
@@ -8,13 +8,14 @@ dct_iris <- dublincore(
   description = "This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris. The species are Iris setosa, versicolor, and virginica."
 )
 
-str(dct_iris)
+
+
 
 test_that("dublincore works", {
   expect_true(is.dublincore(dct_iris))
 })
 
-test_that("dublincore works", {
+test_that("dublincore() works", {
   expect_equal(dct_iris$language, 'en')
   expect_equal(dct_iris$publisher, "American Iris Society")
   expect_equal(dct_iris$date, "1935")
@@ -26,5 +27,14 @@ test_that("dublincore works", {
   expect_equal(dct_iris$description, "This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris. The species are Iris setosa, versicolor, and virginica.")
 })
 
-
-
+test_that("as_dublincore() works", {
+  expect_true(is.dublincore(as_dublincore(iris_dataset)))
+  expect_true(is.list(as_dublincore(x=iris_dataset, type="list")))
+  expect_true(is.dataset(as_dublincore(iris_dataset, type="dataset")))
+  expect_true(is.null(as_dublincore(iris_dataset, type="list")$subject))
+  expect_true(is.null(as_dublincore(iris_dataset, type="list")$contributor))
+  expect_equal(as_dublincore(iris_dataset, type="dataset")$description, ':unas')
+  expect_equal(dataset_title(as_dublincore(iris_dataset, type="dataset")), "The Dublin Core Metadata of `Iris Dataset'")
+  expect_equal(as_dublincore(iris_dataset, type="dataset")$rights, ':tba')
+  expect_equal(as_dublincore(iris_dataset, type="dataset")$coverage, ':unas')
+})
