+++ R/datacite.R
@@ -1,28 +1,15 @@
-#' @title Add DataCite metadata to a bibentry object
-#'
-#' @description Add metadata conforming the \href{https://schema.datacite.org/}{DataCite Metadata Schema}
-#' to a \code{utils::\link[utils]{bibentry}} object, for an accurate and consistent identification
-#' of a resource for citation and retrieval purposes.
-#'
+#' @title Add or get DataCite metadata
+#' @description Add or retrieve metadata conforming
+#' the \href{https://schema.datacite.org/}{DataCite Metadata Schema}.
 #' @details DataCite is a leading global non-profit organisation that provides persistent identifiers
-#' (DOIs) for research data and other research outputs. Organizations within the research
+#' (DOIs) for research data and other research outputs. Organisations within the research
 #' community join DataCite as members to be able to assign DOIs to all their research
-#' outputs. This way, their outputs become discoverable and associated metadata is made
-#' available to the community.\cr
-#' DataCite then develops additional services to improve the DOI management experience,
-#' making it easier for our members to connect and share their DOIs with the broader
-#' research ecosystem and to assess the use of their DOIs within that ecosystem.
-#' DataCite is an active participant in the research community and promotes data sharing
-#' and citation through community-building efforts and outreach activities.
-#' @param x An R object of type data.frame, or inherited data.table, tibble; alternatively a well
-#' structured R list.
+#' outputs. This way, their outputs become discoverable, and associated metadata is made.
+#' available to the community.
 #' @details The \code{ResourceType} property will be by definition "Dataset".
 #' The \code{Size} attribute (e.g. bytes, pages, inches, etc.) will automatically added to the dataset.
 #' @param Title The name(s) or title(s) by which a resource is known. May be the title of a dataset
 #' or the name of a piece of software. Similar to \href{https://purl.org/dc/elements/1.1/title}{dct:title}.\cr
-#' See \code{\link{dataset_title}} for adding further titles.
-#' @param titleType For a single \code{Title} defaults to \code{NULL}. Otherwise you can add
-#' a Subtitle, an Alternative Title and an Other Title. See \code{\link{dataset_title}}.
 #' @param Creator The main researchers involved in producing the data, or the authors of the publication, in
 #' priority order. To supply multiple creators, repeat this property.
 #' @param Identifier The Identifier is a unique string that identifies a resource. For software, determine
@@ -43,7 +30,8 @@
 #' Use \code{\link{subject}} to properly add a key phrase from a controlled vocabulary
 #' and create structured Subject objects with \code{\link{subject_create}}.
 #' @param Contributor Recommended for discovery. The institution or person responsible for collecting, managing, distributing, or otherwise contributing to the development of the resource.
-#' @param Date Recommended for discovery in DataCite. Similar to \href{https://purl.org/dc/elements/1.1/date}{dct:date} in
+#' @param Date Recommended for discovery in DataCite. Similar to
+#' \href{https://purl.org/dc/elements/1.1/date}{dct:date} in
 #' \code{\link{dublincore}}.
 #' @param Publisher The name of the entity that holds, archives, publishes prints,
 #' distributes, releases, issues, or produces the resource. This property will be used to
@@ -57,14 +45,14 @@
 #' specific \code{Type$resourceType} value. See \code{\link{resource_type}}.
 #' @param Language The primary language of the resource. Allowed values are taken from
 #' IETF BCP 47, ISO 639-1 language code. See \code{\link{language}}.
[…]
