+++ R/creator.R
@@ -1,21 +1,19 @@
 #' @title Get/set the Creator of the object.
-#' @description Add the optional \code{Creator} property as an attribute to an R object.
+#' @description Add the optional \code{Creator} property as an attribute to a
+#' dataset object.
 #' @details The \code{Creator}
-#' corresponds to \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/#https://purl.org/dc/elements/1.1/creator}{dct:creator} and Creator in DataCite.
-#' The name of the entity that holds, archives, publishes prints, distributes, releases,
-#' issues, or produces the resource. This property will be used to formulate the citation,
-#' so consider the prominence of the role. For software, use \code{Creator} for the
-#' code repository. If there is an entity other than a code repository, that
-#' "holds, archives, publishes, prints, distributes, releases, issues, or produces" the
-#' code, use the property Contributor/contributorType/hostingInstitution for the code
-#' repository.
-#' @param x An R object, such as a data.frame, a tibble, or a data.table.
+#' corresponds to \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/creator/}{dct:creator}
+#' in Dublin Core and Creator in DataCite.
+#' The name of the entity that holds, archives, publishes prints, distributes,
+#' releases, issues, or produces the datast. This property will be used to
+#' formulate the citation, so consider the prominence of the role.
+#' @param ds A dataset object created by \code{\link{dataset}}.
 #' @param value The \code{Creator} as a \code{utils::\link[utils]{person}} object.
 #' @param overwrite If the attributes should be overwritten. In case it is set to \code{FALSE},
 #' it gives a message with the current \code{Creator} property instead of overwriting it.
 #' Defaults to \code{TRUE} when the attribute is set to \code{value} regardless of previous
 #' setting.
-#' @return The Creator attribute as a character of length 1 is added to \code{x}.
+#' @return The Creator attribute as a character of length one is added to \code{x}.
 #' @importFrom utils person
 #' @examples
 #' ds <- dataset(iris,
