### 2023-12-07 217eba8
Updating almost all metadata functions and long-form documentation.
