+++ R/publication_year.R
@@ -2,41 +2,50 @@
 #' @description Get/set the optional \code{publication_year} property as an attribute to an R object.
 #' @details The \code{PublicationYear} is the year when the data was or will be made
 #' publicly available in \code{YYYY} format.
-#' @param x An R object, such as a data.frame, a tibble, or a data.table.
+#' See
+#' \href{https://support.datacite.org/docs/datacite-metadata-schema-v44-mandatory-properties#publicationyearadditional-guidance}{Publication Year: DataCite Additional Guidance}.
+#' @param x A dataset object created with \code{dataset::\link{dataset}}.
 #' @param value The publication_year as a character set.
 #' @param overwrite If the attributes should be overwritten. In case it is set to \code{FALSE},
 #' it gives a message with the current \code{PublicationYear} property instead of overwriting it.
 #' Defaults to \code{TRUE} when the attribute is set to \code{value} regardless of previous
 #' setting.
-#' @return The \code{publication_year} attribute as a character of length 1 is added to \code{x}.
+#' @return Returns the \code{year} metadata field of the \code{DataBibentry} of
+#' the dataset
 #' @examples
 #' iris_dataset <- iris
-#' publication_year(iris_dataset) <- 1935
 #' publication_year(iris_dataset)
+#' publication_year(iris_dataset) <- 1936
 #' @family Reference metadata functions
 #' @export
 
 #' @export
 publication_year <- function(x) {
+  assert_that(is.dataset(x),
+              msg = "publication_year(x): x must be a dataset object created with dataset() or as_dataset().")
 
-  attr(x, "publication_year")
-
+  ds_bibentry <- dataset_bibentry(x)
+  as.character(ds_bibentry$year)
 }
 
 #' @rdname publication_year
 #' @export
 `publication_year<-`  <- function(x,  overwrite = TRUE, value) {
+  assert_that(is.dataset(x),
+              msg = "publication_year(x) <- value: x must be a dataset object created with dataset() or as_dataset().")
 
-  if (is.null(attr(x, "publication_year"))) {
-    if (is.null(value)) {
-      attr(x, "publication_year") <- NA_character_
-    } else {
-      attr(x, "publication_year") <- value
+  ds_bibentry <- dataset_bibentry(x)
+  publication_year <- ds_bibentry$year
+
+  if (is.null(value)) {
+      value <- ":unas"
     }
-  } else if ( overwrite ) {
-    attr(x, "publication_year") <- value
-  } else {
-    message ("The dataset has already an publication_year: ",  attr(x, "publication_year") )
+
+  if ( overwrite ) {
+    ds_bibentry$year <- as.character(value)
+    attr(x, "DataBibentry") <- ds_bibentry
+   } else {
+    warning ("The dataset has already an publication_year: ",  ds_bibentry$year, "." )
   }
-  x
+  invisible(x)
 }
