### 2023-12-07 587a174
dublincore() corrections
