+++ R/describe.R
@@ -1,16 +1,16 @@
 #' @title Describe a dataset object
-#' @description
+#' @description Print to the screen an easy-to-read summary of a
+#' the attributes of a dataset.
+#' @details
 #' A convenience function to review the most important attributes of a dataset
 #' object. The dataset class adds a wide range of metadata as attributes to
 #' data.frame, tibble, data.table or similar R object that contains tabular data.
 #' Overviewing these attributes becomes cumbersome with base
 #' R \code{attributes()}.
-#'
 #' @param x A dataset object created with \code{dataset::\link{dataset}}.
 #' @return No object is returned, but they key attributes are printed on the screen.
 #' @examples
 #' describe(iris_dataset)
-#'
 #' @export
 describe <- function(x) {
   UseMethod("describe", x)
@@ -19,6 +19,8 @@ describe <- function(x) {
 #' @rdname describe
 #' @exportS3Method
 describe.dataset <- function(x){
+
+  empty_values <- c("", ":unas", ":tba")
 
   assertthat::assert_that(is.dataset(x),
                           msg="describe(ds): ds must be a dataset object.")
@@ -39,26 +41,33 @@ describe.dataset <- function(x){
   row_num <- ifelse (row_num>1, paste0(row_num, " observations (rows)"), paste0(row_num, " observation (row)"))
   col_num <- ifelse (col_num>1, paste0(col_num, " variables (columns)"), paste0(col_num, " variable (column)"))
   cat(paste0(DataBibentry[[1]]$resourceType, " with ", row_num, " and ", col_num, ".\n"  ))
-  if(! is.null(DataBibentry[[1]]$description) ) {
-    cat(paste0("Description: ", DataBibentry[[1]]$description, "\n"))
+
+  dataset_description <- description(x)
+
+  if(! is.null(dataset_author) & ! dataset_description %in% empty_values ) {
+    cat(paste0("Description: ", as.character(dataset_description), "\n"))
+  }
+
+  dataset_author <- creator(x)
+
+  if(! is.null(dataset_author) & ! dataset_author %in% empty_values ) {
+    cat(paste0("Creator: ", as.character(dataset_author), "\n"))
   }
 
   if(! is.null(DataBibentry[[1]]$author) ) {
     cat(paste0("Creator: ", as.character(DataBibentry[[1]]$author), "\n"))
   }
 
+  dataset_publisher <- publisher(x)
 
-  if(! is.null(DataBibentry[[1]]$publisher) ) {
-    cat(paste0("Publisher: ", as.character(DataBibentry[[1]]$publisher), "\n"))
+  if(! is.null(dataset_publisher) & ! dataset_publisher %in% empty_values ) {
+    cat(paste0("Publisher: ", as.character(dataset_publisher), "\n"))
   }
 
-  if(! is.null(DataBibentry[[1]]$rights) ) {
-    cat(paste0("Rights: ", as.character(DataBibentry[[1]]$rights), "\n"))
+  dataset_rights <- rights(x)
+
+  if(! is.null(dataset_rights) & ! dataset_rights %in% empty_values ) {
+    cat(paste0("Rights: ", as.character(dataset_rights), "\n"))
   }
-
-  #if(!is.null(attr(x, "Subject"))) {
-  #  cat(paste0("Subject: ", names(attr(x, "Subject")), "\n"))
-  #}
-
 
 }
