+++ R/dublincore.R
@@ -138,7 +138,21 @@ dublincore <- function(
 
 #' @rdname dublincore
 #' @export
-as_dublincore <- function(x, type = "bibentry") {
+as_dublincore <- function(x, type = "bibentry", ...) {
+
+  citation_author <- person(NULL, NULL)
+
+  is_person <- function(p) ifelse (inherits(p, "person"), TRUE, FALSE)
+
+  arguments <- list(...)
+
+  if (!is.null(arguments$author)) {
+    if ( is_person(arguments$author))  {
+      citation_author <- arguments$author
+    } else {
+      stop("as_datacite(x, ..., author = ): author must be created with utils::person().")
+    }
+  }
 
   if (! type %in% c("bibentry", "list", "dataset")) {
     warning_message <- "as_datacite(ds, type=...) type cannot be "
@@ -147,87 +161,106 @@ as_dublincore <- function(x, type = "bibentry") {
   }
 
   ds_bibentry <- dataset_bibentry(x)
-  title   <- ds_bibentry$title
-  creator <- ds_bibentry$author
+  dataset_title   <- ds_bibentry$title
+  dataset_creator <- ds_bibentry$author
 
 
-  if (!is.null(ds_bibentry$date)) {
+  if (!is.null(ds_bibentry$year)) {
     if(!is.null(ds_bibentry$year)) {
-      date <- ":tba"
+      dataset_date <- ":tba"
     } else {
-      date <- as.character(ds_bibentry$year)
+      dataset_date <- as.character(ds_bibentry$year)
     }
-  } else {
-    date <- as.character(ds_bibentry$date)
   }
 
-  relation <- ifelse (is.null(ds_bibentry$relation), ":unas", as.character(ds_bibentry$relation))
-  identifier <- ifelse (is.null(ds_bibentry$identifier), ":tba", as.character(ds_bibentry$identifier))
-  version <- ifelse (is.null(ds_bibentry$version), ":unas", as.character(ds_bibentry$version))
-  description <- ifelse (is.null(ds_bibentry$description), ":unas", as.character(ds_bibentry$description))
-  language <- ifelse (is.null(ds_bibentry$language), ":unas", as.character(ds_bibentry$language))
-  format <- ifelse (is.null(ds_bibentry$format), ":tba", as.character(ds_bibentry$format))
-  rights <- ifelse (is.null(ds_bibentry$rights), ":tba", as.character(ds_bibentry$rights))
-  coverage  <- ifelse (is.null(ds_bibentry$coverage), ":unas", as.character(ds_bibentry$coverage))
+  dataset_relation <- ifelse (is.null(ds_bibentry$relation), ":unas", as.character(ds_bibentry$relation))
+  dataset_identifier <- ifelse (is.null(ds_bibentry$identifier), ":tba", as.character(ds_bibentry$identifier))
+  dataset_version <- ifelse (is.null(ds_bibentry$version), ":unas", as.character(ds_bibentry$version))
+  dataset_description <- ifelse (is.null(ds_bibentry$description), ":unas", as.character(ds_bibentry$description))
+  dataset_language <- ifelse (is.null(ds_bibentry$language), ":unas", as.character(ds_bibentry$language))
+  dataset_format <- ifelse (is.null(ds_bibentry$format), ":tba", as.character(ds_bibentry$format))
+  dataset_rights <- ifelse (is.null(ds_bibentry$rights), ":tba", as.character(ds_bibentry$rights))
+  dataset_coverage  <- ifelse (is.null(ds_bibentry$coverage), ":unas", as.character(ds_bibentry$coverage))
   datasource <- ifelse (is.null(ds_bibentry$datasource), ":unas", as.character(ds_bibentry$datasource))
[…]
