+++ R/dublincore.R
@@ -42,12 +42,9 @@
 #' lifecycle of the resource. \href{https://purl.org/dc/elements/1.1/date}{dct:date}.
 #' \code{Date} is also recommended for
 #' discovery in \code{\link{datacite}}, but it requires a different formatting.
-#' @param source A related resource from which the described resource is derived.
-#' See \href{https://purl.org/dc/elements/1.1/source}{dct:source} and
-#' \code{\link{datasource}}.
 #' @param language The primary language of the resource. Allowed values are taken from
 #' IETF BCP 47, ISO 639-1 language code. See \code{\link{language}}. Corresponds to Language in Datacite.
-#' @param Format The file format, physical medium, or dimensions of the resource.
+#' @param format The file format, physical medium, or dimensions of the resource.
 #' \href{	https://purl.org/dc/elements/1.1/format}{dct:format}
 #' Examples of dimensions include size and duration. Recommended best practice is to use a controlled
 #' vocabulary such as the list of \href{https://www.iana.org/assignments/media-types/media-types.xhtml}{Internet Media Types, formerly known as MIME}. It is similar to \code{Format} in
