+++ R/dataset.R
@@ -13,7 +13,7 @@
 #' be announced later.
 #' @param year The year of the creation of the dataset. If left empty (NULL),
 #' the current year.
-#' @param version The version of the dataset. If left emtpy (NULL), defaults to
+#' @param version The version of the dataset. If left empty (NULL), defaults to
 #' \code{'0.1.0'}
 #' @param subject The subject of the data frame, as a subject type.
 #' @param description The optional \code{Description} property as an attribute to
