+++ R/publisher.R
@@ -8,7 +8,7 @@
 #' "holds, archives, publishes, prints, distributes, releases, issues, or produces" the
 #' code, use the property Contributor/contributorType/ hostingInstitution for the code
 #' repository.
-#' @param x An R object, such as a data.frame, a tibble, or a data.table.
+#' @param x A dataset object created with \code{dataset::\link{dataset}}.
 #' @param overwrite If the attributes should be overwritten. In case it is set
 #' to \code{FALSE},it gives a warning with the current \code{publisher}
 #' property instead of overwriting it. Defaults to \code{FALSE}.
