+++ cran-comments.md
@@ -1,11 +1,9 @@
 ## Test environments
 
 * Windows10 x86_64-w64-mingw32 (64-bit), R version 4.2.1, locally.
-* Ubuntu 22.04.1 LTS x86_64-pc-linux-gnu (64-bit), R verseion 4.2.1, locally
+* Ubuntu 22.04.1 LTS x86_64-pc-linux-gnu (64-bit), R version 4.2.1, locally
 0 errors ✔ | 0 warnings ✔ | 0 notes ✔
 
-* Fedora Linux, R-devel, clang, gfortran
-* Ubuntu Linux 20.04.1 LTS, R-release, GCC
-* Windows Server 2022, R-devel, 64 bit
-
-This is a documentation improvement.  The latency of the Eurostat data warehouse triggered a broken link alarm on CRAN with one of the builds. That part of the vignette is not evaluated anymore, as the code chunk only served tutorial purposes. The same vignette text was very slightly altered for readability.
+*  Windows Server 2022 x64 (build 20348); R Under development (unstable)
+*  Fedora Linux, R-devel, clang, gfortran
+*  Ubuntu Linux 20.04.1 LTS, R-release, GCC
