### 2023-12-07 091cd41
small documentation errors corrected in manual
