+++ man/dataset_download.Rd
@@ -10,7 +10,7 @@ dataset_download(url, title, author, destfile, type = "csv", ...)
 dataset_download_csv(url, title, author, destfile = NULL, ...)
 }
 \arguments{
-\item{url}{a \code{\link{character}} string (or longer vector e.g.,
+\item{url}{a \code{\link{character}} string (or longer vector
     for the \code{"libcurl"} method) naming the URL of a resource to be
     downloaded.}
 
