### 2023-12-16 0bb35bf
New vignette on _Richer Semantics for the Dataset's Variables_.
