+++ vignettes/variables.Rmd
@@ -1,0 +1,66 @@
+---
+title: "Richer Semantics for the Dataset's Variables"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Richer Semantics for the Dataset's Variables}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+> There are only two hard things in Computer Science: cache invalidation and naming things. 
+-- Phil Karlton
+
+In a tidy dataset, variables are in columns. An R data frame (in its form as data.frame, tibble, tsibble, data.table) allows using column names that do not contain characters reserved with a special meaning in the R language itself. Choosing good variable names is an art for the data scientist: it improves code readability, review, and later reuse. However, the use of often used (and standardised) variable names like `GEO`, `SEX`, `UNIT` is not enough for a third-party review or reuse; the same analyst will likely have problems recalling the _meaning_ of these variables without further documentation.
+
+Let's take a look at the famous `iris` dataset. What does _Sepal.Length_ or _Species_ means? What happens if we make a mutating transformation of _Sepal.Width_? What is the unit of measure if we take the average of _Petal.Widht_?
+
+```{r iris}
+head(iris)
+```
+
+One good practice for increasing reviewability and reproducibility is to include this information in a notebook or similar document. While this is undoubtedly a good practice, often the reuse or review is hampered by the fact that the data and the documentation are separated, and for example, the dataset is available without the documentation, or worse still, the dataset is modified after the documentation was closed, but this is not visible for the reuser.
+
+## Labelling variables
+
+There are some R packages that help to add semantic information about the variables with a free-text variable label that allows normal punctuation and more words than the restricted col.names attribute of the data.frame. Our _dataset_ package also implements variable labelling.
+
+```{r setup}
+library(dataset)
+```
+
+```{r addlabels}
+dsd_iris <- DataStructure(iris_dataset)
+dsd_iris$Sepal.Length$label <- "The sepal length measured in centimeters."
+dsd_iris$Sepal.Width$label  <- "The sepal width measured in centimeters."
+dsd_iris$Petal.Length$label <- "The petal length measured in centimeters."
+dsd_iris$Petal.Width$label  <- "The petal width measured in centimeters."
+dsd_iris$Species$label      <- "The name of the Iris species in the observation."
+
+iris_dataset_labelled <- DataStructure_update(iris_dataset, dsd_iris)
+```
+
+```{r retrievelabel}
+DataStructure(iris_dataset_labelled)[["Sepal.Length"]]$label
+```
+
+### Conceptualisation of variables
+
+Variable labels vary in quality, and they are natural language and context-dependent. They are not machine-actionable and require substantial time investment for interpretation or validation for reuse. 
+
+Professional dataset producers, for example, statistical agencies, therefore encourage using controlled vocabularies, such as taxonomies or formal ontologies, to describe the meaning of the variables. Our main goal with regards to increase the usability, reviewability or knowledge-expanding mutation capability of variables by adding standardised, machine-actionable information to the datasets.
+
+### Increased interoperability
+
+Interoperability is enhanced if we know what is the valid range of the variable. Until you save the `iris_dataset_labelled` R object in native .rds form, basic interoperability is ensured: the sepal length variable will be read as a numeric variable. But even in a CSV serialisation, it is only sometimes clear to the following user's application if the variable should be read as an integer, double or a string.  The _dataset_ package allows simpler or richer range definitions to be added to each variable's structural information set.
+
+```{r retrieverange}
+DataStructure(iris_dataset_labelled)[["Sepal.Length"]]$range
+```
+
