+++ NEWS.md
@@ -1,6 +1,9 @@
 # dataset 0.2.8
 
-- New vignette on _Richer Semantics for the Dataset's Variables_.
+New vignettes on
+- [x] [Richer Semantics for the Dataset's Variables](https://dataset.dataobservatory.eu/articles/columns.html)
+- [x] [Richer Semantics for the Dataset's Observations](https://dataset.dataobservatory.eu/articles/rows.html)
+- [x] [Richer Description of the Datasets's Structure for Meaningful Reuse Opportunities](https://dataset.dataobservatory.eu/articles/DSD.html)
 
 # dataset 0.2.7 
 
