+++ vignettes/DSD.Rmd
@@ -1,0 +1,21 @@
+---
+title: "Richer Description of the Datasets's Structure for Meaningful Reuse Opportunities"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Richer Description of the Datasets's Structure for Meaningful Reuse Opportunities}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+The Statistical Data and Metadata eXchange used the Data Structure Definition to add structural information to datasets. Structural information goes beyond a semantically rich description of the variables (in columns) and observations (in rows). It contains information about the valid subsetting or extension of the dataset.
+
+```{r setup}
+library(dataset)
+```
