+++ DESCRIPTION
@@ -51,7 +51,8 @@ Suggests:
     spelling,
     statcodelists,
     testthat (>= 3.0.0),
-    tidyr
+    tidyr,
+    tibble
 VignetteBuilder: knitr
 Config/testthat/edition: 3
 Language: en-US

[…]
