### 2023-12-17 cfb0792
0.2.8: Documenting the new functionality of semantically enriched columns, rows, and data structure.
