### 2023-12-17 ec7980a
Update package citation information
