+++ R/datasource.R
@@ -1,4 +1,4 @@
-#' @title Get/set the Source of the object.
+#' @title Get/set the source property of a dataset.
 #' @description Get/set the optional \code{Source} property as an attribute to an
 #' R object. Do not confuse with the base R \code{source()} function.
 #' @details The \code{Source} is a related resource from which the described resource is
@@ -10,33 +10,33 @@
 #' @return The \code{Source} attribute as a character of length 1 is added to
 #' \code{x}.
 #' @examples
-#' datasource(iris_dataset) <- "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x"
-#' datasource(iris_dataset)
+#' iris_dataset <- datasource_set(iris_dataset, "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x")
+#' datasource_get(iris_dataset)
 #' @family Reference metadata functions
 #' @importFrom assertthat assert_that
 #' @export
-datasource <- function(x) {
+datasource_get <- function(x) {
 
   assert_that(is.dataset(x),
-              msg = "datasource(x) must be a dataset object created with dataset() or as_dataset().")
+              msg = "datasource_get(x) must be a dataset object created with dataset() or as_dataset().")
   DataBibentry <- dataset_bibentry(x)
   DataBibentry$source
 
 }
 
-#' @rdname datasource
+#' @rdname datasource_get
 #' @param overwrite If the attributes should be overwritten. In case it is set
 #' to \code{FALSE},it gives a warning with the current \code{datasource}
 #' property instead of overwriting it. Defaults to \code{FALSE}.
 #' @export
-`datasource<-` <- function(x,  overwrite = TRUE, value) {
+datasource_set <- function(x,  value, overwrite = TRUE) {
 
   assert_that(is.dataset(x),
-              msg = "datasource(x): x must be a dataset object created with dataset() or as_dataset().")
+              msg = "datasource_set(x): x must be a dataset object created with dataset() or as_dataset().")
 
   DataBibentry <- invisible(dataset_bibentry(x))
 
-  if ( is.null(value) ) {
+    if ( is.null(value) ) {
     DataBibentry$source <- ":unas"
     attr(x, "DataBibentry") <- DataBibentry
     return(invisible(x))
