### 2023-12-27 edfcbd5
subsetting
