+++ R/subsetting.R
@@ -1,0 +1,164 @@
+#' @title Subsetting datasets
+#' @description
+#' Accessing columns, rows, or cells via `$`, `[[`, or `[` is mostly similar to
+#' [regular data frames][base::Extract]. However, the
+#' * `[` always returns a similar dataset by default, even if
+#'   only one column is accessed. The title is modified to
+#'   from \code{'Original Title'} to \code{'Original Title [subset]'}
+#'  * `[[` always returns a vector.
+#'  * `$` always returns a vector.
+#' @name subsetting
+#' @examples
+#' # Subsetting single columns:
+#' iris_dataset[, "Species"]
+#'
+#' # Subsetting single column to vector:
+#' iris_dataset$Species
+#' iris_dataset[, "Species"]
+#'
+#' # Subsetting a single cell in tabular data:
+#' iris_dataset[[1,2]]
+#' @export
+`[.dataset` <- function(x, i, j, drop = FALSE, ...) {
+
+  i_arg <- substitute(i)
+  j_arg <- substitute(j)
+
+  if (missing(i)) {
+    if (nrow(x)>0) {
+      i <- 1:nrow(x)
+      i_arg <- NULL
+    } else {
+      i <- integer()
+    }
+  } else if (is.null(i)) {
+    if (nrow(x)>0) {
+     i <- 1:nrow(x)
+    } else {
+      i <- integer()
+    }
+  }
+
+  if (missing(j)) {
+    j <- 1:ncol(x)
+    j_arg <- NULL
+  } else if (is.null(j)) {
+    j <- integer()
+  }
+
+  DataStructure <- attr(x, "DataStructure")
+  DataBibentry <- dataset_bibentry(x)
+  x_Subject <- subject(x)
+  new_title <- paste0(dataset_title(x), " [subset]")
+  new_creator <- creator(x)
+  classes <- class(x)
+  class(x) <- classes[-which(classes == "dataset")]
+  tmp <- x[i, j, drop = drop]
+  new_DataStructure <- DataStructure[names(tmp)]
+  tmp <- dataset(tmp,
+                 author = new_creator,
+                 title = new_title,
+                 datasubject = x_Subject  )
+  DataBibentry$title <- new_title
+  attr(tmp, "DataBibentry") <-  DataBibentry
+  attr(tmp, "DataStructure") <- DataStructure
+
+  tmp
+}
+
+
+#' @rdname subsetting
+#' @param n Maximum number of rows.
+#' @export
+head.dataset <- function(x, n=6L, ...) {
+
+  if(is.null(x)) return(x)
+
+  if (nrow(x)>0) {
+    n <- min(nrow(x), n)
+    x[1:n, ]
+  } else {
+    x
+  }
+}
+
+
+
+#' @rdname subsetting
+#' @param name A [name] or a string.
+#' @export
+`$.dataset` <- function(x, name) {
+  col_n <- which(names(x) %in% name)
+  if (length(col_n)==0) {
+    stop(paste0("Unknown or uninitialised column: '", name, "'."))
+  } else {
+   unname(unlist(x[, col_n]))
+  }
+}
+
+`[[.dataset` <- function(x, i, j, ..., exact = TRUE) {
+
+  if (!exact) {
+    warning("dataset[[..., exact=FALSE]] ignored.")
+  }
+
+  if (exists("x", envir = environment())) {
+    x_arg <- get("x", envir = environment())
+  }
+
+  n_arguments <- 0
+  if(!missing(i)) n_arguments <- 1
+  if(!missing(j)) n_arguments <- n_arguments+1
+
+  dot_args <- list(...)
+
+  n_dots <- dots_number(...)
+  if (n_dots > 0) {
[…]
