+++ man/subsetting.Rd
@@ -1,0 +1,42 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/subsetting.R
+\name{subsetting}
+\alias{subsetting}
+\alias{[.dataset}
+\alias{head.dataset}
+\alias{$.dataset}
+\title{Subsetting datasets}
+\usage{
+\method{[}{dataset}(x, i, j, drop = FALSE, ...)
+
+\method{head}{dataset}(x, n = 6L, ...)
+
+\method{$}{dataset}(x, name)
+}
+\arguments{
+\item{n}{Maximum number of rows.}
+
+\item{name}{A \link{name} or a string.}
+}
+\description{
+Accessing columns, rows, or cells via \code{$}, \code{[[}, or \code{[} is mostly similar to
+\link[base:Extract]{regular data frames}. However, the
+\itemize{
+\item \code{[} always returns a similar dataset by default, even if
+only one column is accessed. The title is modified to
+from \code{'Original Title'} to \code{'Original Title [subset]'}
+\item \code{[[} always returns a vector.
+\item \code{$} always returns a vector.
+}
+}
+\examples{
+# Subsetting single columns:
+iris_dataset[, "Species"]
+
+# Subsetting single column to vector:
+iris_dataset$Species
+iris_dataset[, "Species"]
+
+# Subsetting a single cell in tabular data:
+iris_dataset[[1,2]]
+}
