### 2023-12-28 cdc3994
New replication dataset data("statwales")
