+++ R/data-statwales.R
@@ -1,0 +1,11 @@
+#' @title Life Expectancy in Regions of Wales by Sex
+#'
+#' @description A small demonstration data set extracted from StatsWales report
+#' number 003311 which describes life expectancy broken down by region
+#' (unitary authority), age and time.
+#'
+#' @format The dataset has originally one column for reference areas, and three
+#' column-pairs in slices, in total 7 column, and 4 reference area rows.
+#' @source  \href{https://www.w3.org/TR/vocab-data-cube/#dataset}{The RDF Data Cube Vocabulary},
+#' example in 5.4.
+"statwales"
