### 2023-12-29 02b06bd
subsetting and consequences
