+++ R/dataset.R
@@ -1,7 +1,7 @@
 #' @title Create a dataset
 #'
 #' @param x An R object that contains the data of the dataset (a data.frame or
-#' inherited from data.frame, for example, tibble, tsibble, data.table).
+#' inherited from [data.frame], for example, tibble, tsibble, data.table).
 #' @param author A single person or a vector of persons as authors, declared with
 #' \code{\link[utils:person]{person}}.
 #' @param title The title of the dataset.
@@ -15,7 +15,7 @@
 #' the current year.
 #' @param version The version of the dataset. If left empty (NULL), defaults to
 #' \code{'0.1.0'}
-#' @param subject The subject of the data frame, as a subject type.
+#' @param datasubject The subject of the data frame, as a [subject] type.
 #' @param description The optional \code{Description} property as an attribute to
 #' an R object.
 #' @param language The primary language of the dataset, for example \code{'eng'}.
@@ -30,6 +30,7 @@
 #' Free text, defaults to \code{":unas"} for unassigned values. See \code{\link{rights}}.
 #' @return A dataset object, which is a data.frame or inherited object with rich
 #' metadata.
+#' @seealso [xsd_convert()]
 #' @examples
 #' ds <- dataset(iris,
 #'         title = "The iris Dataset",
@@ -53,7 +54,7 @@ dataset  <- function(x,
                      publisher = NULL,
                      year = NULL,
                      version = NULL,
-                     subject = NULL,
+                     datasubject = NULL,
                      description = NULL,
                      language = NULL,
                      datasource = NULL,
@@ -62,6 +63,10 @@ dataset  <- function(x,
 
   arguments <- list(...)
 
+  if (is.null(datasubject)) {
+    datasubject <- subject_create(NULL)
+  }
+
   as_dataset(x,
              author=author,
              title=title,
@@ -69,33 +74,40 @@ dataset  <- function(x,
              identifier=identifier,
              year=year,
              version=version,
-             subject=subject,
+             datasubject=datasubject,
              description=description,
              language=language,
              datasource=datasource,
              rights=rights)
 }
 
+
 #' @keywords internal
 new_dataset <- function (x,
                          DataBibentry,
-                         subject) {
+                         datasubject) {
 
   validate_dataset(x = x,
                    DataBibentry = DataBibentry,
-                   subject = subject)
+                   datasubject = datasubject)
+
+  if (ncol(x) >0) {
+    DataStructureNest <- lapply (1:ncol(x),
+                                 function(col)
+                                   initialise_dsd(df = x, col) )
+    DSD <- DataStructureNest[[1]]
+    if (ncol(x)>1) {
+      for ( i in 2:ncol(x)) DSD <- c(DSD,  DataStructureNest[[i]])
+    }
 
 
-  DataStructureNest <- lapply (1:ncol(x), function(col) initialise_dsd(df = x, col) )
-
-  DataStructure <- DataStructureNest[[1]]
-  if (ncol(x)>1) {
-    for ( i in 2:ncol(x)) DataStructure <- c(DataStructure,  DataStructureNest[[i]])
+  } else {
+    DSD <- list()
   }
 
   attr(x, "DataBibentry") <- DataBibentry
-  attr(x, "Subject") <- subject
-  attr(x, "DataStructure") <- DataStructure
+  attr(x, "Subject") <- datasubject
+  attr(x, "DataStructure") <- DSD
   class(x) <- c("dataset", class(x))
[…]
