+++ R/as_dataset.R
@@ -9,7 +9,7 @@ as_dataset <- function(x,
                        identifier = NULL,
                        description = NULL,
                        version = NULL,
-                       subject = NULL,
+                       datasubject = NULL,
                        language = NULL,
                        datasource = NULL,
                        rights = NULL) {
@@ -27,7 +27,7 @@ as_dataset.data.frame <- function(x,
                                   identifier = NULL,
                                   description = NULL,
                                   version = NULL,
-                                  subject = NULL,
+                                  datasubject = NULL,
                                   language = NULL,
                                   datasource = NULL,
                                   rights = NULL) {
@@ -45,10 +45,10 @@ as_dataset.data.frame <- function(x,
                                datasource = datasource,
                                rights = rights)
 
-  if (is.null(subject)) subject <- new_Subject("")
+  if (is.null(datasubject)) datasubject <- new_Subject("")
 
   new_dataset(x,
               DataBibentry = DataBibentry,
-              subject = subject)
+              datasubject = datasubject)
 
 }
