+++ R/data-dataset_namespace.R
@@ -1,0 +1,11 @@
+#' @title Popular Namespace
+#'
+#' @description Some popular namespace prefix abbreviations for the linking
+#' of datasets.
+#'
+#' @format
+#' The dataset_namespace has 2 columns and 20 rows. The \code{prefix} column
+#' contains the generally used abbreviations of ontologies, vocabularies and
+#' other common standard definitions, the \code{uri} contains their unique
+#' resource indicators.
+"dataset_namespace"
