+++ man/ttl_dataset_write.Rd
@@ -1,0 +1,41 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ttl_dataset_write.R
+\name{ttl_dataset_write}
+\alias{ttl_dataset_write}
+\title{Write a dataset into Turtle serialisation}
+\usage{
+ttl_dataset_write(
+  tdf,
+  ttl_namespace = NULL,
+  file_path = NULL,
+  overwrite = TRUE
+)
+}
+\arguments{
+\item{tdf}{A dataset in exactly three columns.}
+
+\item{overwrite}{If the file exists, overwrite it? Defaults to \code{TRUE}.}
+
+\item{ttle_namespace}{The namespace definitions of the dataset}
+
+\item{file}{The path to the file that should be written or appended.}
+}
+\value{
+A text file with the prefix and the observation serialisations.
+}
+\description{
+Write a dataset into a Turtle serialisation.
+}
+\examples{
+testtdf <- data.frame (s = c("eg:o1", "eg:01", "eg:02"),
+                       p = c("a", "eg-var:", "eg-var"),
+                       o = c("qb:Observation",
+                             "\"1\"^^<xs:decimal>",
+                             "\"2\"^^<xs:decimal>"))
+
+examplefile <- file.path(tempdir(), "ttl_dataset_write.ttl")
+
+ttl_dataset_write(tdf=testtdf, file_path = examplefile)
+
+readLines(examplefile)
+}
