+++ man/dataset_namespace.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data-dataset_namespace.R
+\docType{data}
+\name{dataset_namespace}
+\alias{dataset_namespace}
+\title{Popular Namespace}
+\format{
+The dataset_namespace has 2 columns and 20 rows. The \code{prefix} column
+contains the generally used abbreviations of ontologies, vocabularies and
+other common standard definitions, the \code{uri} contains their unique
+resource indicators.
+}
+\usage{
+dataset_namespace
+}
+\description{
+Some popular namespace prefix abbreviations for the linking
+of datasets.
+}
+\keyword{datasets}

[…]
