+++ data-raw/prepare_iris_dataset.R
@@ -15,7 +15,16 @@ iris_dataset <- dataset(
   description = "The famous (Fisher's or Anderson's) iris data set."
 )
 
+DSD <- DataStructure(iris_dataset)
+DSD$Sepal.Length$label <- "The sepal length of iris specimen in centimeters."
+DSD$Petal.Length$label <- "The petal length of iris specimen in centimeters."
+DSD$Sepal.Width$label <- "The sepal width of iris specimen in centimeters."
+DSD$Petal.Width$label <- "The petal width of iris specimen in centimeters."
+DSD$Species$label <- "The iris species of the observed plant."
+DataStructure_update(x = iris_dataset, value=DSD)
+
 describe(iris_dataset)
+attr(iris_dataset, "DataStructure") <- DSD
 usethis::use_data(iris_dataset, overwrite = TRUE)
 
 #snakecase::to_title_case("Edgar Anderson's Iris Data (For Testing the dataset R package)")
