+++ data-raw/prepare_dataset_namespace.R
@@ -1,0 +1,9 @@
+
+library(tidyverse)
+dataset_namespace <- readxl::read_excel(file.path("data-raw", "namespace.xlsx")) %>%
+  mutate ( prefix = stringr::str_trim(prefix, "both"),
+           uri  = stringr::str_trim(uri, "both")) %>%
+  arrange(prefix)
+
+
+usethis::use_data(dataset_namespace, overwrite = TRUE)
