+++ man/iris_dataset.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/iris_dataset.R
+% Please edit documentation in R/data-iris_dataset.R
 \docType{data}
 \name{iris_dataset}
 \alias{iris_dataset}
