+++ cran-comments.md
@@ -11,6 +11,7 @@ Appveyor:
 
 r_hub:
 *  Windows Server 2022 x64 (build 20348); R Under development (unstable)
+*  Windows Server 2022, R-devel, 64 bit
 *  Fedora Linux, R-devel, clang, gfortran
 *  Ubuntu Linux 20.04.1 LTS, R-release, GCC
 *  r-release-macosx-arm64|4.3.0|macosx|macOS 13.3.1 (22E261)|Mac mini|Apple M1||en_US.UTF-8|macOS 11.3|clang-1403.0.22.14.1|GNU Fortran (GCC) 12.2.0
