+++ R/dataset.R
@@ -2,7 +2,7 @@
 #'
 #' @param x An R object that contains the data of the dataset (a data.frame or
 #' inherited from [`data.frame`][base::data.frame()], for example,
-#' [tibble](tibble()), tsibble, [data.table](data.table::data.table()).
+#' [tibble::tibble()], [tsibble::tsibble()], [DT::datatable()].
 #' @param author A single person or a vector of persons as authors, declared with
 #' \code{\link[utils:person]{person}}.
 #' @param title The title of the dataset.
