+++ R/head-tail.R
@@ -13,6 +13,7 @@
 #' # Subsetting the top n rows (observations)
 #' head(iris_dataset, n=3)
 #' @export
+
 head.dataset <- function(x, n=6L, ...) {
 
   if(is.null(x)) return(x)
@@ -30,15 +31,16 @@ head.dataset <- function(x, n=6L, ...) {
 #' (or 1, for non-dimensioned objects). A \code{logical} is silently coerced to
 #' integer. Values specify the indices to be selected in the corresponding
 #' dimension (or along the length) of the object. A positive value of \code{n[i]}
-#' includes the first/last \code{n[i]} indices in that dimension, while a negative
-#' value excludes the last/first abs(n[i]), including all remaining indices.
+#' includes the first/last \code{n[i]} indices in that dimension,
+#' while a negative value excludes the last/first
+#' \code{abs(n[i])}, including all remaining indices.
 #' \code{NA} or non-specified values (when \code{length(n) < length(dim(x))}) select all
 #' indices in that dimension. Must contain at least one non-missing value.
 #' @param keepnums in each dimension, if no names in that dimension are present,
 #' create them using the indices included in that dimension.
 #' Ignored if \code{dim(x)} is \code{NULL} or its length 1.
-#' @param ... arguments to be passed to or from other methods.
 #' @method tail dataset
+#' @importFrom utils tail head
 #' @examples
 #'
 #' # Subsetting the last n rows (observations)
