+++ man/head.dataset.Rd
@@ -16,12 +16,13 @@
 (or 1, for non-dimensioned objects). A \code{logical} is silently coerced to
 integer. Values specify the indices to be selected in the corresponding
 dimension (or along the length) of the object. A positive value of \code{n[i]}
-includes the first/last \code{n[i]} indices in that dimension, while a negative
-value excludes the last/first abs(n\link{i}), including all remaining indices.
+includes the first/last \code{n[i]} indices in that dimension,
+while a negative value excludes the last/first
+\code{abs(n[i])}, including all remaining indices.
 \code{NA} or non-specified values (when \code{length(n) < length(dim(x))}) select all
 indices in that dimension. Must contain at least one non-missing value.}
 
-\item{...}{arguments to be passed to or from other methods.}
+\item{...}{Ignored.}
 
 \item{keepnums}{in each dimension, if no names in that dimension are present,
 create them using the indices included in that dimension.
