+++ NAMESPACE
@@ -67,6 +67,8 @@ importFrom(assertthat,assert_that)
 importFrom(stats,setNames)
 importFrom(utils,bibentry)
 importFrom(utils,download.file)
+importFrom(utils,head)
 importFrom(utils,object.size)
 importFrom(utils,person)
 importFrom(utils,read.csv)
+importFrom(utils,tail)
