+++ NAMESPACE
@@ -66,6 +66,7 @@ export(xsd_convert)
 importFrom(assertthat,assert_that)
 importFrom(stats,setNames)
 importFrom(utils,bibentry)
+importFrom(utils,data)
 importFrom(utils,download.file)
 importFrom(utils,head)
 importFrom(utils,object.size)
