+++ R/ttl_dataset_write.R
@@ -27,7 +27,7 @@ ttl_dataset_write <- function(tdf,
                               overwrite = TRUE) {
 
   ## load dataset_namespace
-  default_namespace <- getdata(dataset_namespace)
+  default_namespace <- getdata("dataset_namespace")
   default_namespace <- dataset_namespace[which(dataset_namespace$prefix %in% c("rdf:", "rdfs:", "owl:", "qb:", "dcat:")),]
 
   ## validate dataset
