+++ man/subsetting.Rd
@@ -18,7 +18,7 @@
 
 \item{i, j}{Row and column indices. If \code{j} is omitted, \code{i} is used as column index.}
 
-\item{drop}{ogical. If \code{TRUE} the result is coerced to the lowest possible
+\item{drop}{logical. If \code{TRUE} the result is coerced to the lowest possible
 dimension. The default is to drop if only one column is left, but not to
 drop if only one row is left.}
 
