+++ vignettes/motivation.Rmd
@@ -90,9 +90,9 @@ my_dataset
 ### The data cube model
 The tidy data is designed to be easily handled in relational databases and it is an ideal format for analysis in a closed system – for example, when all your data for the analysis is stored in a well-designed relational database. This semantics is, however, insufficient for exchanging data with several data sources. This problem has been attracting solutions for a long time by statistical agencies and designers of hyperlinked, web-based products. 
 
-The dataset package is inspired by how statisticians have been organizing data for publications for a long time: the data cube model. The data cube model has additional, non-normative prescriptions for the structure of the (tidy) dataset:  columns must be grouped into measurements, dimensions, and attributes. The dimensions are non-measured variables, such as reference geographical areas and reference time periods, that are usually used to filter out subsets (slices) from a datacube. Attributes provide information about the measurements, such as the unit of information and estimation method, and usually can be seen as metadata for modelling purposes: they contain observations (measurements) level metadata that is also often used for filtering (selecting actual, imputed or missing cases, for example.)  Measurements are, in fact, the observations that we want to analyze.
+The dataset package is inspired by how statisticians have been organizing data for publications for a long time: the data cube model. The data cube model has additional, non-normative prescriptions for the structure of the (tidy) dataset:  columns must be grouped into measurements, dimensions, and attributes. The dimensions are non-measured variables, such as reference geographical areas and reference time periods, that are usually used to filter out subsets (slices) from a datacube. Attributes provide information about the measurements, such as the unit of information and estimation method, and usually can be seen as metadata for modelling purposes: they contain observations (measurements) level metadata that is also often used for filtering (selecting actual, imputed or missing cases, for example.)  Measurements are, in fact, the observations that we want to analyse.
 
-The data cube model has long been the preferred format to organize data for analytical use by SDMX. Statistical agencies exchange with each other and publish data in datasets that conform to the data cube model. Organizing your R tabular data like statisticians makes your work easier to synchronize with statistical data sources: you can download or exchange data in this format. Following the data cube model helps you to understand the data more easily and bring your data to be analyzed to a tidy form in one or two steps --- usually using attributes for subsetting, dimensions for grouping, and measurements to carry out arithmetic calculations like computing the group average values.
+The data cube model has long been the preferred format to organize data for analytical use by SDMX. Statistical agencies exchange with each other and publish data in datasets that conform to the data cube model. Organizing your R tabular data like statisticians makes your work easier to synchronize with statistical data sources: you can download or exchange data in this format. Following the data cube model helps you to understand the data more easily and bring your data to be analysed to a tidy form in one or two steps --- usually using attributes for subsetting, dimensions for grouping, and measurements to carry out arithmetic calculations like computing the group average values.
 
 ## Interoperability
 While tidying aims and preventing reinventing the wheel and many times repeating the work of bringing the data from data storage, such as a database, into a form that an analyst can work with, reproducible research has further aims that add to the quality and inefficiency of the analysis: it aims to make the tidy datasets and the research output easier to reuse and to review and replicate for quality control and peer review.
