+++ vignettes/data-linking.Rmd
@@ -25,7 +25,7 @@ library(dataset)
 For linking data over the World Wide Web, it is essential to have unambiguous, unique names for the things that your dataset refers to.  The cross-comparison of statistical data around the globe is possible, because the Statistical Data and Metadata eXchange (SDMX) information model, among many things, have standardised the observed values, their aggregation dimensions and measurement or description 
 attributes. If we break up a statistical dataset by the dimension of time, geography and sex (gender), then we can use the very same time period, geography and sex definitions and names to make sure that we can join data about women in the Netherlands for the time period 2022 correctly across datasets; or add data about men and women in Belgium in a way that allows for a semantically correct comparison with Dutch women.
 
-To allow for a linking of datasets, we need unique resource identifiers (URIs) for the definition of time, geography and sex dimensions of the statistical dataset, and uniform variables names. For example, the `http://purl.org/linked-data/sdmx/2009/dimension#sex` URI provides a unique identifier for any variable that breaks up a statistical dataset by sex; conveniently, the URI also resolves to a human-readable URL (or web page) where we can read how to code observed population into male, female and other groups so that the comparison will be meaningful. To uniquely identify a cell in a dataset, we must have a globally unique identifier for each of its rows (observations) and columns (variables, inclusing dimensions and attributes of observed values.) 
+To allow for a linking of datasets, we need unique resource identifiers (URIs) for the definition of time, geography and sex dimensions of the statistical dataset, and uniform variables names. For example, the `http://purl.org/linked-data/sdmx/2009/dimension#sex` URI provides a unique identifier for any variable that breaks up a statistical dataset by sex; conveniently, the URI also resolves to a human-readable URL (or web page) where we can read how to code observed population into male, female and other groups so that the comparison will be meaningful. To uniquely identify a cell in a dataset, we must have a globally unique identifier for each of its rows (observations) and columns (variables, including dimensions and attributes of observed values.) 
 
 ## Namespace and URI
 In linking data, the namespace helps records have a unique resource identifier (URI). The namespace is the shared part of the URI and is often abbreviated. The abbreviation of the namespace is the prefix.
