+++ DESCRIPTION
@@ -55,4 +55,4 @@ Suggests:
     tibble
 VignetteBuilder: knitr
 Config/testthat/edition: 3
-Language: en-US
+Language: en-GB
