+++ R/data-iris_dataset.R
@@ -1,7 +1,7 @@
 #' @title Edgar Anderson's Iris Data
 #'
 #' @description This famous (Fisher's or Anderson's) iris data set gives the measurements in
-#' centimeters of the variables sepal length and width and petal length and
+#' centimetres of the variables sepal length and width and petal length and
 #' width, respectively, for 50 flowers from each of 3 species of iris.
 #' The species are \emph{Iris setosa}, \emph{versicolor}, and \emph{virginica}.
 #' This is a replication of \code{datasets::\link[datasets:iris]{iris}} as
