+++ tests/testthat/test-dataset_ttl_write.R
@@ -1,14 +1,14 @@
-testfile <- file.path(tempdir(), "test-ttl_dataset_write.ttl")
+testfile <- file.path(tempdir(), "test-dataset_ttl_write.ttl")
 testtdf <- data.frame ( s = c("eg:o1", "eg:01", "eg:02"),
                         p = c("a", "eg-var:", "eg-var"),
                         o = c("qb:Observation", "\"1\"^^<xs:decimal>", "\"2\"^^<xs:decimal>") )
 
-ttl_dataset_write(tdf=testtdf,
+dataset_ttl_write(tdf=testtdf,
                   ttl_namespace = NULL,
                   file_path=testfile,
                   overwrite=TRUE)
 
-test_that("ttl_dataset_write() works:", {
+test_that("dataset_ttl_write() works:", {
   expect_true(file.exists(testfile))
   expect_true(grepl("@prefix  owl:", readLines(testfile)[1]))
   expect_equal(sum(
@@ -19,7 +19,7 @@ test_that("ttl_dataset_write() works:", {
 , 1)
 })
 
-test_that("ttl_dataset_write() validation works:", {
-  expect_error(ttl_dataset_write(tdf=iris, file = tempfile()))
-  expect_error(ttl_dataset_write(tdf=list(), file = tempfile()))
+test_that("dataset_ttl_write() validation works:", {
+  expect_error(dataset_ttl_write(tdf=iris, file = tempfile()))
+  expect_error(dataset_ttl_write(tdf=list(), file = tempfile()))
 })
