+++ vignettes/dataset.Rmd
@@ -19,7 +19,7 @@ knitr::opts_chunk$set(
 library(dataset)
 ```
 
-In the R language, datasets are usually contained in a [data.frame()](https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/data.frame) object, or in one of their modernized versions.  For example, [tibble::tibble()(https://tibble.tidyverse.org/reference/tibble.html) or [data.table::data.table()](https://rdatatable.gitlab.io/data.table/articles/datatable-intro.html) are inherited from the base data.frame().
+In the R language, datasets are usually contained in a [data.frame()](https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/data.frame) object, or in one of their modernized versions.  For example, [tibble::tibble()](https://tibble.tidyverse.org/reference/tibble.html) or [data.table::data.table()](https://rdatatable.gitlab.io/data.table/articles/datatable-intro.html) are inherited from the base data.frame().
 
 > This documentation is not updated yet to the development version of the [dataset] package.
 
@@ -31,122 +31,119 @@ attr(foo, "Title") <- "My Foo Object"
 attributes(foo)
 ```
 
-For reproducible research, publication, or linking resources on the web, the standardization of metadata is critically important. The aim `dataset()` class is a modernized data.frame that has standardized attributes.
+Metadata standardisation is critically essential for reproducible research, publication, or linking resources on the web. The aim `dataset()` class is the creation of semantically enriched data frames with as much interoperability as possible across various sub-classes of the base R `data.frame()`. 
 
-```{r iris}
-describe(iris_dataset)
-publisher(iris_dataset)
+```{r mtcars}
+head (dataset(mtcars, 
+        title="The Motor Trend [mtcar] Dataset", 
+        author=person("Motor Trend Magazine"), 
+        year=1974, 
+        publisher="Motor Trend Magazine" ))
+
 ```
 
-```{r, include=FALSE, message=FALSE}
-library(here)
-library(knitr)
-```
+```{r mtcars-ttl, warning=FALSE}
+temp_ttl_file  <- file.path(tempdir(), "temp_ttl.ttl")
+mtcars_dataset <- dataset(mtcars, 
+        title="The Motor Trend [mtcar] Dataset", 
+        author=person("Motor Trend Magazine"), 
+        year=1974, 
+        publisher="Motor Trend Magazine")
 
-According to the [RDF Data Cube Vocabulary](https://www.w3.org/TR/vocab-data-cube/) DataSet is a collection of statistical data that corresponds to a defined structure. The data in a data set can be roughly described as belonging to one of the following kinds:
+mtcars_namespace <- dataset_namespace[
+  dataset_namespace$prefix %in% c("owl:", "rdf:", "rdfs:", "qb:", "eg:"), ]
 
-- `Observations`: these are the measured values, and the cells of a data frame object in R.
-- `Organizational structure`: To locate an observation within the hypercube, one has at least to know the value of each dimension at which the observation is located, so these values must be specified for each observation. Datasets can have additional organizational structure in the form of slices as described in section 7.2.
-- `Structural metadata`: Metadata to interpret the data. What is the unit of measurement? Is it a normal value or a series break? Is the value measured or estimated? These metadata are provided as attributes and can be attached to individual observations, or to higher levels.
-- `Reference metadata`: Metadata that describes the dataset as a whole, such as categorization of the dataset, its publisher, or an endpoint where it can be accessed.
-
-| Information  | dataset |
-|     :--:     |  ----|
-| dimensions   |  first column section of the dataset   |
-| measurements |  second column section of the dataset  |
[…]
