+++ vignettes/RDF.Rmd
@@ -15,43 +15,162 @@ knitr::opts_chunk$set(
   comment = "#>",
   out.width = '90%'
 )
+if (file.exists(file.path(tempdir(), "example_ttl.ttl"))) {
+  file.remove(file.path(tempdir(), "example_ttl.ttl"))
+}
+
 library(here)
-library(readxl)
-library(kableExtra)
 ```
 
-```{r setup}
-library(dataset)
-```
-
-> This documentation is not updated yet to the development version of the [dataset] package.
-
-Our datasets are defined in a way that their dimensions can be easily and unambiguously reduced to triples for RDF applications; they can be easily serialized to, or synchronized with semantic web applications with the [rdflib](https://cran.r-project.org/package=rdflib/vignettes/rdf_intro.html) package^[Carl Boettiger: [A tidyverse lover’s intro to RDF](https://cran.r-project.org/package=rdflib/vignettes/rdf_intro.html)].
-
-Read more about how we adopted to datacube model of SDMX and the [RDF Data Cube Vocabulary](https://www.w3.org/TR/vocab-data-cube/) for datasets as R objects in the [The dataset S3 Class](https://dataset.dataobservatory.eu/articles/RDF.html) vignette article.
-
-Because our datasets conform the tidy data concept, they can be reduced into long-form triples.
+Long-form triples are tidy datasets with an explicit row (observation) identifier among the columns.
 
 
 
 &nbsp;          | &nbsp;  | &nbsp;      | &nbsp;
 ----------------|---------|-------------|-------------
-**RDF**         | subject | predicate   | object
 **JSON**        | object  | property    | value
 **spreadsheet** | row id  | column name | cell
 **data.frame**  | key     | variable    | measurement
 **data.frame**  | key     | attribute   | value
-
+**RDF**         | subject | predicate   | object
 
 Table source: [rdflib](https://cran.r-project.org/package=rdflib/vignettes/rdf_intro.html)
 
-## Dimension reductions of the dataset
+```{r setup}
+library(dataset)
+```
 
-Our datasets are tidy.
+Let's take a small subset of the `iris_dataset()`, which is the semantically enriched version of the base R `iris` dataset. Limiting the the dataset to the top 3 rows, we have exactly 2 x 5 = 10 data cells.
 
-```{r importexamplexlsx}
-example_df <- readxl::read_excel(
-  system.file("extdata", "rdf_example.xlsx", package = "dataset"),
-  sheet = "dataset-wide")
+```{r head-iris}
+head(iris_dataset, 2)
+```
+
+```{r head-iris-xsd}
+xsd_convert(head(iris_dataset, 2))
+```
+
+Let us arrange this to subject-predicate-object triples.
+
+```{r dataset-to-triples}
+iris_triples <- dataset_to_triples(xsd_convert(head(iris_dataset,2)))
+iris_triples
+```
+
+We receive 2x5 = 10 rows; each with an identifier. The identifiers are made from `row.names()`, and we have exactly 5 statements about the first observation (`iris:o1`), and 5 statements about the second (`iris:o2`). Each statement simply states the observed value.
+
+
+```{r define-vars}
+iris_triples$p <- paste0("iris:", iris_triples$p)
+iris_triples
+```
+
+
+
+```{r row-names}
+row.names(head(iris_dataset,2))
+```
+```{r write-to-ttle}
+vignette_temp_file <- file.path(tempdir(), "example_ttl.ttl")
+dataset_ttl_write(dataset_to_triples(iris_triples), 
+                  file_path = vignette_temp_file)
+```
+
+
+We see a standard metadata file expressed in the [Turtle](https://www.w3.org/TR/rdf12-turtle/) language. The definitions are
[…]
