### 2024-01-08 41ad804
Update suggested packages for vignette documentation.
