+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Date: 2024-01-02
-Version: 0.2.9.004
+Date: 2024-01-08
+Version: 0.3.0
 Authors@R: 
     c(person(given = "Daniel", family = "Antal", 
            email = "daniel.antal@dataobservatory.eu", 
@@ -52,7 +52,11 @@ Suggests:
     statcodelists,
     testthat (>= 3.0.0),
     tidyr,
-    tibble
+    tibble,
+    nycflights,
+    nycflights13,
+    tsibble,
+    data.table
 VignetteBuilder: knitr
 Config/testthat/edition: 3
 Language: en-GB
