+++ _pkgdown.yml
@@ -42,15 +42,19 @@ articles:
   - title: "Metadata"
     desc: >
      Our `dataset` R package creates new classes and an interface to easily add,
-     retrieve or modify descriptive or structural metadata to a dataset.
+     retrieve or modify descriptive, provenance or structural metadata to a
+     dataset.
     contents:
+      - metadata
       - datacite
       - dublincore
+      - provenance
   - title: "Data Linking"
     desc: >
        Linking datasets.
     contents:
        - data-linking
+       - RDF
   - title: "Planned functionality"
     desc: >
      This package is in an early development phase. The current dataset S3 class is inherited
@@ -60,7 +64,6 @@ articles:
          [data.table](https://cran.r-project.org/web/packages/data.table/index.html) package remains a top development priority.
     contents:
       - survey
-      - RDF
 reference:
   - title: "Dataset"
     desc: >
