+++ DESCRIPTION
@@ -53,7 +53,6 @@ Suggests:
     testthat (>= 3.0.0),
     tidyr,
     tibble,
-    nycflights,
     nycflights13,
     tsibble,
     data.table
