+++ man/rights.Rd
@@ -27,7 +27,7 @@ R object.
 }
 \details{
 \code{Rights} corresponds to
-\href{https://purl.org/dc/elements/1.1/rights}{dct:rights} and
+\href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/rights}{dct:rights} and
 \code{\link{datacite}} Rights. Information about rights held in and over the resource.
 Typically, rights information includes a statement about various property rights associated with the resource,
 including intellectual property rights.
