+++ man/dublincore.Rd
@@ -5,7 +5,7 @@
 \alias{as_dublincore}
 \alias{is.dublincore}
 \alias{is.dublincore.dublincore}
-\title{Add or get DublinCore metadata}
+\title{Add or get Dublin Core metadata}
 \source{
 \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/format/}{	DCMI Metadata Terms}.
 }
@@ -35,10 +35,10 @@ is.dublincore(x)
 \method{is.dublincore}{dublincore}(x)
 }
 \arguments{
-\item{title}{\href{https://purl.org/dc/elements/1.1/title}{dct:title}, a name given to the resource.
+\item{title}{\href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/title/}{dct:title}, a name given to the resource.
 \code{\link{datacite}} allows the use of alternate titles, too. See \code{\link{dataset_title}}.}
 
-\item{creator}{An entity primarily responsible for making the resource. \href{https://purl.org/dc/elements/1.1/creator}{dct:creator}
+\item{creator}{An entity primarily responsible for making the resource. \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/creator/}{dct:creator}
 Corresponds to \code{Creator} in \code{\link{datacite}}. See \code{\link{creator}}.}
 
 \item{identifier}{An unambiguous reference to the resource within a given context.
@@ -73,7 +73,7 @@ Format element.}
 \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/contributor/}{DCMI: Contributor}.}
 
 \item{date}{Corresponds to a point or period of time associated with an event in the
-lifecycle of the resource. \href{https://purl.org/dc/elements/1.1/date}{dct:date}.
+lifecycle of the resource. \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/date}{dct:date}.
 \code{Date} is also recommended for
 discovery in \code{\link{datacite}}, but it requires a different formatting.}
 
@@ -81,13 +81,14 @@ discovery in \code{\link{datacite}}, but it requires a different formatting.}
 \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/language/}{DCMI: Language}.}
 
 \item{relation}{A related resource. Recommended best practice is to identify the related
-resource by means of a string conforming to a formal identification system. See: \href{https://purl.org/dc/elements/1.1/relation}{dct:relation}.
+resource by means of a string conforming to a formal identification system.
+See: \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/relation}{dct:relation}.
 Similar to \code{RelatedItem} in \code{\link{datacite}}, which is recommended for discovery.}
 
 \item{format}{The file format, physical medium, or dimensions of the dataset. See
 \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/source/}{DCMI: Format}.}
 
-\item{rights}{Corresponds to \href{https://purl.org/dc/elements/1.1/rights}{dct:rights} and
+\item{rights}{Corresponds to \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/rights}{dct:rights} and
 \code{\link{datacite}} Rights. Information about rights held in and over the resource.
 Typically, rights information includes a statement about various property rights associated with the resource,
 including intellectual property rights. See \code{\link{rights}}.}
@@ -99,7 +100,7 @@ which corresponds to a \code{relatedItem} in the DataCite vocabulary. We use
 
 \item{description}{An account of the resource. It may include but is not limited to:
[…]

[…]
