+++ R/dublincore.R
@@ -15,7 +15,8 @@
 #' The \code{Size} attribute (e.g. bytes, pages, inches, etc.) will automatically added to the dataset.
 #' @param title \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/title/}{dct:title}, a name given to the resource.
 #' \code{\link{datacite}} allows the use of alternate titles, too. See \code{\link{dataset_title}}.
-#' @param creator An entity primarily responsible for making the resource. \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/creator/}{dct:creator}
+#' @param creator An entity primarily responsible for making the resource.
+#' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/creator/}{dct:creator}
 #' Corresponds to \code{Creator} in \code{\link{datacite}}. See \code{\link{creator}}.
 #' @param identifier An unambiguous reference to the resource within a given context.
 #' Recommended practice is to identify the resource by means of a string conforming to an
@@ -34,18 +35,18 @@
 #' "holds, archives, publishes, prints, distributes, releases, issues, or produces" the
 #' code, use the property Contributor/contributorType/hostingInstitution for the code
 #' repository. See \code{\link{publisher}}.
-#' @param subject In \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/subject}{dct:subject}. In
+#' @param subject In \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/subject/}{dct:subject}. In
 #' \code{\link{datacite}} it is a recommended property for discovery. In DataCite, a more complex
 #' referencing is used. See \code{\link{subject}} and create structured Subject objects with
 #' \code{\link{subject_create}}.
 #' @param date Corresponds to a point or period of time associated with an event in the
-#' lifecycle of the resource. \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/date}{dct:date}.
+#' lifecycle of the resource. \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/date/}{dct:date}.
 #' \code{Date} is also recommended for
 #' discovery in \code{\link{datacite}}, but it requires a different formatting.
 #' @param language The primary language of the resource. Allowed values are taken from
 #' IETF BCP 47, ISO 639-1 language code. See \code{\link{language}}. Corresponds to Language in Datacite.
 #' @param format The file format, physical medium, or dimensions of the resource.
-#' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/format}{dct:format}
+#' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/format/}{dct:format}
 #' Examples of dimensions include size and duration. Recommended best practice is to use a controlled
 #' vocabulary such as the list of \href{https://www.iana.org/assignments/media-types/media-types.xhtml}{Internet Media Types, formerly known as MIME}. It is similar to \code{Format} in
 #' \code{\link{datacite}}.
@@ -55,11 +56,11 @@
 #' including intellectual property rights. See \code{\link{rights}}.
 #' @param description An account of the resource. It may include but is not limited to:
 #' an abstract, a table of contents, a graphical representation, or a free-text account of the resource.
-#' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/description}{dct:description}. In
+#' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/description/}{dct:description}. In
 #' \code{\link{datacite}} it is recommended for discovery. See \code{\link{description}}.
[…]
