+++ R/description.R
@@ -2,7 +2,8 @@
 #' @description Get/set the optional \code{Description} property as an attribute to an R object.
 #' @details The \code{Description} is recommended for discovery in DataCite. All additional information that does not
 #' fit in any of the other categories. May be used for technical information. A free text.
-#' Similar to \href{https://purl.org/dc/elements/1.1/description}{dct:description}.
+#' Similar to
+#' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/description/}{dct:description}.
 #' @param x A dataset object created with \code{dataset::\link{dataset}}.
 #' @param value The \code{Description} as a character set.
 #' @param overwrite If the \code{Description} attribute should be overwritten. In case it is set to \code{FALSE},
