+++ R/datacite.R
@@ -1,32 +1,36 @@
 #' @title Add or get DataCite metadata
 #' @description Add or retrieve metadata conforming
 #' the \href{https://schema.datacite.org/}{DataCite Metadata Schema}.
-#' @details DataCite is a leading global non-profit organisation that provides persistent identifiers
-#' (DOIs) for research data and other research outputs. Organisations within the research
-#' community join DataCite as members to be able to assign DOIs to all their research
-#' outputs. This way, their outputs become discoverable, and associated metadata is made.
-#' available to the community.
+#' @details DataCite is a leading global non-profit organisation that provides
+#' persistent identifiers (DOIs) for research data and other research outputs.
+#' Organisations within the research community join DataCite as members to be
+#' able to assign DOIs to all their research outputs. This way, their outputs
+#' become discoverable, and associated metadata is made available to the community.
 #' @details The \code{ResourceType} property will be by definition "Dataset".
-#' The \code{Size} attribute (e.g. bytes, pages, inches, etc.) will automatically added to the dataset.
-#' @param Title The name(s) or title(s) by which a resource is known. May be the title of a dataset
-#' or the name of a piece of software. Similar to \href{https://purl.org/dc/elements/1.1/title}{dct:title}.\cr
+#' The \code{Size} attribute (e.g. bytes, pages, inches, etc.) will automatically added to
+#' the dataset.
+#' @param Title The name(s) or title(s) by which a resource is known. May be the
+#' title of a dataset or the name of a piece of software.
+#' Similar to
+#' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/title/}{dct:title}.\cr
 #' @param Creator The main researchers involved in producing the data, or the authors of the publication, in
 #' priority order. To supply multiple creators, repeat this property.
 #' @param Identifier The Identifier is a unique string that identifies a resource. For software, determine
 #' whether the identifier is for a specific version of a piece of software, (per the
 #' \href{https://force11.org/info/software-citation-principles-published-2016/}{Force11 Software Citation Principles},
-#' or for all versions. Similar to \code{dct:title} in \code{\link{dublincore}}.
+#' or for all versions. Similar to \code{dct:title} in [dublincore()].
 #' @param Publisher The name of the entity that holds, archives, publishes prints, distributes,
 #' releases, issues, or produces the resource. This property will be used to formulate the
 #' citation, so consider the prominence of the role. For software, use Publisher for the
 #' code repository. If there is an entity other than a code repository, that "holds, archives,
 #' publishes, prints, distributes, releases, issues, or produces" the code, use the property
 #' Contributor/contributorType/ hostingInstitution for the code repository. Corresponds to dct:Publisher in
-#' \code{\link{dublincore}}.
+#' [dublincore()].
 #' @param PublicationYear The year when the data was or will be made publicly available in
-#' \code{YYYY} format.See \code{\link{publication_year}}.
+#' \code{YYYY} format.See [publication_year()].
 #' @param Subject Recommended for discovery. Subject, keyword, classification code, or key
-#' phrase describing the resource. Similar to \href{https://purl.org/dc/elements/1.1/subject}{dct:subject}. \cr
+#' phrase describing the resource. Similar to
+#' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/subject/}{dct:subject}. \cr
 #' Use \code{\link{subject}} to properly add a key phrase from a controlled vocabulary
 #' and create structured Subject objects with \code{\link{subject_create}}.
[…]
