+++ man/datasource_get.Rd
@@ -28,7 +28,7 @@ R object. Do not confuse with the base R \code{source()} function.
 }
 \details{
 The \code{Source} is a related resource from which the described resource is
-derived. See \href{https://purl.org/dc/elements/1.1/source}{dct:source}. In Datacite,
+derived. See \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/source}{dct:source}. In Datacite,
 the source is described by a \code{relatedIdentifierType} with the property
 \code{relationType="isDerivedFrom"}.
 }

[…]
