### 2024-01-08 c1bf5ee
URIs and links formatting
