+++ man/datasource_get.Rd
@@ -28,8 +28,9 @@ R object. Do not confuse with the base R \code{source()} function.
 }
 \details{
 The \code{Source} is a related resource from which the described resource is
-derived. See \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/source}{dct:source}. In Datacite,
-the source is described by a \code{relatedIdentifierType} with the property
+derived. See
+\href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/elements11/source/}{dct:source}.
+In Datacite, the source is described by a \code{relatedIdentifierType} with the property
 \code{relationType="isDerivedFrom"}.
 }
 \examples{
