+++ cran-comments.md
@@ -3,7 +3,6 @@
 Local:
 * Windows10 x86_64-w64-mingw32 (64-bit), R version 4.3.2, locally.
 * Ubuntu 22.04.1 LTS x86_64-pc-linux-gnu (64-bit), R version 4.2.1, locally
-0 errors ✔ | 0 warnings ✔ | 0 notes ✔
 
 Appveyor:
 * Windows Server 2019 x64 (build 17763); R version 4.3.2 Patched (2023-12-01 r85659 ucr)
@@ -15,4 +14,6 @@ r_hub:
 *  Ubuntu Linux 20.04.1 LTS, R-release, GCC
 *  r-release-macosx-arm64|4.3.0|macosx|macOS 13.3.1 (22E261)|Mac mini|Apple M1||en_US.UTF-8|macOS 11.3|clang-1403.0.22.14.1|GNU Fortran (GCC) 12.2.0
 
+0 errors ✔ | 0 warnings ✔ | 0 notes ✔
+No extra comments
 
