+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Date: 2024-01-08
-Version: 0.3.0
+Date: 2024-01-27
+Version: 0.3.1
 Authors@R: 
     c(person(given = "Daniel", family = "Antal", 
            email = "daniel.antal@dataobservatory.eu", 
@@ -27,7 +27,7 @@ URL: https://github.com/dataobservatory-eu/dataset
 BugReports: https://github.com/dataobservatory-eu/dataset/issues
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.0
 Depends: 
     R (>= 2.10)
 LazyData: true
