### 2024-01-27 70025d1
Correction of internal function cbind, which referred to a non-existing cbind generic method.
