### 2024-01-27 b1dca41
Update README, codemeta, citation, DOI.
