### 2024-02-28 #553 @ldecicco-USGS
Hi @antaldaniel Since you mentioned "The new version (which is an entire rewrite since the first review) ", we're going to treat this as a new submission and get new reviewers. Can you work on the 2 outstanding issues above while I look for a new editor?

Thanks @annakrystalli for the initial work!
