### 2024-10-15 #553 @adamhsparks
Hi @antaldaniel, just checking in, the editor checks indicate a few minor issues that could be addressed fairly easily, I think. Are you in a position to fix these issues so we can resume this review?
