### 2024-10-16 #553 @antaldaniel
Yes, I am. I just created in the last days a plan to improve this package, and add an inheritated package for a specific use, because I think that the mass use was missing that would have created interest and contributions to the package.  I will make these small changes, but also include for review a new conceptual vignette to explain better the mission statement.
