### 2024-10-18 cc1b636
0.3.20: new requirement settings
