+++ NEWS.md
@@ -1,3 +1,7 @@
+# dataset 0.3.2
+
+- [New Requirement](https://dataset.dataobservatory.eu/articles/new-requirements.html) setting for further works.
+
 # dataset 0.3.0
 
 - Released on CRAN.

[…]
