+++ vignettes/new-requirements.Rmd
@@ -1,5 +1,6 @@
 ---
 title: "New Requirements"
+subtitle: "New concept for the dataset package"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{New Requirements}
@@ -18,9 +19,6 @@ knitr::opts_chunk$set(
 library(dataset)
 ```
 
-
-New concept for the dataset package
-
 The first CRAN release and RopenSci review brought very useful experience and feedback. The dataset package had been defined with a very broad requirement. While the very general requirement setting has advantages, a clear disadvantage is that without a specific use case, it is difficult to raise enough user and co-developer interest.
 
 In line with some of the legitimate criticism of version 0.1.0—0.3.1, I envision a dataset package that has more specific inheritance packages that work with datasets in a more specific use case.
@@ -28,6 +26,7 @@ In line with some of the legitimate criticism of version 0.1.0—0.3.1, I envisi
 It is probably too wide of a claim to create a package that will bring the base R data.frame object in line with any disciplinary requirements of datasets. The original concept closely followed the SDMX (statistical) dataset definition to the extent that one reviewer recommended the `datacube` name for the package. In light of some further use experiences, this is a valid criticism because datasets used in digital humanities, for example, have slightly different specification needs.
 
 R is primarily a statistical environment and language; therefore, broad conformity with the SDMX statistical metadata standards is desirable. However, the dataset package should remain generic enough to support non-statistical datasets. Along these lines, the current aim is to triangulate three packages:
+
 -    A `datacube` package, which follows more closely the SDMX definition of a dataset and the more general, multi-dimensional datacube definition
 -    A `wbdataset` package, which follows more closely the Wikibase Data Model that is increasingly used for digital collections management and other scenarios for statistically not aggregated datasets.
 -    A `dataset` package that is sufficiently generic that both the `datacube` and the `wbdataset` package rely on it as a joint dependency.
@@ -39,7 +38,7 @@ Another important lesson was that the first version of the dataset package wante
 
 ## New requirement settings
 
-The new dataset package would be streamlined to provide a tidier version of the tidy data definition. "Tidy datasets provide a standardised way to link the structure of a dataset (its physical layout) with its semantics (its meaning)." The aim of the dataset package is to improve the semantic infrastructure of tidy datasets beyond the current capabilities of the tidyverse packages, relaxing the exclusive use of the semantic definitions of the SDMX statistical metadata standards.
+The new dataset package would be streamlined to provide a tidier version of the [tidy data definition](https://cran.r-project.org/web/packages/tidyr/vignettes/tidy-data.html). "Tidy datasets provide a standardised way to link the structure of a dataset (its physical layout) with its semantics (its meaning)." The aim of the dataset package is to improve the semantic infrastructure of tidy datasets beyond the current capabilities of the tidyverse packages, relaxing the exclusive use of the semantic definitions of the SDMX statistical metadata standards.
 
 The dataset package and its dependencies (or, more specifically, new extensions) will only work with tidy datasets. In the context of tidy data, "A dataset is a collection of values, usually either numbers (if quantitative) or strings (if qualitative). Values are organised in two ways. Every value belongs to a variable and an observation."
 
@@ -49,21 +48,51 @@ Our aim is to improve the semantic description of the observations and the varia
 -    Every observation must contain the machine-readable definition of the observational unit.
 -    Whenever possible, such an observation should also be supported with a formal persistent identifier such as a URI, but this is not a strict requirement.
 
-For example, when working with this tiny dataset:
+For example, when working with this small dataset:
+
+```{r smallcountrydataset}
[…]
