### 2024-12-11 #553 @emilyriederer
Hi @antaldaniel ! I've taken over from Adam as rotating editor in chief. Just wanted to check in and see if you have a timeline in mind for making the updates discussed above? Or shall we put this review on hold?
