### 2024-12-11 #553 @antaldaniel
I received a message the night before that I cannot find here from Emily. Anyhow:
1. The package was kicked out from CRAN because of a small documenting issue, but I thought that this is a good time to go back for a conceptual issue, as normally rOpenSci suggest to ask for feedback before CRAN release
2. I took the package apart into two packages, one providing the basic classes and concept, and one a practical use case.
3. The package under review here remains dataset, which is thoroughly rewritten in terms of s3 classes: https://github.com/antaldaniel/dataset
4. These were the [requirements](https://dataset.dataobservatory.eu/articles/new_requirements.html) that I set out before the rewrite
5. The connecting package [wbdataset](https://wbdataset.dataobservatory.eu/).

In very short, dataset aims to provide tidier datasets that conform 3NF but their semantics is much richer: their variables can have strict definitions from global name spaces, units of measures, and they are labelled. The dataset as a whole has provenance information, and all Dublin Core - DataCite variables.

What I tried to avoid is what dataspice does, to detach the metadata; these metadata are all stored as attributes in the R object.  I also want a full interoperability with the rdflib package in rOpenSci, because my datasets can be serialised in a very rich way to RDF, and for example, can be sent via API calls to statistical data catalogues like the EU Open Data Portal.
