+++ R/dataset_to_triples.R
@@ -3,7 +3,7 @@
 #' columns \code{s} for subject, \code{p} for predicate and \code{o} for
 #' object.
 #' @param x An R object that contains the data of the dataset (a data.frame or
-#' inherited from [`data.frame`][base::data.frame()]), for example, [dataset()]
+#' inherited from [`data.frame`][base::data.frame()]), for example, [dataset_df()]
 #' [tibble::tibble()], [tsibble::tsibble()], [data.table::data.table()].
 #' @param idcol The identifier column. If \code{idcol} is \code{NULL} it attempts to
 #' use the \code{row.names(df)} as an \code{idcol}.
@@ -13,11 +13,11 @@
 
 dataset_to_triples <- function(x, idcol=NULL) {
 
-  is_dataset <- inherits(x, "dataset")
+  is_dataset <- inherits(x, "dataset_df")
 
   if (is_dataset) {
     new_title = paste0(dataset_title(x), " [triple form]")
-    DataBibentry <- dataset_bibentry(x)
+    DataBibentry <- get_bibentry(x)
     new_Subject <- subject(x)
   }
 
@@ -32,21 +32,14 @@ dataset_to_triples <- function(x, idcol=NULL) {
   }
 
   triple_list <- lapply (seq_along_cols, function(i) {
-    data.frame(s = x[[idcol]],
+    data.frame(s = as.character(x[[idcol]]),
                p = names(x)[i],
-               o = x[[i]])
+               o = as.character(x[[i]]))
     })
 
   tmp <- do.call(rbind, triple_list)
 
-  if (is_dataset) {
-    tmp2 <- dataset(x=tmp, author=creator(x), title = new_title)
-    tmp_DSD <- DataStructure(tmp2)
-    tmp_DSD$s$label <- "Subject"
-    tmp_DSD$s$label <- "Object"
-    tmp_DSD$p$label <- "Predicate"
-    DataStructure_update(tmp2, tmp_DSD)
-  }
   tmp
+
 }
 
