### 2024-12-16 87dd990
CRAN comments, missing examples
