+++ cran-comments.md
@@ -1,21 +1,20 @@
 ## Test environments
 
 Local:
-* Windows10 x86_64-w64-mingw32 (64-bit), R version 4.3.2, locally.
-* Ubuntu 22.04.1 LTS x86_64-pc-linux-gnu (64-bit), R version 4.2.1, locally
-
-Appveyor:
-* Windows Server 2019 x64 (build 17763); R version 4.3.2 Patched (2023-12-01 r85659 ucr)
+* Windows10 x86_64-w64-mingw32 (64-bit), R version 4.4.0, locally.
 
 r_hub:
-*  Windows Server 2022 x64 (build 20348); R Under development (unstable)
-*  Windows Server 2022, R-devel, 64 bit
-*  Fedora Linux, R-devel, clang, gfortran
-*  Ubuntu Linux 20.04.1 LTS, R-release, GCC
-*  r-release-macosx-arm64|4.3.0|macosx|macOS 13.3.1 (22E261)|Mac mini|Apple M1||en_US.UTF-8|macOS 11.3|clang-1403.0.22.14.1|GNU Fortran (GCC) 12.2.0
+* macos (latest release)
+* windows-latest (release)
+* ubuntu-latest (release) - ubuntu-latest pipelines will use ubuntu-24.04 soon. 
+* ubuntu-latest (oldrel) - ubuntu-latest pipelines will use ubuntu-24.04 soon.
+* ubuntu-latest (devel)
 
 0 errors ✔ | 0 warnings ✔ | 0 notes ✔
 
-Comment: On CRAN the Mac and Windows old release has thrown an error in on of the testthat files. In my opinion, the error was not caused by the function, but the test itself. I changed the test. 
+Comment:
+The 0.3.1 version of this package was on CRAN but archived due to a documentation problem.
+The new version solve documentation problems, but also has a thoroughly rewritten s3 class system, which 
+does not affect the functionality of must functions.
 
 
