### 2024-12-16 98a7755
document expected returns
