+++ NAMESPACE
@@ -21,13 +21,13 @@ S3method(var_namespace,default)
 S3method(var_unit,default)
 S3method(xsd_convert,Date)
 S3method(xsd_convert,POSIXct)
-S3method(xsd_convert,boolean)
 S3method(xsd_convert,character)
 S3method(xsd_convert,data.frame)
 S3method(xsd_convert,dataset)
 S3method(xsd_convert,factor)
 S3method(xsd_convert,haven_labelled_defined)
 S3method(xsd_convert,integer)
+S3method(xsd_convert,logical)
 S3method(xsd_convert,numeric)
 S3method(xsd_convert,tibble)
 export("creator<-")
