+++ NAMESPACE
@@ -109,5 +109,6 @@ importFrom(rlang,env_is_user_facing)
 importFrom(stats,setNames)
 importFrom(tibble,new_tibble)
 importFrom(utils,bibentry)
+importFrom(utils,citation)
 importFrom(utils,person)
 importFrom(vctrs,df_list)
