### 2024-12-17 25bf9f2
update CRAN comments
