+++ cran-comments.md
@@ -8,7 +8,7 @@ r_hub:
 * macos-arm64
 * windows-latest (release)
 * ubuntu-latest (release) - ubuntu-latest pipelines will use ubuntu-24.04 soon. 
-* ubuntu-latest (oldrel) - ubuntu-latest pipelines will use ubuntu-24.04 soon.
+* ubuntu-latest (oldrel)  - ubuntu-latest pipelines will use ubuntu-24.04 soon.
 * ubuntu-latest (devel)
 * atlas
 
