### 2024-12-17 bcc7b3e
Function had no return value documented
