+++ R/datacite.R
@@ -83,7 +83,7 @@
 #'    Geolocation = "US",
 #'    Language = "en")
 #'
-#' as_datacite(iris_dataset, type="dataset")
+#' as_datacite(iris_dataset)
 #' @export
 
 datacite <- function(Title,
