+++ tests/testthat/test-datacite.R
@@ -17,7 +17,6 @@ test_that("as_datacite() works", {
   expect_true(is.datacite(as_datacite(iris_dataset)))
   expect_true(is.list(as_datacite(iris_dataset, type="list")))
   expect_equal(as_datacite(iris_dataset, type="list")$Publisher, 'American Iris Society')
-  expect_equal(dataset_title(as_datacite(iris_dataset, type="dataset")), "The DataCite Metadata of `Iris Dataset'")
   expect_equal(creator(as_datacite(iris_dataset, type ="dataset", author = person("Jane", "Doe"))),
             person("Jane", "Doe"))
 })
