### 2024-12-18 #553 @emilyriederer
Hi @antaldaniel - just to double check, is the right URL https://github.com/antaldaniel/dataset or https://github.com/antaldaniel/dataset now? What will the final home of the main branch of the package be?

Lets make sure the top entry is pointing to the right URL.

If https://github.com/antaldaniel/dataset is the final home, can you please set up CI there also?
