### 2024-12-18 #553 @emilyriederer
Hey @mpadge - any intuition for why the review bot isn't picking up on CI in this case? (Repo uses GitHub Actions)

I can try to dig into the bot code if it's a mystery but wanting to check if it's a known issue. Obviously I can see the CI checks so it's not a blocker for this review; just a curious phenomenon generally
