### 2024-12-18 #553 @antaldaniel
> Hey @mpadge - any intuition for why the review bot isn't picking up on CI in this case? (Repo uses GitHub Actions)
>
> I can try to dig into the bot code if it's a mystery but wanting to check if it's a known issue. Obviously I can see the CI checks so it's not a blocker for this review; just a curious phenomenon generally

I manage several packages on CRAN and the new [rhubv2](https://blog.r-hub.io/2024/04/11/rhub2/) infrastructure was also a big challenge for me, I think it is a genuinly new CI infrastructure, so maybe it is not caught by your earlier workflow analysis.  However, I think it is also very-very good so perhaps it would make sense for R packages to only rely on that. It is very well functioning, and provides tests on up to 30 scenarios with varous combinations of old, new, devel versions of both R and the various Windows, Mac, Linux distributions.
