### 2024-12-20 #553 @emilyriederer
Thanks for the context, @antaldaniel ! I'll escalate this if rhubv2 is the issue. On a prelimary look at the [pkgcheck](https://github.com/ropensci-review-tools/pkgcheck/blob/main/R/info-ci.R) repo, however, I don't believe that's the case.

In the meantime, under the `goodpractice` results, I see that you have a few `R CMD check` issues. Could you please take a look at these? (particularly the fails)

```
R CMD check with [rcmdcheck](https://r-lib.github.io/rcmdcheck/)

    checking Rd cross-references ... NOTE
    Package unavailable to check Rd xrefs: ‘tsibble’

    no_description_date
    no_import_package_as_a_whole
```
