### 2024-12-20 #553 @emilyriederer
@antaldaniel - could you please also try adding a CI badge to your README? This may be the problem.

If Rhubv2 does not provide one, there are instructions to [make your own here](https://docs.github.com/en/actions/monitoring-and-troubleshooting-workflows/monitoring-workflows/adding-a-workflow-status-badge)
