+++ README.md
@@ -7,8 +7,7 @@
 
 [![Codecov test
 coverage](https://codecov.io/gh/dataobservatory-eu/dataset/graph/badge.svg)](https://app.codecov.io/gh/dataobservatory-eu/dataset)
-![example
-workflow](https://github.com/dataobservatory-eu/dataset/actions/workflows/rhub.yaml/badge.svg)
+[![rhub](https://github.com/dataobservatory-eu/dataset/actions/workflows/rhub.yaml/badge.svg)](https://github.com/dataobservatory-eu/dataset/actions/workflows/rhub.yaml)
 [![lifecycle](https://lifecycle.r-lib.org/articles/figures/lifecycle-experimental.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental)
 [![Project Status:
 WIP](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.org/#wip)
@@ -16,7 +15,7 @@ WIP](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.o
 [![CRAN_time_from_release](https://www.r-pkg.org/badges/ago/dataset)](https://cran.r-project.org/package=dataset)
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/553_status.svg)](https://github.com/ropensci/software-review/issues/553)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14499632.svg)](https://zenodo.org/record/6950435#.YukDAXZBzIU)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.14537352.svg)](https://zenodo.org/record/6950435#.YukDAXZBzIU)\]
 [![devel-version](https://img.shields.io/badge/devel%20version-0.3.3.0007-blue.svg)](https://github.com/antaldaniel/dataset)
 [![dataobservatory](https://img.shields.io/badge/ecosystem-dataobservatory.eu-3EA135.svg)](https://dataobservatory.eu/)
 <!-- badges: end -->
@@ -73,6 +72,8 @@ The current version of the `dataset` package is in an early,
 experimental stage. You can follow the discussion of this package on
 [rOpenSci](https://github.com/ropensci/software-review/issues/553).
 
+## Semantically richer data frames
+
 ``` r
 library(dataset)
 iris_ds <- dataset_df(
@@ -92,14 +93,15 @@ iris_ds <- dataset_df(
 It is mandatory to add a `title`, `author` to a dataset, and if the
 `date` is not specified, the current date will be added.
 
-As the dataset at this point is just created, if it is not published
-yet, the `identifer` receives the default `:tba` value, a `version` of
-0.1.0 and the `:unas` (unassigned) `publisher` field.
+As the `dataset_df` at this point is just created, if it is not
+published yet, the `identifer` receives the default `:tba` value, a
+`version` of 0.1.0 and the `:unas` (unassigned) `publisher` field.
 
-The dataset behaves as expected, with all data.frame methods applicable.
-If the dataset was originally a tibble or data.table object, it retained
-all methods of these s3 classes because the dataset class only
-implements further methods in the attributes of the original object.
+The `dataset_df` behaves as expected from a data.frame-like object. See
+more information about the enhanced semantic capabilities of these data
+frames in the vignette article [Create Datasets that are Easy to Share
+Exchange and
+Extend](https://dataset.dataobservatory.eu/articles/dataset_df.html)
 
 ``` r
 summary(iris_ds)
@@ -113,7 +115,7 @@ summary(iris_ds)
 #>  Max.   :7.900000     Max.   :4.400000     Max.   :6.900        Max.   :2.5000000    NA
 ```
 
-A brief description of the extended metadata attributes:
[…]
