+++ NEWS.md
@@ -1,3 +1,7 @@
+# dataset 0.3.4
+
+- New release candidate on CRAN.
+
 # dataset 0.3.2
 
 - [New Requirement](https://dataset.dataobservatory.eu/articles/new-requirements.html) setting for further works.
