+++ R/dataset_to_triples.R
@@ -3,8 +3,7 @@
 #' columns \code{s} for subject, \code{p} for predicate and \code{o} for
 #' object.
 #' @param x An R object that contains the data of the dataset (a data.frame or
-#' inherited from [`data.frame`][base::data.frame()]), for example, [dataset_df()]
-#' [tibble::tibble()], [tsibble::tsibble()], [data.table::data.table()].
+#' inherited from [`data.frame`][base::data.frame()]), for example, [dataset_df()].
 #' @param idcol The identifier column. If \code{idcol} is \code{NULL} it attempts to
 #' use the \code{row.names(df)} as an \code{idcol}.
 #' @return The long form version of the original dataset, retaining the attributes
@@ -18,9 +17,9 @@ dataset_to_triples <- function(x, idcol=NULL) {
   is_dataset <- inherits(x, "dataset_df")
 
   if (is_dataset) {
-    new_title = paste0(dataset_title(x), " [triple form]")
+    new_title    <  paste0(dataset_title(x), " [triple form]")
     DataBibentry <- get_bibentry(x)
-    new_Subject <-  subject(x)
+    new_Subject  <-  subject(x)
   }
 
   if (is.null(idcol)) {
