### 2024-12-20 #553 @antaldaniel
@emilyriederer I hopefully made a well-functioning badge (it appears to be working to me), and I also ran goodpractice() locally, should have done that before, and made some minor cosmetic documentation changes that were not flagged by the rhub checks (some functions and methods did not have return values, now hopefully all of them have.)  This is now the 0.3.4 version.

Regarding the issues flagged by the goodpractice are in fact seem to be devtools::document() issues that I [reported](https://github.com/r-lib/devtools/issues/2488#issuecomment-2557487476).  The tsibble example is no longer an issue. I ran goodpractice locally, and there are a few instances of bad practice; some are related to long lines in the documentation, which  I will try to resolve after the review when possible (sometimes long URLs cause the problem and there is nothing to do about it.) And there a few instances when accidentally = is used instead of the <- assignment, I fixed these.
