+++ tests/testthat/test-var_namespace.R
@@ -19,8 +19,8 @@ test_that("set_namespace_attribute() works", {
 
 
 test_that("`var_namespace<-` works", {
-  var_namespace(qid) <- "https://www.wikibase.se/"
   qid <- defined(c("Q275912", "Q116196078"), namespace = "https://www.wikidata.org/wiki/")
+  var_namespace(qid) <- "https://www.wikibase.se/"
   expect_equal(get_namespace_attribute(qid), "https://www.wikibase.se/")
   expect_error(var_namespace(qid) <-  c("https://www.wikidata.org/wiki/", "https://www.wikiba.se/"))
 })
