### 2024-12-20 65fe1c3
further tests for subject
