### 2024-12-21 #553 @emilyriederer
Thanks @antaldaniel ! That seems to have fixed the CI issue.

I think we are in good shape to look for an editor. This may be slow these next few weeks (e.g. rOpenSci is actually closed to new submissions through year end) but we will be in touch soon.
