### 2024-12-23 #553 @antaldaniel
@emilyriederer Thank you very much. In the meantime, the package got back to CRAN, so I think it technically works. In my opinion, the real conceptual issue would be how this package could be well integrated with some other rOpenSci packages, mainly rdflib (they could form a super powerful usecase together, i.e., to export in a semantically rich way any R dataset to RDF, making them truly interoperable), and perhaps with dataspice, as I think that this package will solve in a much more general and broader sense what that pacakge is aiming for.

I am looking forward to hear from you and I wish a very nice festive season to you and all rOpenSci reviewers
