### 2024-12-24 03baa5e
Update DOI, CRAN release data
