+++ R/provenance.R
@@ -40,23 +40,28 @@ provenance <- function(x) {
 
 #' @keywords internal
 default_provenance <- function(dataset_id = "http://example.com/dataset#",
-                               creator_id =NULL,
-                               started_at_time = started_at_time,
-                               ended_at_time = ended_at_time) {
+                               author = NULL,
+                               dtm = NULL,
+                               generated_at_time = NULL) {
   cite_dataset <- utils::citation("dataset")
-  if(is.null(creator_id)) creator_statement <- NULL else {
-    creator_statement <- n_triple(creator_id, "a", "http://www.w3.org/ns/prov#Agent")
-  }
+
+  agent_triples <- prov_author(author)
+  if((!is.null(dtm))) c(agent_tripels, prov_author(dtm))
+
+  if (is.null(generated_at_time)) generated_at_time <- Sys.time()
+  bundle_id <- gsub("#", "_prov.nt", dataset_id)
   prov <- n_triples(
-    c(n_triple(dataset_id, "a", "http://purl.org/linked-data/cube#DataSet"),
-      creator_statement,
+    c(n_triple(bundle_id, "a", "http://www.w3.org/ns/prov#Bundle"),
+      n_triple(dataset_id, "a", "http://www.w3.org/ns/prov#Entity"),
+      n_triple(dataset_id, "a", "http://purl.org/linked-data/cube#DataSet"),
+      n_triple(dataset_id, "a", "http://purl.org/linked-data/cube#DataSet"),
+      agent_triples,
+      n_triple("https://doi.org/10.32614/CRAN.package.dataset", "a", "http://www.w3.org/ns/prov#SoftwareAgent"),
       n_triple("http://example.com/creation", "a", "http://www.w3.org/ns/prov#Activity"),
-      n_triple("http://example.com/creation", "http://www.w3.org/ns/prov#startedAtTime", xsd_convert(started_at_time) ),
-      n_triple("http://example.com/creation", "http://www.w3.org/ns/prov#endedAtTime", xsd_convert(ended_at_time) ),
+      n_triple("http://example.com/creation", "http://www.w3.org/ns/prov#generatedAtTime", generated_at_time ),
       n_triple(paste0("https://doi.org/", cite_dataset[[2]]$doi), "a", "http://www.w3.org/ns/prov#SoftwareAgent")
     )
   )
+
   prov
 }
-
-
