+++ tests/testthat/test-dataset_df.R
@@ -2,7 +2,22 @@ test_that("dataset_df() works", {
   expect_equal(is.dataset_df(dataset_df(mtcars)), TRUE)
   expect_false(is.dataset_df(mtcars))
   expect_equal(get_bibentry(dataset_df(mtcars))$author, person("Author", "Unknown"))
+  expect_equal(dataset_df(mtcars, identifier = c(mt="http:/mtcars.com/dataset#"))$rowid,
+               defined(paste0("mt:", 1:nrow(mtcars)), namespace=c(mt="http:/mtcars.com/dataset#")))
+  my_dataset <- dataset_df(
+    country_name = defined(
+      c("AD", "LI"),
+      definition = "http://data.europa.eu/bna/c_6c2bb82d",
+      namespace = "https://www.geonames.org/countries/$1/"),
+    gdp = defined(
+      c(3897, 7365),
+      label = "Gross Domestic Product",
+      unit = "million dollars",
+      definition = "http://data.europa.eu/83i/aa/GDP")
+  )
+  expect_equal(var_label(my_dataset$gdp), "Gross Domestic Product")
 })
+
 
 test_that("dataset_df() works", {
   test_dataset <- dataset_df(a=3, dataset_bibentry = datacite(Title="Hello", Creator = "Jane Doe"))
@@ -17,9 +32,9 @@ test_that("subsetting works", {
 })
 
 test_that("new_my_tibble() works", {
-  myiris <- new_my_tibble(x=iris)
+  myiris <- new_my_tibble(x=iris, identifier="example")
   expect_error(new_my_tibble(2))
-  expect_equal(class(new_my_tibble(iris)), c("dataset_df", "tbl_df", "tbl", "data.frame"))
+  expect_equal(class(new_my_tibble(iris, identifier = "example")), c("dataset_df", "tbl_df", "tbl", "data.frame"))
   expect_output(print(provenance(myiris)), "<http://example.com/dataset#>")
 })
 
@@ -35,7 +50,7 @@ test_that("rbind works", {
 })
 
 test_that("print.dataset_df() works", {
-  expect_output(print(iris_dataset), "Anderson E", ignore.case = FALSE)
+  expect_output(print(iris_dataset), "E. Anderson.", ignore.case = FALSE)
 })
 
 test_that("as_dataset_df() works", {

[…]
