### 2024-12-24 36eaed2
remove unnecessary test
