+++ README.Rmd
@@ -62,7 +62,7 @@ iris_ds <- dataset_df(
     creator = person("Edgar", "Anderson", role = "aut"),
     publisher = "American Iris Society",
     datasource = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x",
-    date = 1935,
+    dataset_date = 1935,
     language = "en",
     description = "This famous (Fisher's or Anderson's) iris data set."
   )
@@ -94,7 +94,7 @@ The descriptive metadata are added to a `utils::bibentry` object which has many 
 
 ## Semantically richer data frame columns
 
-It is important to see that we do not only increase the semantics of the dataset as a whole, but also the semantics of each variable. R users often have a problem with the reusability of their data frames because, by default, a variable is only described by a programmatically usable name label; for example, in the famous `iris` dataset, the length of the sepal for each observation (row) is in the `iris$Sepal.Length` column. If we would like to add rows to this dataset, it is essential to know if the numbers in the `iris$Sepal.Length` are measured in millimetres centimetres or inches. 
+It is important to see that we do not only increase the semantics of the dataset as a whole, but also the semantics of each variable. R users often have a problem with the reusability of their data frames because, by default, a variable is only described by a programmatically usable name label; for example, in the famous `iris` dataset, the length of the sepal for each observation (row) is in the `iris$Sepal.Length` column. If we would like to add rows to this dataset, it is essential to know if the numbers in the `iris$Sepal.Length` are measured in millimetres, centimetres or inches. 
 
 When working with datasets that receive their components from different linked open data sources, it is particularly important to have a more precise semantic definition and description of each variable.
 
