+++ README.md
@@ -16,7 +16,7 @@ WIP](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.o
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/553_status.svg)](https://github.com/ropensci/software-review/issues/553)
 [![DOI](https://zenodo.org/badge/DOI/10.32614/CRAN.package.dataset.svg)](https://zenodo.org/record/6950435#.YukDAXZBzIU)
-[![devel-version](https://img.shields.io/badge/devel%20version-0.3.4-blue.svg)](https://github.com/dataobservatory-eu/dataset)
+[![devel-version](https://img.shields.io/badge/devel%20version-0.3.4001-blue.svg)](https://github.com/dataobservatory-eu/dataset)
 [![dataobservatory](https://img.shields.io/badge/ecosystem-dataobservatory.eu-3EA135.svg)](https://dataobservatory.eu/)
 <!-- badges: end -->
 
@@ -80,7 +80,7 @@ iris_ds <- dataset_df(
     creator = person("Edgar", "Anderson", role = "aut"),
     publisher = "American Iris Society",
     datasource = "https://doi.org/10.1111/j.1469-1809.1936.tb02137.x",
-    date = 1935,
+    dataset_date = 1935,
     language = "en",
     description = "This famous (Fisher's or Anderson's) iris data set."
   )
@@ -102,7 +102,14 @@ Extend](https://dataset.dataobservatory.eu/articles/dataset_df.html)
 
 ``` r
 summary(iris_ds)
-#> Anderson E (1935). "Iris Dataset."
+#> [1] E. Anderson. _Iris Dataset_. En. DCMITYPE:Dataset. 1935.
+#>     rowid          
+#>  Length:150        
+#>  Class :character  
+#>  Mode  :character  
+#>                    
+#>                    
+#>                    
 #>    x.Sepal.Length        x.Sepal.Width       x.Petal.Length        x.Petal.Width          x.Species     
 #>  Min.   :4.300000     Min.   :2.000000     Min.   :1.000        Min.   :0.1000000    setosa    :50      
 #>  1st Qu.:5.100000     1st Qu.:2.800000     1st Qu.:1.600        1st Qu.:0.3000000    versicolor:50      
@@ -116,21 +123,7 @@ The dataset_df A brief description of the extended metadata attributes:
 
 ``` r
 print(get_bibentry(iris_ds), "Bibtex")
-#> @Misc{,
-#>   title = {Iris Dataset},
-#>   author = {Edgar Anderson},
-#>   identifier = {:tba},
-#>   publisher = {American Iris Society},
-#>   year = {1935},
-#>   language = {en},
-#>   relation = {:unas},
-#>   format = {:unas},
-#>   rights = {:tba},
-#>   description = {This famous (Fisher's or Anderson's) iris data set.},
-#>   type = {DCMITYPE:Dataset},
-#>   datasource = {https://doi.org/10.1111/j.1469-1809.1936.tb02137.x},
-#>   coverage = {:unas},
-#> }
+#> [1] E. Anderson. _Iris Dataset_. En. DCMITYPE:Dataset. 1935.
 ```
 
 ``` r
@@ -154,7 +147,7 @@ by default, a variable is only described by a programmatically usable
 name label; for example, in the famous `iris` dataset, the length of the
 sepal for each observation (row) is in the `iris$Sepal.Length` column.
 If we would like to add rows to this dataset, it is essential to know if
[…]
