+++ DESCRIPTION
@@ -3,7 +3,7 @@ Title: Create Data Frames that are Easier to Exchange and Reuse
 Version: 0.3.4001
 Date: 2024-12-24
 DOI: 10.32614/CRAN.package.dataset
-Language: en-US
+Language: en-GB
 Authors@R: 
     c(person(given = "Daniel", family = "Antal", 
            email = "daniel.antal@dataobservatory.eu", 
