+++ tests/testthat/test-publication_year.R
@@ -4,6 +4,7 @@ test_that("publication_year() works", {
   expect_warning(publication_year(iris_dataset, overwrite=F) <- 1934)
   })
 
+
 value <- 1936
 
 test_that("publication_year() <- assignment works", {
