+++ vignettes/rdf.Rmd
@@ -19,14 +19,69 @@ library(dataset)
 library(rdflib)
 ```
 
+The **RDF** (Resource Description Framework) annotation significantly enhances the interoperability and exchangeability of datasets in data repositories by leveraging a standardised, machine-readable format for describing and linking data. This vignette shows how to leverage the capabilities of the _dataset_ package with [rdflib](https://docs.ropensci.org/rdflib/index.html), an R-user-friendly wrapper on ROpenSci to work with the _redland_ Python library for performing common tasks on rdf data, such as parsing and converting between formats including rdfxml, turtle, nquads, ntriples, and trig, creating rdf graphs, and performing SPARQL queries.
 
-```{r prov}
-provenance(iris_dataset)
+## Standardised Semantic Framework
+RDF provides a common framework to describe resources and their relationships using triples (subject-predicate-object). This standardisation ensures that data from different systems can be understood in a unified way, regardless of the original source or format.  Notice that this format is a stricter version of the tidy dataset concept, where not only on every observation is in a row, but there are always strictly three columns.
+
+```{r iris}
+head(iris)[1:2,]
 ```
+Instead of placing the relevant measurement of an observed flower into the intersection of columns and rows, in the triple format we put them next to each other: 
+
+- the first flower's sepal length is 5.1
+- the second flowers's sepal length is 4.9
+
+
+```{r triples}
+dataset_to_triples(iris[1:2,])[1:10,]
+```
+We describe the `dataset_df` datasets in such triplets, where each triplet is a semantic statement: it connects a single observation unit with a single measurement.
+
+##  Enhanced Interoperability
+RDF uses globally unique identifiers (URIs) for resources, ensuring that different datasets can reference the same entities unambiguously. This allows seamless data integration and querying across repositories, even if the datasets come from diverse domains.
+
+Our `defined` class supports this enhanced interoperability. In the example below, an application can look up that the numeric values in your table conform the statistical definition of GDP, and they are expressed in millions of dollars; meaning that you have to multiply them by 1000 if you want to join them with different data expressed in thousands of dolllars. 
+
+```{r defined}
+gdp_vector <- defined(
+  c(3897, 7365, 6753),
+  label = "Gross Domestic Product",
+  unit = "https://rdf.vegdata.no/V440/v440-doc/v440-brudata-owl-doc/unit_MillionUSD.html",
+  definition = "http://data.europa.eu/83i/aa/GDP"
+ )
+```
+
+There are several ways to add permanent identifiers to observational units, variable definitions, and specific observed values. The simplest (but certainly not the easiest to read for a human eye) standard format for writing them into a plain text file that you can share online is the [RDF 1.1 N-Triples](https://www.w3.org/TR/n-triples/) format.The NTriple format creates URIs (similarly formatted as URLs) for the definitions that can be looked up in an online resource. This can be combined with literal strings that may also include information if they should be read back to a system as strings, doubles, integers, dates or date-time variables.
+
+```{r ntriples}
+n_triple(s="https://doi.org/10.5281/zenodo.10396807",  # permanent, global ID of the dataset
+         p="http://purl.org/dc/terms/description",     # library definition of 'description' 
+         o="The famous (Fisher's or Anderson's) iris data set.")  # literal string
+```
+## Richer metaadata
+RDF supports linking datasets through shared URIs, enabling the creation of interconnected knowledge graphs. Linked Data principles help relate datasets in meaningful ways, making it easier to discover, navigate, and integrate information. RDF annotations allow datasets to include detailed metadata about their structure, provenance, usage rights, and content. This metadata provides critical context, enabling automated tools to interpret and process the data effectively.
+
[…]
