+++ R/identifier.R
@@ -1,18 +1,23 @@
 #' @title Get/set the Identifier of the object.
-#' @description Add the optional Identifier property as an attribute to an R object.
-#' @details The \code{Identifier} is an unambiguous reference to the resource within a given context.
-#' Recommended practice is to identify the resource by means of a string conforming to an
-#' identification system. Examples include International Standard Book Number (ISBN),
+#' @description Add the optional Identifier property as an attribute to an R
+#' object.
+#' @details The \code{Identifier} is an unambiguous reference to the resource
+#' within a given context. Recommended practice is to identify the resource by
+#' means of a string conforming to an identification system. Examples include
+#' International Standard Book Number (ISBN),
 #' Digital Object Identifier (DOI), and Uniform Resource Name (URN).
 #' Select and identifier scheme from
 #' \href{https://www.ukoln.ac.uk/metadata/dcmi-ieee/identifiers/index.html}{registered URI schemes maintained by IANA}.
 #' More details: \href{https://www.ukoln.ac.uk/metadata/dcmi-ieee/identifiers/}{Guidelines for using resource identifiers in Dublin Core metadata and IEEE LOM}.
 #' Similar to \code{Identifier} in \code{\link{datacite}}.
 #' \href{https://support.datacite.org/docs/datacite-metadata-schema-v44-mandatory-properties#1-identifier}{DataCite 4.4}.\cr
-#' It is not part of the "core" Dublin Core terms, but we always add it to the metadata attributes
-#' of a dataset (in case you use a strict Dublin Core property sheet you can omit it.)
+#' It is not part of the "core" Dublin Core terms, but we always add it to the
+#' metadata attributes of a dataset (in case you use a strict Dublin Core
+#' property sheet you can omit it.)
 #' \href{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/}{Dublin Core metadata terms}.
-#' @param x An R object, such as a data.frame, a tibble, or a data.table.
+#' @param x An \code{\link{dataset_df}} object or a
+#' \code{\link[utils:bibentry]{utils::bibentry}} object, including possibly an
+#' instance of its \code{\link{dublincore}} or \code{\link{datacite}} subclass.
 #' @param value The  \code{Identifier} as a character string.
 #' @param overwrite If the attributes should be overwritten. In case it is set to \code{FALSE},
 #' it gives a message with the current \code{Identifier} property instead of overwriting it.
@@ -27,33 +32,71 @@
 
 #' @export
 identifier <- function(x) {
-  assert_that(is.dataset_df(x),
-              msg = "identifier(x): x must be a dataset object created with dataset_df() or as_dataset_df().")
+  assert_that(inherits(x, "bibentry") | inherits(x, "dataset_df"),
+              msg = "identifier(x): x must be a dataset_df or a bibentry object.")
 
-  ds_bibentry <- get_bibentry(x)
+  if (is.dataset_df(x)) {
+    ds_bibentry <- get_bibentry(x)
+  } else {
+    ds_bibentry <- x
+  }
+
   ds_bibentry$identifier
 }
 
 #' @rdname identifier
 #' @export
 `identifier<-`  <- function(x,  overwrite = TRUE, value) {
-  assert_that(is.dataset_df(x),
-              msg = "identifier(x) <- value: x must be a dataset object created with dataset_df() or as_dataset_df().")
 
-  ds_bibentry <- get_bibentry(x)
+  if (is.numeric(value)) value <- as.character(value)
+
[…]
