+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: dataset
 Title: Create Data Frames that are Easier to Exchange and Reuse
-Version: 0.3.4001
-Date: 2024-12-24
+Version: 0.3.4002
+Date: 2024-12-26
 DOI: 10.32614/CRAN.package.dataset
 Language: en-GB
 Authors@R: 

[…]
