### 2025-01-20 #553 @maelle
@antaldaniel is https://dataset.dataobservatory.eu/articles/new_requirements.html still relevant?
