### 2025-01-20 #553 @antaldaniel
@maelle I think it may be relevant, but not strictly speaking. As there was nobody to review it, I actually implemented a large part of it with slight deviations, and other parts are not yet implemented.  I hope that the intention of what I would like to develop (co-develop with others...) is clear without it, too, but a quick read can potentially help for context.
