### 2025-01-21 #553 @maelle
@antaldaniel thanks again for your submission. I'm preparing my editorial comments.

I'd like to ask you, though, to create a new submission in a new issue, with a link to this one, if that's ok with you. This way, newcomers to the submission like reviewers won't have to sift through to so many comments (some of them relevant, some of them not) so it'd be less overwhelming. Thank you!
