### 2025-01-21 #553 @antaldaniel
@maelle I also see merits of keeping a review in one thread, especially as it looks that I am chasing a moving target, the requirements changed significantly since this proces started. Nevertheless, I tried to fill in the best way I could the new form here> https://github.com/ropensci/software-review/issues/681
