### 2025-01-21 #553 @maelle
Followed in https://github.com/ropensci/software-review/issues/681
